DFDC Airfoil Type

The DFDC Airfoil type is very similar to the XROTOR airfoil type, but includes additions for cascade corrections based on stagger and solidity. The cascade corrections aren't particularly accurate, but they do apply ballpark effects resulting from high solidity blade sections. The main benefit to this airfoil type is its simplicity and that the post-stall behavior is already in a format allowing more robust convergence of the DuctAPE solvers.

DuctAPE.C4Blade.DFDCairfoil — Type

DFDCairfoil

Fields

alpha0::Vector{Float} : zero lift angle of attack

clmax::Vector{Float} : maximum cl

clmin::Vector{Float} : minimum cl

dclda::Vector{Float} : lift curve slope (1/radians)

dclda_stall::Vector{Float} : lift curve slope post-stall (1/radians)

dcl_stall::Vector{Float} : cl increment from initial to total stall.

cdmin::Vector{Float} : minimum cd

cldmin::Vector{Float} : cl at cdmin

dcddcl2::Vector{Float} : quadratic curve factor for cl vs cd curve $\left(\frac{d(c_d)}{d(c_l^2)}\right)$

cmcon::Vector{Float} : pitching moment constant (unused right now)

Re_ref::Vector{Float} : reference Reynolds number at which cd values apply

Re_exp::Vector{Float} : Reynolds number exponent scaling $\left( c_d = c_d(Re/Re_{ref})^{Re_{exp}}\right)$ should be 0.2 for fully laminar and 0.5 for fully turbulent

mcrit::Vector{Float} : critical Mach number

correct_for_mach::Vector{Float} : flag to add Prandtl-Glauert correction

correct_for_cascade::Vector{Float} : flag to add cascade corrections

correct_for_reynolds::Vector{Float} : flag to add reynolds drag correction

correct_for_transonic::Vector{Float} : flag to add drag correction above critical mach number