Actuator Disk Type

Warning

Actuator disk types are currently in development and not ready for general use.

DuctAPE currently implements an actuator disk type that can be used to directly define the rotor blade circulation.

DuctAPE.C4Blade.ADM — Type

Fields:

prescribed_circulation::Float=0.0 : Prescribed circulation strength

prescribed_source_strength::Float=0.0 : Prescribed source panel strength