Cascade Types

Warning

Cascade types are currently in development and not ready for general use.

Cascade types are defined analogous to CCBlade airfoil types. Instead of angle of attack, however, cascade types take in both inflow and stagger angles. In addition, cascade types are dependent on local solidity.

DuctAPE.C4Blade.InReStSoMaCAS — Type

InReStSoMaCAS(inflow, Re, stagger, solidity, Mach, cl, cd, info)
InReStSoMaCAS(inflow, Re, stagger, solidity, Mach, cl, cd)
InReStSoMaCAS(filenames::Matrix{String}; radians=true)

Data is fit recursively with Akima splines.

Arguments:

inflow::Vector{Float64}: inflow angles

Re::Vector{Float64}: Reynolds numbers

stagger::Vector{Float64}: stagger angles

solidity::Vector{Float64}: local solidity

Mach::Vector{Float64}: Mach numbers

cl::Array{Float64}: lift coefficients where cl[i, j, k, ell] corresponds to stagger[i], Re[j], Mach[k], solidity[ell]

cd::Array{Float64}: drag coefficients where cd[i, j, k, ell] corresponds to stagger[i], Re[j], Mach[k], solidity[ell]

info::String: a description of this airfoil data (just informational)

or files with one per Re/Stagger/Solidty/Mach combination

Arguments:

filenames::Matrix{String}: name/path of files to read in. filenames[i, j, k, ell] corresponds to Re[i] Stagger[j] Stagger[k] and Solidity[k] with each in ascending order.

radians::Bool: true if angle of attack in file is given in radians

DuctAPE.C4Blade.interp5d — Method

 interp5d(interp1d, x1data, x2data, x3data, x4data, fdata, x1pt, x2pt, x3pt, x4pt)

Same as FLOWMath.interp4d, ex1cept in five dimensions.

DuctAPE.C4Blade.parsecascadefile — Method

parsefile(filename, radians, solidity)

Cascade version of parsefile function from CCBlade. Assumes stagger is given before reynolds and Mach number, and solidity is given after

DuctAPE.C4Blade.writecascadefile — Method

writecascadefile(filename, info, Re, Mach, stagger, inflow, cl, cd, radians)

Cascade version of writecascadefile function from CCBlade. Writes solidity after Mach number