Airfoil Polar Corrections

In some cases various airfoil polar corrections may be required. Of specific note are modifications to airfoil polars for post-stall behavior. Thus far, DuctAPE is much more robust if the post-stall behavior in the lift polars does not exhibit a decrease in lift at angles of attack beyond that of the maximum lift coefficient. Therefore a function is provided to help modify polars as needed:

DuctAPE.C4Blade.stall_limiters — Function

stall_limiters(
    aoa,
    cl,
    cd;
    clminid=nothing,
    clmaxid=nothing,
    cl_cutoff_slope=0.1,
    cd_cutoff_slope=0.9,
    N=20,
    blend_hardness=50
)

Cuts off coefficient vs alpha curve at min and max coefficient and places rest of curve from -pi to min coeff and max coeff to pi according to user defined clcutoffslope (default 0.1)

Arguments:

aoa::AbstractVector{Float} : input angles of attack, in radians

cl::AbstractVector{Float} : input lift coefficients

cd::AbstractVector{Float} : input drag coefficients

Keyword Arguments:

clminid::Float=nothing : manually set index for minimum cl

clmaxid::Float=nothing : manually set index for maximum cl

cl_cutoff_slope::Float=0.1 : "post-stall" slope for cl

cd_cutoff_slope::Float=0.1 : "post-stall" slope for cd

blend_hardness::Float=50 : hardenss of blend between nominal polar and post-stall modifications.

Returns:

aoa_ext::AbstractVector{Float} : angles of attack for modified polar, in radians

cl_ext::AbstractVector{Float} : modified lift coefficients

cd_ext::AbstractVector{Float} : modified drag coefficients

Various other correction methods are available, including the cascade corrections inherent in the DuctAPE.C4Blade.DFDCairfoil type. The following methods are in addition to the various corrections available alongside the CCBlade Airfoil Types.

DuctAPE.C4Blade.corrected_clcd — Method

corrected_clcd(af::AlphaReAF, alpha, Re, Mach, solidity, stagger; kwargs...)

Evaluates and applies on-the-fly corrections for airfoil lift and drag. On-the-fly airfoil polar corrections include solidity/stagger corrections, Prandtl-Glauert compressibility corrections, and transonic lift limits and drag additions.

corrected_clcd!(cl, cd, af::AlphaReAF, Re, alpha, Mach, solidity, stagger; kwargs...)

Evaluates and applies on-the-fly corrections for airfoil lift and drag in place.

corrected_clcd!(cl, cd, Mach, solidity, stagger; kwargs...)

Applies on-the-fly corrections for airfoil lift and drag in place.

corrected_clcd!(cl, cd, af::AlphaAF, alpha, Re, Mach, solidity, stagger; kwargs...)

Evaluates and applies on-the-fly corrections, including Reynolds corrections, for airfoil lift and drag in place

corrected_clcd(cas::InReStSoMaCAS, inflow, Re, Mach, solidity, stagger)

Evaluates cascade lift and drag.

Arguments:

Coefficients

cl::Float : local lift coefficient

cd::Float : local drag coefficient

Airfoil Object

af::AlphaReAF : airfoil object of CCBlade type dependent on angle of attack and Reynolds number

or

af::AlphaAF : airfoil object of CCBlade type dependent on angle of attack only

or

cas::InReStSoMaCAS : cascade object depentent on inflow angle, Reynolds number, stagger, solidity, and Mach number.

Flow Angle

alpha::Float : angle of attack, radians. Used with airfoil types

or

inflow::Float : inflow angle, radians. Used with cascade types

Flow Conditions

Re::Float : Reynolds number

Mach::Float : Mach number

Geometry

solidity::Float : Local solidity

stagger::Float : Stagger angle, radians

Keyword Arguments:

mcrit::Float=0.7 : Critical Mach number

rotorzloc airfoil type parameters for post-stall behavior

dcl_stall::Float=0.1 : change in cl from incipient to total stall, used in transonic lift limiter correction

dclda_stall::Float=0.1 : Post-stall lift curve slope

Correction factors that were hard coded in rotorzloc and DFDC

cdmfactor::Float=10.0 :

clmfactor::Float=0.25 :

mexp::Float=3.0 :

cdmstall::Float=0.1 :

cdmdd::Float=0.0020 :

Smoothing Paramters

ssblend_hardness::Float=100.0 : sigmoid blending hardness for solidity/stagger corrections

transblendhardness::Float=75.0 : sigmoid blending hardness for transonic corrections

absdx::Float=0.0625 : smooth absolute value Δα (radians) for transonic drag addition

Miscellaneous

verbose::Bool=false : Boolean of whether to print warnings, etc.

DuctAPE.C4Blade.prandtl_glauert! — Method

prandtl_glauert!(cl, ma)

In place version of pradtl_glauert.

DuctAPE.C4Blade.prandtl_glauert — Method

prandtl_glauert(cl, ma)

Applies Prandtl-Glauert correction

Arguments:

cl::Float : local lift coefficient

Returns

cl_corr::AbstractVector{Float} : corrected lift coefficients

DuctAPE.C4Blade.prandtl_glauert_factor — Method

prandtl_glauert_factor(mach; verbose=false, blend_range=0.02)

Smoothed Prandtl-Glauert Mach correction factor

Arguments:

mach::Float : Mach number

Keyword Arguments:

blend_range::Float=0.02 : range for blending factor and max cutoff (allowing Mach >= 1.0 for continuity)

DuctAPE.C4Blade.quadspline — Method

quadspline(xdata, ydata, xpoint)

Sample data in quadratic spline at give point.

DuctAPE.C4Blade.re_drag! — Method

re_drag!(cd, re, re_ref; re_exp=0.5)

In-place version of re_drag.

DuctAPE.C4Blade.re_drag — Method

re_drag(cd, re, re_ref; re_exp=0.5)

Arguments:

cd::AbstractVector{Float} : input drag coefficients

re::Float : Current Reynolds number

re_ref::Float : Reference Reynolds number (at which the cd's were generated)

Keyword Arguments:

re_exp::Float=0.5 : should be 0.2 for laminar and 0.5 for turbulent flow

Returns:

cd_corr::AbstractVector{Float} : Reynolds corrected drag coefficients

DuctAPE.C4Blade.solidity_and_stagger! — Method

solidity_and_stagger!(cl, solidity, stagger; blend_hardness=100)

In-place version of solidity_and_stagger.

DuctAPE.C4Blade.solidity_and_stagger — Method

solidity_and_stagger(cl, solidity, stagger; blend_hardness=100)

Apply smoothed Wallis' cascade correction (see solidity_and_stagger_factor_smooth) to local lift.

Arguments:

cl::AbstractVector{Float} : input lift coefficients

solidity::Float : local solidity

stagger::Float : local stagger (in radians)

Keyword Arguments:

blend_hardness::Float=100 : hardness of smoothing blends

Returns:

cl_corr::AbstractVector{Float} : corrected lift coefficients.

DuctAPE.C4Blade.solidity_and_stagger_factor — Method

solidity_and_stagger_factor(solidity, stagger; blend_hardness=100)

Correction for airfoil data used in a high-solidity cascade application. Correction is used in DFDC airfoils nominally and come from quadratic fits to curves in fig 6-29 "Axial Flow Fans and Ducts" by Wallis (1983). Note that the corrections are really only meant for Wallis' custom airfoil design and specific conditions mentioned in the book.

Arguments:

solidity::Float : local solidity

stagger::Float : local stagger (in radians)

Keyword Arguments:

blend_hardness::Float=100 : hardness for smoothing blends

DuctAPE.C4Blade.solidity_and_stagger_factor_smooth — Method

solidity_and_stagger_factor_smooth(solidity, stagger; blend_hardness=100)

A smoothed version of solidity_and_stagger_factor.

Arguments:

solidity::Float : local solidity

stagger::Float : local stagger (in radians)

Keyword Arguments:

blend_hardness::Float=100 : hardness for smoothing blends

DuctAPE.C4Blade.stall_limiters — Method

stall_limiters(
    aoa,
    cl,
    cd;
    clminid=nothing,
    clmaxid=nothing,
    cl_cutoff_slope=0.1,
    cd_cutoff_slope=0.9,
    N=20,
    blend_hardness=50
)

Cuts off coefficient vs alpha curve at min and max coefficient and places rest of curve from -pi to min coeff and max coeff to pi according to user defined clcutoffslope (default 0.1)

Arguments:

aoa::AbstractVector{Float} : input angles of attack, in radians

cl::AbstractVector{Float} : input lift coefficients

cd::AbstractVector{Float} : input drag coefficients

Keyword Arguments:

clminid::Float=nothing : manually set index for minimum cl

clmaxid::Float=nothing : manually set index for maximum cl

cl_cutoff_slope::Float=0.1 : "post-stall" slope for cl

cd_cutoff_slope::Float=0.1 : "post-stall" slope for cd

blend_hardness::Float=50 : hardenss of blend between nominal polar and post-stall modifications.

Returns:

aoa_ext::AbstractVector{Float} : angles of attack for modified polar, in radians

cl_ext::AbstractVector{Float} : modified lift coefficients

cd_ext::AbstractVector{Float} : modified drag coefficients

DuctAPE.C4Blade.transonic_drag_addition! — Method

transonic_drag_addition!(
    cd,
    cl,
    clcdmin,
    mach;
    mcrit=0.7,
    cdmfactor=10.0,
    clmfactor=0.25,
    mexp=3.0,
    cdmdd=0.0020,
    cdmstall=0.1000,
    absdx=0.0625,
    blend_hardness=50,
)

Smoothed, vecotrized, in-place version of transonic_drag_addition.

Different Arguments:

cd::AbstractVector{Float} : vector of drag coefficients

cl::AbstractVector{Float} : vector of lift coefficients

Additional Keyword Argument:

blend_hardness::Float=50 : hardenss of smoothing blends

DuctAPE.C4Blade.transonic_drag_addition — Method

transonic_drag_addition(
    cd,
    cl,
    clcdmin,
    mach;
    mcrit=0.7,
    cdmfactor=10.0,
    clmfactor=0.25,
    mexp=3.0,
    cdmdd=0.0020,
    cdmstall=0.1000,
    absdx=0.0625,
)

Drag augmentation due to transonic effects as found in XROTOR and DFDC. Note this is nominally applied to DFDC airfoil evaluation.

Arguments:

cd::Float : input drag coefficient

cl::Float : input lift coefficient

clcdmin::Float : lift coefficient at minimum drag coefficient.

mach::Float : Mach number

Keyword Arguments

mcrit::Float=0.7 : critical Mach number

cdmfactor::Float=10.0 : factor hard coded in XROTOR and DFDC

clmfactor::Float=0.25 : factor hard coded in XROTOR and DFDC

mexp::Float=3.0 : factor hard coded in XROTOR and DFDC

cdmstall::Float=0.1000 : factor hard coded in XROTOR and DFDC

absdx::Float=0.0625 : smoothing factor for smooth absolute value function

Returns:

cl_corr:Float : corrected lift coefficient

DuctAPE.C4Blade.transonic_lift_limiter — Method

transonic_lift_limiter(
    cl,
    mach,
    clcdmin,
    clmax,
    clmin,
    dclda;
    mcrit=0.7,
    dcl_stall=0.1,
    dclda_stall=0.1,
    cdmfactor=10.0,
    clmfactor=0.25,
    mexp=3.0,
    cdmstall=0.1000,
)

Airfoil polar corrections due to transonic effects as found in XROTOR and DFDC. Note that this correction is done nominally in the DFDC airfoil evaluation.

Arguments:

cl::Float : input lift coefficient

mach::Float : Mach number

clcdmin::Float : lift coefficient at minimum drag coefficient.

clmax::Float : maximum lift coefficient

clmin::Float : minimum lift coefficient

dclda::Float : lift-curve slope

mcrit::Float=0.7 : critical Mach number

dcl_stall::Float=0.1 : cl increment from initial to total stall

dclda_stall::Float=0.1 : lift curve slope post-stall (1/radians)

cdmfactor::Float=10.0 : factor hard coded in XROTOR and DFDC

clmfactor::Float=0.25 : factor hard coded in XROTOR and DFDC

mexp::Float=3.0 : factor hard coded in XROTOR and DFDC

cdmstall::Float=0.1000 : factor hard coded in XROTOR and DFDC

Returns:

cl_corr:Float : corrected lift coefficient

DuctAPE.C4Blade.transonic_lift_limiter_smooth! — Method

transonic_lift_limiter_smooth!(
    cl,
    mach,
    clcdmin,
    clmax,
    clmin,
    dclda;
    mcrit=0.7,
    dcl_stall=0.1,
    dclda_stall=0.1,
    cdmfactor=10.0,
    clmfactor=0.25,
    mexp=3.0,
    cdmstall=0.1000,
    blend_hardness=50,
)

Smoothed, vectorized, in-place version of transonic_lift_limiter.

Different Arguments:

cl::AbstractVector{Float} : vector of lift coefficients

Additional Keyword Argument:

blend_hardness::Float=50 : hardenss of smoothing blends