API Reference

API Reference

Public API

Input Types

Options

General Options

Integration Options

Solver Options

Elliptic Grid Solve

Aero Solve

Analysis

Preprocess

Process

Postprocess

Private API

Option Types

Bookkeeping

Caching

Allocation

Reshaping

Preprocess

General

Geometry

Wake

Bodies

Rotors

Induced Velocities

State Initialization

Analysis

Post-process

Convenience Functions

Index

Public API

Input Types

DuctAPE.Propulsor — Type

Propulsor(duct_coordinates, centerbody_coordinates, rotorstator_parameters, operating_point, paneling_constants, reference_parameters)

Arguments

duct_coordinates::AbstractMatrix : The [z, r] coordinates of the duct geometry beginning at the inner (casing) side trailing edge and proceeding clockwise. Note that the duct geometry absolute radial position does not need to be included here if the autoshiftduct option is selected.

centerbody_coordinates::AbstractMatrix : The [z, r] coordinates of the centerbody beginning at the leading edge and ending at the trailing edge. Note that the leading edge is assumed to be placed at a radial distance of 0.0 from the axis of rotation.

operating_point::OperatingPoint : The operating point values.

paneling_constants::PanelingConstants : Constants used in re-paneling the geometry.

rotorstator_parameters::RotorStatorParameters : Rotor (and possibly stator) geometric paramters.

reference_parameters::ReferenceParameters : Reference Parameters.

DuctAPE.RotorStatorParameters — Type

RotorStatorParameters(
    B, rotorzloc, r, Rhub, Rtip, chords, twists, tip_gap, airfoils, fliplift
)

Composite type containing the rotor(s) geometric properties.

Note that the actual struct requires the inputs to be arrays, but there is a constructor function that will take in scalars and automatically build the array-based struct.

Arguments

B::AbstractVector{Float} : The number of blades for each rotor. May not be an integer, but usually is.

rotorzloc::AbstractVector{Float} : Dimensional, axial position of each rotor.

r::AbstractArray{Float} : Non-dimensional radial locations of each blade element.

Rhub::AbstractVector{Float} : Dimensional hub radius of rotor. (may be changed if it does not match the radial position of the centerbody geometry at the selected rotorzloc.

Rtip::AbstractVector{Float} : Dimensional tip radius of rotor. Is used to determine the radial position of the duct if the autoshiftduct option is selected.

chords::AbstractArray{Float} : Dimensional chord lengths of the blade elements.

twists::AbstractArray{Float} : Blade element angles, in radians.

tip_gap::AbstractVector{Float} : Currently unused, do not set to anything other than zeros.

airfoils::AbstractArray{AFType} : Airfoil types describing the airfoil polars for each blade element. Currently only fully tested with C4Blade.DFDCairfoil types.

fliplift::AbstractVector{Bool} : flag to indicate if the airfoil lift values should be flipped or not.

DuctAPE.OperatingPoint — Type

OperatingPoint(Vinf, rhoinf, muinf, asound, Omega)

Propulsor operating point information.

Note that the actual struct requires the inputs to be arrays, but there is a constructor function that will take in scalars and automatically build the array-based struct.

Also note that even though each field is required to be a vector, only Omega should have more than one entry, and only then if there are more than one rotor. The purpose behind having vector rather than constant scalar inputs here is for ease of redefinition in an optimization setting when freestream design variables may be present.

Arguments

Vinf::AbstractVector{Float} : Freestream velocity magnitude (which is only in the axial direction).

rhoinf::AbstractVector{Float} : Freestream density

muinf::AbstractVector{Float} : Freestream viscosity

asound::AbstractVector{Float} : Freestream speed of sound

Omega::AbstractVector{Float} : Rotor rototation rate(s)

DuctAPE.PanelingConstants — Type

PanelingConstants(
    nduct_inlet,
    ncenterbody_inlet,
    npanels,
    dte_minus_cbte,
    nwake_sheets,
    wake_length=1.0,
)

Constants used in re-paneling geometry.

Note that unlike other input structures, this one, in general, does not define fields as vectors. This is because these values should not change throughout an optimization, even if the geometry may change. Otherwise, discontinuities could be experienced.

Arguments

nduct_inlet::Int : The number of panels to use for the duct inlet (this number is used for both the casing and nacelle re-paneling)

ncenterbody_inlet::Int : The number of panels to use for the centerbody inlet.

npanels::AbstractVector{Int} : A vector containing the number of panels between discrete locations inside the wake. Specifically, the number of panels between the rotors, between the last rotor and the first body trailing edge, between the body trailing edges (if different), and between the last body trailing edge and the end of the wake. The length of this vector should be N+1 (where N is the number of rotors) if the duct and centerbody trailing edges are aligned, and N+2 if not.

dte_minus_cbte::Float : An indicator concerning the hub and duct trailing edge relative locations. Should be set to -1 if the duct trailing edge axial position minus the centerbody trailing edge axial position is negative, +1 if positive (though any positive or negative number will suffice), and zero if the trailing edges are aligned.

nwake_sheets::Int : The number of wake sheets to use. Note this will also be setting the number of blade elements to use.

wake_length::Float=1.0 : Non-dimensional (based on the length from the foremost body leading edge and the aftmost body trailing edge) length of the wake extending behind the aftmost body trailing edge.

DuctAPE.ReferenceParameters — Type

ReferenceParameters(Vref, Rref)

Reference parameters for post-process non-dimensionalization.

Note that the actual struct requires the inputs to be arrays, but there is a constructor function that will take in scalars and automatically build the array-based struct.

Arguments

Vref::AbstractVector{Float} : Reference velocity.

Rref::AbstractVector{Float} : Reference rotor tip radius.

Options

General Options

DuctAPE.Options — Type

struct Options{
    TB,
    TBwo,
    TF,
    TI,
    TSf,
    TSt,
    Tin,
    TIo<:IntegrationOptions,
    TSo<:SolverOptionsType,
    WS<:GridSolverOptionsType,
}

Type containing (nearly) all the available user options.

Fields

General Options

verbose::TB = false : flag to print verbose statements

silence_warnings::TB = true : flag to silence warnings

multipoint_index::TI = [1] : holds current index of multi-point solver (no need for user to change this usually)

Geometry Re-interpolation and generation options` :

finterp::Tin = FLOWMath.akima : interpolation method used for re-paneling bodies

autoshiftduct::TB = true : flag as to whether duct geometry should be shifted based on rotor tip location

lu_decomp_flag::TB = false : flag indicating if panel method LHS matrix factorization was successful

paneling options

itcpshift::TF = 0.05 : factor for internal trailing edge psuedo-panel placement (default is DFDC hard-coded value)

axistol::TF = 1e-15 : tolerance for how close the the axis of rotation should be considered on the axis

tegaptol::TF = 1e1 * eps() : tolerance for how large of a trailing edge gap should be considered a gap

Integration Options

integration_options::TIo = IntegrationOptions() : integration options

Post-processing Options

write_outputs::TBwo = [false] : Bool for whether to write the outputs of the analysis to an external file (slow)

outfile::TSf = ["outputs.jl"] : External output file name (including path information) for files to write

checkoutfileexists::TB = false : Flag for whether to check if file exists before overwriting

output_tuple_name::TSt = ["outs"] : variable name for named tuple written to out file

Solving Options

grid_solver_options::WS = GridSolverOptions() : elliptic grid solver options

solver_options::TSo = ChainSolverOptions() : solver options

DuctAPE.set_options — Function

set_options(; kwargs...)
set_options(multipoint; kwargs...)

Set the options for DuctAPE to use.

Note that the vast majority of the available options are defined through keyword arguments. See the documentation for the various option types for more information.

Arguments

multipoint::AbstractVector{OperatingPoint} : a vector of operating points to use if running a multi-point analysis.

Integration Options

DuctAPE.IntegrationOptions — Type

struct IntegrationOptions{TN<:IntegrationMethod,TS<:IntegrationMethod}

A struct used to hold the integration options for both the nominal and singular cases.

Fields

nominal::IntegrationMethod=GaussLegendre(8) : the integration options to use for the nominal case.

singular::IntegrationMethod=GaussLegendre(8) : the integration options to use for the self-induced case.

DuctAPE.GaussLegendre — Type

struct GaussLegendre{TN,TW} <: IntegrationMethod

Options for Gauss-Legendre integration method

Fields

sample_points::TN : Sample Points

weights::TW : Gauss weights

DuctAPE.GaussKronrod — Type

struct GaussKronrod{TF,TI} <: IntegrationMethod

Options for Gauss-Kronrod integration method

Fields

order::TI = 7 : order of Legendre polynomial to use on each interval

maxevales::TI = 10^7 : maximum number of evaluations in the adaptive method

atol::TF = 0.0 : absolute error tolerance. (note, if zero, QuadGK uses sqrt(eps()) relative tolerance).

DuctAPE.Romberg — Type

struct Romberg{TF,TI} <: IntegrationMethod

Options for Romberg integration method

Fields

max_subdivisions::TI = 10 : maximum number of subdivisions. Note, total number of internvals is 2^N, where N is number of subdivisions.

atol::TF = 1e-6 : absolute error tolerance.

Solver Options

Elliptic Grid Solve

DuctAPE.SLORGridSolverOptions — Type

struct SLORGridSolverOptions{TB,TF,TI} <: GridSolverOptionsType

Options for SLOR (successive line over relaxation) elliptic grid solver.

Fields

iteration_limit::TI = 100 : maximum number of iterations

atol::TF = 1e-9 : absolute convergence tolerance

`converged::AbstractVector{TB} = [false]

DuctAPE.GridSolverOptions — Type

struct GridSolverOptions{TB,TF,TI,TSym} <: GridSolverOptionsType

Options for Newton elliptic grid solver.

Fields

iteration_limit::TI = 10 : maximum number of iterations

atol::TF = 1e-14 : absolute convergence tolerance

algorithm::TSym = :newton : algorithm to use in NLsolve.jl

autodiff::TSym = :forward : differentiation method to use in NLsolve.jl

`converged::AbstractVector{TB} = [false]

Aero Solve

DuctAPE.ChainSolverOptions — Type

struct ChainSolverOptions{TB,TS<:Union{ExternalSolverOptions,PolyAlgorithmOptions}} <:PolyAlgorithmOptions

Options for Chain Solvers (try one solver, if it doesn't converge, try another)

Fields

`solvers::AbstractVector{TS} = [ NLsolveOptions(; algorithm=:anderson, atol=1e-12), MinpackOptions(; atol=1e-12), NonlinearSolveOptions(; algorithm=SimpleNonlinearSolve.SimpleNewtonRaphson, atol=1e-12, additional_kwargs=(; autodiff=SimpleNonlinearSolve.AutoForwardDiff()), ), ] : Vector of solver options to use.

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.CompositeSolverOptions — Type

struct CompositeSolverOptions{
    TB,TS<:Union{ExternalSolverOptions,PolyAlgorithmOptions}
} <: PolyAlgorithmOptions

Options for Composite Solvers (start with a partial solve of one solve, then finish with another starting where the first left off).

Fields

`solvers::AbstractVector{TS} = [ NLsolveOptions(; algorithm=:newton, iteration_limit=3), NLsolveOptions(; algorithm=:anderson, atol=1e-12), ]' : Vector of solver options to use.

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.NLsolveOptions — Type

struct NLsolveOptions{TB,TF,TK,Tls,Tlsk,TSym} <: ExternalSolverOptions

Options for the NLsolve pacakge solvers

Fields

algorithm::TSym = :anderson : algorithm to use

additional_kwargs::TK = (;) : any additional keyword arguments for the solver

atol::TF = 1e-12 : absolute convergence tolerance

iteration_limit::TF = 25 : maximum number of iterations

linesearch_method::Tls = LineSearches.MoreThuente : line search method to use

linesearch_kwargs::Tlsk = (;) : any additional lineseach keyword arguments

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.NonlinearSolveOptions — Type

struct NonlinearSolveOptions{TA,TB,TF,TI,TT} <: ExternalSolverOptions

Options for the SimpleNonlinearSolve pacakge solvers

Fields

algorithm::TA = SimpleNonlinearSolve.SimpleNewtonRaphson : algorithm to use

additional_kwargs::TK = (;) : any additional keyword arguments for the solver

atol::TF = 1e-12 : absolute convergence tolerance

iteration_limit::TF = 25 : maximum number of iterations

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.MinpackOptions — Type

struct MinpackOptions{TB,TF,TI,TSym} <: ExternalSolverOptions

Options for the MINPACK's HYBRJ solver

Fields

algorithm::TSym = :hybr : algorithm to use in MINPACK.jl (hybr is HYBRJ when the jacobian is provided)

atol::TF = 1e-12 : absolute convergence tolerance

iteration_limit::TF = 100 : maximum number of iterations

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.SIAMFANLEOptions — Type

struct SIAMFANLEOptions{TA,TB,TF,TI,TK} <: ExternalSolverOptions

Options for the SIAMFANLEquations pacakge solvers

Fields

algorithm::TA = SIAMFANLEquations.nsoli : algorithm to use

rtol::TF = 0.0 : relative convergence tolerance

atol::TF = 1e-10 : absolute convergence tolerance

iteration_limit::TF = 1000 : maximum number of iterations

linear_iteration_limit::TF = 5 : maximum number of linear solve iterations (GMRES)

additional_kwargs::TK = (;) : any additional keyword arguments for the solver

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.SpeedMappingOptions — Type

struct SpeedMappingOptions{TB,TF,TI,TL,TSm,TU} <: ExternalSolverOptions

Options for the SpeedMapping.jl package solver

Fields

`orders::AbstractVector{TI} = [3, 2]

sig_min::TSm = 0 : maybe set to 1?

stabilize::TB = false : stabilizes before extrapolation

check_obj::TB = false : checks for inf's and nan's and starts from previous finite point

atol::TF = 1e-10 : absolute convergence tolerance

iteration_limit::TF = 1000 : maximum number of iterations

time_limit::TF = Inf : time limit in seconds

lower::TL = nothing : box lower bounds

upper::TU = nothing : box upper bounds

buffer::TF = 0.01 : if using bounds, buffer brings x inside bounds by buffer amountd

Lp::TF = Inf : p value for p-norm for convergence criteria

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.FixedPointOptions — Type

struct FixedPointOptions{TB,TF,TI} <: ExternalSolverOptions

Options for the FixedPoint.jl package solver

Fields

iteration_limit::TF = 1000 : maximum number of iterations

vel::TF = 0.9 : vel keyword argument, default is package default

ep::TF = 0.01 : ep keyword argument, default is package default

atol::TF = 1e-12 : absolute convergence tolerance

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.CSORSolverOptions — Type

struct CSORSolverOptions{TB,TC<:ConvergenceType,TF,TS} <: SolverOptionsType

Type containing all the options for the CSOR (controlled successive over relaxation) solver.

Note that the defaults match DFDC with the exception of the relaxation schedule, which is an experimental feature.

Fields

var::type :

verbose::TB = false : flag to print verbose statements

iteration_limit::TF = 1e2 : maximum number of iterations

nrf::TF = 0.4 : nominal relaxation factor

bt1::TF = 0.2 : backtracking factor 1

bt2::TF = 0.6 : backtracking factor 2

pf1::TF = 0.4 : press forward factor 1

pf2::TF = 0.5 : press forward factor 2

btw::TF = 0.6 : backtracking factor for wake

pfw::TF = 1.2 : press forward factor for wake

relaxation_schedule::TS = [[0.0;1e-14;1e-13;1e10]), [1.0;1.0;0.0;0.0])] : values used in spline definition for scaling the relaxation factors (second vector) after various convergence values (first vector).

f_circ::TF = 1e-3 : convergence tolerance for rotor circulation

f_dgamw::TF = 2e-4 : convergence tolerance for wake vortex strength

convergence_type::TC = Relative() : dispatch for relative or absolute convergence criteria.

Vconv::TF = 1.0 : velocity used in relative convergence criteria (should be set to Vref).

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

Analysis

DuctAPE.analyze — Function

analyze(
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze propulsor, including preprocessing.

Arguments

propulsor::Propulsor : Propulsor input object (see docstring for Propulsor type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options=set_options();
    return_inputs=false,
    solve_container_caching=nothing,
)

Analyze propulsor, assuming setup_analysis has been called and the outputs thereof are being passed in here.

Arguments

propulsor::Propulsor : Propulsor input object

prepost_containers::NamedTuple : An output from setup_analysis containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : An output from setup_analysis containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : An output from setup_analysis containing dimensions used for reshaping the solve parameter cache

airfoils::Vector{AFType} : An output from setup_analysis contiaining the blade element airfoil polar objects

A_bb_LU::LinearAlgebra.LU : An output from setup_analysis that is the LU decomposition of the AIC matrix used in the panel method

idmaps::NamedTuple : An output from setup_analysis containing bookkeeping information (index mappings)

problem_dimensions::NamedTuple : An output from setup_analysis contiaining bookkeeping information (problem dimensions)

options::Options=set_options() : Options object

Keyword Arguments

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    multipoint::AbstractVector{OperatingPoint},
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze propulsor, including preprocessing, for a set of operating points.

Arguments

multipoint::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the propulsor (note that the operating point within the propulsor input will be overwritten with these)

propulsor::Propulsor : Propulsor input object

options::Options=set_options() : Options object

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Vector of named tuples of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    multipoint::Vector{OperatingPoint},
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options=set_options();
    return_inputs=false,
    solve_container_caching=nothing,
)

Analyze propulsor, assuming setup_analysis has been called and the inputs are being passed in here.

Arguments

multipoint::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the propulsor (note that the operating point within the propulsor input will be overwritten with these)

propulsor::Propulsor : Propulsor input object

prepost_containers::NamedTuple : An output from setup_analysis containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : An output from setup_analysis containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : An output from setup_analysis containing dimensions used for reshaping the solve parameter cache

airfoils::Vector{AFType} : An output from setup_analysis contiaining the blade element airfoil polar objects

A_bb_LU::LinearAlgebra.LU : An output from setup_analysis that is the LU decomposition of the AIC matrix used in the panel method

idmaps::NamedTuple : An output from setup_analysis containing bookkeeping information (index mappings)

problem_dimensions::NamedTuple : An output from setup_analysis contiaining bookkeeping information (problem dimensions)

options::Options=set_options() : Options object

Keyword Arguments

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true. Note that some inputs will be overwritten (e.g. the linear system RHS components related to the freestream) and only those associated with the final operating point will be returned.

convergence_flag : Flag for successful solve convergence

Preprocess

DuctAPE.setup_analysis — Function

setup_analysis(
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
)

Perform pre-processing and cache setup (as needed) for propuslor analysis.

Arguments

propulsor::Propulsor : Propulsor input object (see docstring for Propulsor type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

Returns

problem_dimensions::NamedTuple : Named Tuple contiaining bookkeeping information (problem dimensions)

prepost_containers::NamedTuple : Named Tuple containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : Vector containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : Named Tuple containing dimensions used for reshaping the solve parameter cache

A_bb_LU::LinearAlgebra.LU : The LU factorization of the AIC matrix used in the panel method

lu_decomp_flag::Bool : flag indicating if the LU decomposition was successful

airfoils::Matrix{AFType} : Matrix contiaining the blade element airfoil polar objects

idmaps::NamedTuple : Named Tuple containing bookkeeping information (index mappings)

Process

Postprocess

Private API

Option Types

DuctAPE.DFDC_options — Function

function DFDC_options(;
    grid_solver_options=SLORGridSolverOptions(),
    solver_options=CSORSolverOptions(),
    kwargs...,
)

Convenience function to select options used in DFDC.

Arguments

grid_solver_options=SLORGridSolverOptions() : elliptic grid solver options

solver_options=CSORSolverOptions() : solver options

DuctAPE.ConvergenceType — Type

abstract type ConvergenceType

Used in dispatching the CSOR (controlled successive over relaxation) residual as relative or absolute.

DuctAPE.Relative — Type

struct Relative <: ConvergenceType

Used to dispatch the relative residual for CSOR (controlled successive over relaxation) method

DuctAPE.Absolute — Type

struct Absolute <: ConvergenceType

Used to dispatch the absolute residual for CSOR (controlled successive over relaxation) method

DuctAPE.SolverOptionsType — Type

abstract type SolverOptionsType

Used for solver dispatch.

DuctAPE.ExternalSolverOptions — Type

abstract type ExternalSolverOptions <: SolverOptionsType

Used for solver dispatch.

DuctAPE.PolyAlgorithmOptions — Type

abstract type PolyAlgorithmOptions <: SolverOptionsType

Used for solver dispatch.

DuctAPE.GridSolverOptionsType — Type

abstract type GridSolverOptionsType

Used for elliptic grid solver dispatch

DuctAPE.IntegrationMethod — Type

abstract type IntegrationMethod

Used in integration method dispatch

Bookkeeping

DuctAPE.get_problem_dimensions — Function

get_problem_dimensions(paneling_constants::PanelingConstants)
get_problem_dimensions(body_vortex_panels, rotor_source_panels, wake_vortex_panels)

Determine all relevant dimensions to the problem based either on the paneling_constants or the panels themselves.

Arguments

paneling_constants::PanelingConstants : Rotor (and possibly stator) geometric paramters.

Returns

problem_dimensions::ProblemDimensions : ProblemDimensions object.

Caching

Allocation

The following are various helper functions used in preallocating the various caches.

DuctAPE.allocate_wake_panel_container! — Function

allocate_wake_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

wake_vortex_panels::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to the wake vortex panel object

DuctAPE.allocate_panel_containers! — Function

allocate_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

panels::NamedTuple : A named tuple of named tuples containing the dimensions needed to reshape the cache with regards to the panel objects

DuctAPE.allocate_panel_container! — Function

allocate_panel_container!(total_length, nn, np, tn, tp, nb)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

nn::Int : number of nodes in each body, rotor, or wake sheet

np::Int : number of panels in each body, rotor, or wake sheet

tn::Int : number of total nodes among the bodies, rotors, or wake sheets

tp::Int : number of total panels among the bodies, rotors, or wake sheets

nb::Int : number of bodies, rotors, or wake sheets

Returns

panel::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to an arbitrary panel set

DuctAPE.allocate_body_panel_container! — Function

allocate_body_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

body_vortex_panels::NamedTuple : A named tuple containing the dimensions needed to reshape the cache with regards to the body vortex panel object

DuctAPE.allocate_rotor_panel_container! — Function

allocate_rotor_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

rotor_source_panels::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to the rotor source panel object

DuctAPE.allocate_solve_parameter_extras! — Function

allocate_solve_parameter_extras!(
    solver_options::SolverOptionsType, input_length, total_length
)

Includes additional caching for various solvers. Currently only does anything for SIAMFANLEOptions types.

Arguments

input_length::Int : the number of state variables in the solver

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

Returns

solve_parameter_extras::NamedTuple : A named tuple containing dimensions related to extra caching parameters used in various solvers.

Reshaping

The following are used internally to reshape the cache vectors into more usable formats.

DuctAPE.withdraw_prepost_container_cache — Function

withdraw_prepost_container_cache(vec, dims)

Reshape the prepost cache vector using the saved dimensions tuple.

Arguments

vec::Vector{Float} : vector cache of pre- and post-processing intermediate containers.

dims::NamedTuple : Named tuple containing the indices and shape of the various items stored in the cache vector.

Returns

prepost_container_caching::NamedTuple : Named tuple containing reshaped views of sections of the cache vector.

DuctAPE.withdraw_solve_parameter_cache — Function

withdraw_solve_parameter_cache(solver_options::SolverOptionsType, vec, dims)

Reshape the solve parameter cache vector using the saved dimensions tuple.

Arguments

solver_options::SolverOptionsType : Solver options type for dispatch.

vec::Vector{Float} : vector cache of pre- and post-processing intermediate containers.

dims::NamedTuple : Named tuple containing the indices and shape of the various items stored in the cache vector.

Returns

solve_parameter_caching::NamedTuple : Named tuple containing reshaped views of sections of the cache vector.

DuctAPE.withdraw_solve_container_cache — Function

withdraw_solve_container_cache(solver_options::SolverOptionsType, vec, dims)

Reshape the intermediate solve container cache vector using the saved dimensions tuple.

Arguments

solver_options::SolverOptionsType : Solver options type for dispatch.

vec::Vector{Float} : vector cache of pre- and post-processing intermediate containers.

dims::NamedTuple : Named tuple containing the indices and shape of the various items stored in the cache vector.

Returns

solve_container_caching::NamedTuple : Named tuple containing reshaped views of sections of the cache vector.

Preprocess

General

DuctAPE.set_index_maps — Function

set_index_maps(
    npanels,
    ncenterbody_inlet,
    nwake_sheets,
    dte_minus_cbte,
    wnm,
    wenids,
    nwp,
    nwsp,
    nbn,
    ndp,
    riiw,
    nrotor,
)

Set values for index map to be used throughout solve and post-process.

Arguments

npanels : paneling_constants.npanels

ncenterbody_inlet : paneling_constants.ncenterbody_inlet

nwake_sheets : paneling_constants.nwake_sheets

dte_minus_cbte : paneling_constants.dte_minus_cbte

wnm : wake_vortex_panels.nodemap

wenids : wake_vortex_panels.endnodeidxs

nwp : problem_dimensions.nwp

nwsp : problem_dimensions.nwsp

nbn : problem_dimensions.nbn

ndp : body_vortex_panels.npanel[1]

riiw : rotor_indices_in_wake

nrotor : problem_dimensions.nrotor

Returns

idmaps::NamedTuple : A named tuple containing index mapping used in bookkeeping throughout solve and post-process

DuctAPE.precompute_parameters — Function

precompute_parameters(
    propulsor;
    grid_solver_options=GridSolverOptions(),
    integration_options=IntegrationOptions(),
    autoshiftduct=true,
    itcpshift=0.05,
    axistol=1e-15,
    tegaptol=1e1 * eps(),
    finterp=FLOWMath.akima,
    silence_warnings=true,
    verbose=false,
)

Out of place main pre-processing function that computes all the required parameters for the solve.

Arguments

propulsor::Propulsor : A Propuslor object

Keyword Arguments

grid_solver_options::GridSolverOptionsType=GridSolverOptions() : A GridSolverOptionsType object

integration_options::IntegrationMethod=IntegrationOptions() : An IntegrationMethod object

autoshiftduct::Bool=true : flag to shift duct geometry based on rotor tip radius

itcpshift::Float=0.05 : value used in positioning the internal pseudo control point in the solid bodies. Default is DFDC hard-coded value.

axistol::Float=1e-15 : tolerance for how close to the axis of rotation to be considered on the axis.

tegaptol::Float=1e1 * eps() : tolerance for how large of a trailing edge gap is considered a gap.

finterp::Function=FLOWMath.akima : interpolation method for re-interpolating body coordinates

silence_warnings::Bool=true : flag to silence warnings

verbose::Bool=false : flag to print verbose statements

Returns

ivr::NamedTuple : A named tuple containing arrays of induced velocities on the rotors

ivw::NamedTuple : A named tuple containing arrays of induced velocities on the wake

ivb::NamedTuple : A named tuple containing arrays of induced velocities on the bodies

linsys::NamedTuple : A named tuple containing cacheable data for the linear system, including:

A_bb::Array{Float} : AIC (LHS) matrix for the panel method system

b_bf::Array{Float} : Initial system RHS vector based on freestrem magnitude

A_br::Array{Float} : Unit normal velocity from rotors onto body panels

A_pr::Array{Float} : Unit normal velocity from rotors onto body internal psuedo control points

A_bw::Array{Float} : Unit normal velocity from wake onto body panels

A_pw::Array{Float} : Unit normal velocity from wake onto body internal psuedo control points

A_bb_LU::LinearAlgebra.LU : LinearAlgebra LU factorization of the LHS matrix

lu_decomp_flag::Vector{Bool} : flag for whether factorization was successful

blade_elements::NamedTuple : A named tuple containing cacheable blade element information (see docs for interpolate_blade_elements)

airfoils::Vector{AFType} : A matrix of airfoil types associated with each of the blade elements

wakeK::Matrix{Float} : A matrix of precomputed geometric constants used in the calculation of the wake vortex strengths

idmaps::NamedTuple : A named tuple containing index mapping used in bookkeeping throughout solve and post-process

panels::NamedTuple : A named tuple of panel objects including:

body_vortex_panels::NamedTuple : the named tuple containing the body vortex panel information

rotor_source_panels::NamedTuple : the named tuple containing the rotor source panel information

wake_vortex_panels::NamedTuple : the named tuple containing the wake vortex panel information

problem_dimensions::ProblemDimensions : A ProblemDimensions object

precompute_parameters(
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    wake_grid,
    rotor_indices_in_wake,
    Rtips,
    Rhubs,
    rotorstator_parameters,
    paneling_constants,
    operating_point,
    integration_options,
    problem_dimensions=nothing;
    itcpshift=0.05,
    axistol=1e-15,
    tegaptol=1e1 * eps(),
    silence_warnings=true,
    verbose=false,
)

An alternate version of precompute_parameters allowing for user defined geometry that does not go through a re-panling step (use with caution).

The first inputs are the outputs of the reinterpolate_geometry and get_blade_ends_from_body_geometry functions.

DuctAPE.precompute_parameters! — Function

precompute_parameters!(
    ivr,
    ivw,
    blade_element_cache,
    linsys,
    wakeK,
    propulsor,
    prepost_containers,
    problem_dimensions;
    grid_solver_options=GridSolverOptions(),
    integration_options=IntegrationOptions(),
    autoshiftduct=true,
    itcpshift=0.05,
    axistol=1e-15,
    tegaptol=1e1 * eps(),
    finterp=fm.akima,
    silence_warnings=true,
    verbose=false,
)

In-place version of nominal version of precompute_parameters.

precompute_parameters!(
    ivr,
    ivw,
    blade_element_cache,
    linsys,
    wakeK,
    wake_grid,
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    rotor_indices_in_wake,
    rotorstator_parameters,
    paneling_constants,
    operating_point,
    prepost_containers,
    problem_dimensions=nothing;
    integration_options=IntegrationOptions(),
    itcpshift=0.05,
    axistol=1e-15,
    tegaptol=1e1 * eps(),
    finterp=fm.akima,
    silence_warnings=true,
    verbose=false,
)

In-place version of the precompute_parameters function by-passing the geometry reinterpolateion. (Use with caution)

Geometry

DuctAPE.reinterpolate_geometry — Function

reinterpolate_geometry(
    problem_dimensions,
    duct_coordinates,
    centerbody_coordinates,
    rotorstator_parameters,
    paneling_constants;
    autoshiftduct=true,
    grid_solver_options=GridSolverOptions(),
    finterp=FLOWMath.akima,
    verbose=false,
    silence_warnings=true,
)

Re-interpolate the body geometry and return compatible body and way geometry.

Arguments

problem_dimensions::ProblemDimensions : A ProblemDimensions object

duct_coordinates::Matrix{Float} : [z,r] coordinates of duct geometry

centerbody_coordinates::Matrix{Float} : [z,r] coordinates of centerbody geometry

rotorstator_parameters::RotorStatorParameters : A RotorStatorParameters object

paneling_constants::PanelingConstants : A PanelingConstants object

Keyword Arguments

autoshiftduct::Bool=true : flag to shift duct geometry based on rotor tip radius

grid_solver_options::SolverOptionsType=GridSolverOptions() : options for the wake grid position solver

finterp::Function=FLOWMath.akima : interpolation method for re-interpolating body coordinates

verbose::Bool=false : flag to print verbose statements

silence_warnings::Bool=true : flag to silence warnings

Returns

wake_grid::Array{Float} : array containig the z and r elliptic grid points defning the wake geometry.

rp_duct_coordinates::Matrix{Float} : matrix containing the re-paneled duct coordinates

rp_centerbody_coordinates::Matrix{Float} : matrix containing the re-paneled centerbody coordinates

rotor_indices_in_wake::Vector{Int} : vector containing the indices of where in the wake the rotors reside (used later to define the rotor panel edges).

DuctAPE.reinterpolate_geometry! — Function

reinterpolate_geometry!(
    wake_grid,
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    rotor_indices_in_wake,
    duct_coordinates,
    centerbody_coordinates,
    rotorstator_parameters,
    blade_element_cache,
    paneling_constants;
    autoshiftduct=true,
    grid_solver_options=GridSolverOptions(),
    finterp=FLOWMath.akima,
    verbose=false,
    silence_warnings=true,
)

In-place version of reinterpolate_geometry.

DuctAPE.generate_all_panels — Function

generate_all_panels(
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    nwake_sheets,
    rotor_indices_in_wake,
    rotorzloc,
    wake_grid;
    itcpshift=0.05,
    axistol=1e-15,
    tegaptol=1e1 * eps(),
    silence_warnings=true,
)

Function that calls all of the various panel generation functions are returns a named tuple containing all the panels

Arguments

rp_duct_coordinates::Matrix{Float} : matrix containing the re-paneled duct coordinates

rp_centerbody_coordinates::Matrix{Float} : matrix containing the re-paneled centerbody coordinates

nwake_sheets::Int : number of wake sheets

rotor_indices_in_wake::Vector{Int} : vector containing the indices of where in the wake the rotors reside (used later to define the rotor panel edges).

rotorzloc:Vector{Float} : axial locations of rotor lifting lines (contained in RotorStatorParameters)

wake_grid::Array{Float} : array containig the z and r elliptic grid points defning the wake geometry.

Keyword Arguments

itcpshift::Float=0.05 : value used in positioning the internal pseudo control point in the solid bodies. Default is DFDC hard-coded value.

axistol::Float=1e-15 : tolerance for how close to the axis of rotation to be considered on the axis.

tegaptol::Float=1e1 * eps() : tolerance for how large of a trailing edge gap is considered a gap.

silence_warnings::Bool=true : flag to silence warnings

Returns

panels::NamedTuple : A named tuple of named tuples containing paneling information, including:

body_vortex_panels::NamedTuple

rotor_source_panels::NamedTuple

wake_vortex_panels::NamedTuple

DuctAPE.generate_all_panels! — Function

generate_all_panels!(
    panels,
    wake_grid,
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    rotor_indices_in_wake,
    rotorzloc,
    nwake_sheets;
    itcpshift=0.05,
    axistol=1e-15,
    tegaptol=1e1 * eps(),
    silence_warnings=true,
)

In-place version of generateallpanels.

Wake

DuctAPE.discretize_wake — Function

discretize_wake(
    duct_coordinates,
    centerbody_coordinates,
    rotorzloc, # rotor axial locations
    wake_length,
    npanels,
    dte_minus_cbte;
)

Calculate wake sheet panel node z-coordinates.

Arguments

duct_coordinates::Matrix{Float} : Array of input duct coordinates

centerbody_coordinates::Matrix{Float} : Array of input centerbody_coordinates coordinates

rotorzloc ::Vector{Float} : rotor axial locations

wake_length::Float : non-dimensional length of wake to extend beyond aft-most body trailing edge.

npanels::Vector{Int} : A vector of the number of panels between each discrete point. For example: [number of panels between the rotors; number of panels between the stator and the first trailing edge; number of panels between the trailing edges; number of panels between the last trailing edge and the end of the wake]

dte_minus_cbte::Float : indicator as to whether the duct trailing edge minus the centerbody trailing edge is positive, zero, or negative.

DuctAPE.generate_wake_grid — Function

generate_wake_grid(
    problem_dimensions,
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    Rhub1,
    Rtip1,
    tip_gap1,
    zwake;
    grid_solver_options=GridSolverOptions(),
    verbose=false,
    silence_warnings=true,
)

Initialize and solve for elliptic grid on which wake sheets are defined.

Arguments

problem_dimensions:: : A ProblemDimensions object

rp_duct_coordinates:: : repaneled duct coordinates

rp_centerbody_coordinates:: : repaneled centerbody coordinates

Rhub1:: : Hub radius of first rotor

Rtip1:: : Tip radius of first rotor

tip_gap1:: : Tip gap of first rotor (MUST BE ZERO for now)

zwake:: : axial positions of wake sheet panel nodes

Keyword Arguments

grid_solver_options::GridSolverOptionsType=GridSolverOptions() : options for solving the elliptic grid.

verbose::Bool=false : flag to print verbose statements

silence_warnings::Bool=true : flag to supress warnings

Returns

wake_grid::Array{Float,3} : 3D Array of axial and radial wake_grid points after solution of elliptic system.

DuctAPE.generate_wake_grid! — Function

generate_wake_grid!(
    wake_grid,
    rp_duct_coordinates,
    rp_centerbody_coordinates,
    Rhub1,
    Rtip1,
    tip_gap1,
    zwake;
    grid_solver_options=grid_solver_options,
    verbose=false,
    silence_warnings=true,
)

In-place version of generatewakegrid.

DuctAPE.initialize_wake_grid — Function

initialize_wake_grid(rp_duct_coordinates, rp_centerbody_coordinates, zwake, rwake)

Intialize the wake grid.

Arguments:

rp_duct_coordinates::Matrix{Float} : The re-paneled duct coordinates

rp_centerbody_coordinates::Matrix{Float} : The re-paneled centerbody coordinates

zwake::Vector{Float} : The axial positions of the wake sheet panel nodes

rwake::Vector{Float} : The radial positions of the blade elements for the foremost rotor

Returns:

wake_grid::Array{Float,3} : 3D Array of axial and radial wake_grid points

DuctAPE.initialize_wake_grid! — Function

initialize_wake_grid!(
    wake_grid, rp_duct_coordinates, rp_centerbody_coordinates, zwake, rwake
)

In-place version of initializewakegrid.

DuctAPE.relax_grid! — Function

relax_grid!(
    grid_solver_options::GridSolverOptionsType,
    wake_grid;
    verbose=false,
    silence_warnings=true,
    tabchar="    ",
    ntab=1,
)

Relax/Solve initial wake grid according to elliptic system of equations.

Arguments

`gridsolveroptions::GridSolverOptionsType' : options for elliptic grid solver

wake_grid::Array{Float,3} : Initialized wake grid

Keyword Arguments

`verbose=false::' : flag for printing verbose statements

`silence_warnings=true::' : flag for supressing warnings

`tabchar::String=" "::' : string to use for tabbing over verbose statements.

`ntab::Int=1' : number of tabs for printing verbose statements

relax_grid!(xg, rg, nxi, neta; iteration_limit, atol)

Relax wakegrid using elliptic wakegrid solver.

Arguments:

Keyword Arguments:

iteration_limit::Int : maximum number of iterations to run, default=100

atol::Float : convergence tolerance, default = 1e-9

Returns:

DuctAPE.generate_wake_panels — Function

generate_wake_panels(wake_grid)

Generate paneling for each wake sheet emanating from the rotor blade elements.

Arguments:

wake_grid::Array{Float,3} : axial and radial locations of each wake_grid point (after relaxation/solution)

Returns:

wake_vortex_panels::NamedTuple : A named tuple of panel values describing the wake vortex panels

DuctAPE.generate_wake_panels! — Function

generate_wake_panels!(wake_panels, wake_grid)

In-place version of generatewakepanels.

DuctAPE.get_wake_k — Function

get_wake_k(r, nwn)

Calculate geometric constant for use in later calculation of wake panel node strengths.

Arguments

r::Vector{Float} : Vector of wake panel node radial positions

Returns

K::Vector{Float} : Vector of geometric constants used in calculation of panel node strengths.

DuctAPE.get_wake_k! — Function

get_wake_k!(K, r)

In-place version of getwakek.

Bodies

Rotors

Induced Velocities

DuctAPE.calculate_unit_induced_velocities — Function

calculate_unit_induced_velocities(problem_dimensions, panels, integration_options)

Calculate all the unit-induced velocties of all panels on all control points

Arguments

problem_dimensions::ProblemDimensions : A ProblemDimensions object

panels::NamedTuple : A named tuple containing all the paneling information

integration_options::IntegrationOptions : Options used for integration of velocity kernals across panels

Returns

ivr::NamedTuple : A named tuple containing arrays of induced velocities on the rotors

ivw::NamedTuple : A named tuple containing arrays of induced velocities on the wake

ivb::NamedTuple : A named tuple containing arrays of induced velocities on the bodies

DuctAPE.calculate_unit_induced_velocities! — Function

calculate_unit_induced_velocities!(ivr, ivw, ivb, panels, integration_options)

In-place version of calculateunitinduced_velocities.

DuctAPE.initialize_linear_system — Function

initialize_linear_system(
    ivb,
    body_vortex_panels,
    rotor_source_panels,
    wake_vortex_panels,
    Vinf,
    integration_options,
)

Set up the linear system used in the panel method solve.

Arguments

ivb::NamedTuple : the named tuple containing all the unit induced velocities on the bodies

body_vortex_panels::NamedTuple : the named tuple containing the body vortex panel information

rotor_source_panels::NamedTuple : the named tuple containing the rotor source panel information

wake_vortex_panels::NamedTuple : the named tuple containing the wake vortex panel information

Vinf::Vector{Float} : the one-element vector containing the Freestream velocity magnitude

integration_options::IntegrationOptions : the integration options used in integrating the panel induced velocities

Returns

linsys::NamedTuple : A named tuple containing cacheable data for the linear system, including:

A_bb::Array{Float} : AIC (LHS) matrix for the panel method system

b_bf::Array{Float} : Initial system RHS vector based on freestrem magnitude

A_br::Array{Float} : Unit normal velocity from rotors onto body panels

A_pr::Array{Float} : Unit normal velocity from rotors onto body internal psuedo control points

A_bw::Array{Float} : Unit normal velocity from wake onto body panels

A_pw::Array{Float} : Unit normal velocity from wake onto body internal psuedo control points

A_bb_LU::LinearAlgebra.LU : LinearAlgebra LU factorization of the LHS matrix

lu_decomp_flag::Vector{Bool} : flag for whether factorization was successful

DuctAPE.initialize_linear_system! — Function

initialize_linear_system!(
    linsys,
    ivb,
    body_vortex_panels,
    rotor_source_panels,
    wake_vortex_panels,
    Vinf,
    intermediate_containers,
    integration_options,
)

In-place version of initializelinearsystem.

State Initialization

TODO: still need to do everything in initializestates as well as everything in the geometry and velocities directories

Analysis

DuctAPE.analyze_multipoint — Function

analyze_multipoint(
    operating_point::OperatingPoint,
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options;
    solve_container_caching=nothing,
    return_inputs=false,
)

Identical to the single analyze function assuming setup_analysis has been called; except here we are running a single operating point for a multipoint analysis, and overwriting the operating point in the propulsor with the explicit operating point input.

Post-process

DuctAPE.post_process — Function

Convenience Functions

DuctAPE.promote_propulosor_type — Function

promote_propulosor_type(propulsor)

Convenience function for promoting types based on any potential elements of the propulsor object dependent on optimization design variables.

Arguments

propulsor::Propulsor : the propulsor input

Index

DuctAPE.Absolute

DuctAPE.CSORSolverOptions

DuctAPE.ChainSolverOptions

DuctAPE.CompositeSolverOptions

DuctAPE.ConvergenceType

DuctAPE.ExternalSolverOptions

DuctAPE.FixedPointOptions

DuctAPE.GaussKronrod

DuctAPE.GaussLegendre

DuctAPE.GridSolverOptions

DuctAPE.GridSolverOptionsType

DuctAPE.IntegrationMethod

DuctAPE.IntegrationOptions

DuctAPE.MinpackOptions

DuctAPE.NLsolveOptions

DuctAPE.NonlinearSolveOptions

DuctAPE.OperatingPoint

DuctAPE.Options

DuctAPE.PanelingConstants

DuctAPE.PolyAlgorithmOptions

DuctAPE.Propulsor

DuctAPE.ReferenceParameters

DuctAPE.Relative

DuctAPE.Romberg

DuctAPE.RotorStatorParameters

DuctAPE.SIAMFANLEOptions

DuctAPE.SLORGridSolverOptions

DuctAPE.SolverOptionsType

DuctAPE.SpeedMappingOptions

DuctAPE.DFDC_options

DuctAPE.allocate_body_panel_container!

DuctAPE.allocate_panel_container!

DuctAPE.allocate_panel_containers!

DuctAPE.allocate_prepost_container_cache

DuctAPE.allocate_rotor_panel_container!

DuctAPE.allocate_solve_container_cache

DuctAPE.allocate_solve_parameter_cache

DuctAPE.allocate_solve_parameter_extras!

DuctAPE.allocate_wake_panel_container!

DuctAPE.analyze

DuctAPE.analyze_multipoint

DuctAPE.calculate_unit_induced_velocities

DuctAPE.calculate_unit_induced_velocities!

DuctAPE.discretize_wake

DuctAPE.generate_all_panels

DuctAPE.generate_all_panels!

DuctAPE.generate_wake_grid

DuctAPE.generate_wake_grid!

DuctAPE.generate_wake_panels

DuctAPE.generate_wake_panels!

DuctAPE.get_problem_dimensions

DuctAPE.get_wake_k

DuctAPE.get_wake_k!

DuctAPE.initialize_linear_system

DuctAPE.initialize_linear_system!

DuctAPE.initialize_wake_grid

DuctAPE.initialize_wake_grid!

DuctAPE.post_process

DuctAPE.precompute_parameters

DuctAPE.precompute_parameters!

DuctAPE.promote_propulosor_type

DuctAPE.reinterpolate_geometry

DuctAPE.reinterpolate_geometry!

DuctAPE.relax_grid!

DuctAPE.set_index_maps

DuctAPE.set_options

DuctAPE.setup_analysis

DuctAPE.withdraw_prepost_container_cache

DuctAPE.withdraw_solve_container_cache

DuctAPE.withdraw_solve_parameter_cache