Analysis

Since the remaining API references here involve the inputs and outputs of the analysis function, we begin by looking at the analyze function. In general, there are two main dispatches that will be used in the vast majority of use cases. The first is for a single operating point, and the second is for multiple operating points. They are identical, except for the multiple operating point dispatch requiring a vector of operating points and returning a vector of output objects.

DuctAPE.analyze — Method

analyze(
    ducted_rotor::DuctedRotor,
    operating_point::OperatingPoint,
    reference_parameters::ReferenceParameters,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze ducted_rotor, including preprocessing.

Arguments

ducted_rotor::DuctedRotor : DuctedRotor input object (see docstring for DuctedRotor type)

operating_point::OperatingPoint : OperatingPoint input object (see docstring for OperatingPoint type)

reference_parameters::ReferenceParameters : ReferenceParameters input object (see docstring for ReferenceParameters type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

DuctAPE.analyze — Method

analyze(
    ducted_rotor::DuctedRotor,
    operating_point::AbstractVector{OperatingPoint},
    reference_parameters::ReferenceParameters,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze ducted_rotor, including preprocessing, for a set of operating points.

Arguments

ducted_rotor::DuctedRotor : DuctedRotor input object

operating_point::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the ducted_rotor

reference_parameters::ReferenceParameters : ReferenceParameters input object (see docstring for ReferenceParameters type)

options::Options=set_options() : Options object

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Vector of named tuples of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence