Option Types

DuctAPE.DFDC_options — Function

function DFDC_options(;
    grid_solver_options=SLORGridSolverOptions(),
    solver_options=CSORSolverOptions(),
    kwargs...,
)

Convenience function to select options used in DFDC.

function DFDC_options(
    multipoint;
    grid_solver_options=SLORGridSolverOptions(),
    solver_options=CSORSolverOptions(),
    kwargs...,
)

Convenience function to select options used in DFDC and run multipoint analysis.

Arguments

multipoint::Vector : doesn't need to be anything but a vector of the length of multipoints.

DuctAPE.ConvergenceType — Type

abstract type ConvergenceType

Used in dispatching the CSOR (controlled successive over relaxation) residual as relative or absolute.

DuctAPE.Relative — Type

struct Relative <: ConvergenceType

Used to dispatch the relative residual for CSOR (controlled successive over relaxation) method

DuctAPE.Absolute — Type

struct Absolute <: ConvergenceType

Used to dispatch the absolute residual for CSOR (controlled successive over relaxation) method

DuctAPE.SolverOptionsType — Type

abstract type SolverOptionsType

Used for solver dispatch.

DuctAPE.ExternalSolverOptions — Type

abstract type ExternalSolverOptions <: SolverOptionsType

Used for solver dispatch.

DuctAPE.PolyAlgorithmOptions — Type

abstract type PolyAlgorithmOptions <: SolverOptionsType

Used for solver dispatch.

DuctAPE.GridSolverOptionsType — Type

abstract type GridSolverOptionsType

Used for elliptic grid solver dispatch

DuctAPE.IntegrationMethod — Type

abstract type IntegrationMethod

Used in integration method dispatch

Bookkeeping

DuctAPE.get_problem_dimensions — Function

get_problem_dimensions(paneling_constants::PanelingConstants)
get_problem_dimensions(body_vortex_panels, rotor_source_panels, wake_vortex_panels)

Determine all relevant dimensions to the problem based either on the paneling_constants or the panels themselves.

Arguments

paneling_constants::PanelingConstants : Rotor (and possibly stator) geometric paramters.

Returns

problem_dimensions::ProblemDimensions : ProblemDimensions object.

Caching

Allocation

The following are various helper functions used in preallocating the various caches.

DuctAPE.initialize_all_caches — Function

initialize_all_caches(solver_options, paneling_constants)

Convenience function to initialize all caches before calling analysis.

Arguments

solver_options::SolverOptionsType : solver options used for cache allocation dispatch

paneling_constants::PanelingConstants : PanelingConstants object upon which all cache sizing depends

Keyword Arguments

fd_chunk_size::Int=12 : chunk size to use for PreallocationTools caches. Note that the automated chuck size for DuctAPE will always be the ForwardDiff threshold of 12 due to the size of the system, so it will be best to leave this at the default unless further development allows for chunk size selection for individual solvers.

levels::Int=1 : levels for nested duals. Note that since ImplicitAD is being used for all solves, there should be no need for more than 1 level.

Returns

prepost_container_caching::NamedTuple : A named tuple containing the PreallocationTools DiffCache and a named tuple with relevant dimensions for accessing the cache.

solve_parameter_caching::NamedTuple : A named tuple containing the PreallocationTools DiffCache and a named tuple with relevant dimensions for accessing the cache.

solve_container_caching::NamedTuple : A named tuple containing the PreallocationTools DiffCache and a named tuple with relevant dimensions for accessing the cache.

DuctAPE.allocate_wake_panel_container! — Function

allocate_wake_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

wake_vortex_panels::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to the wake vortex panel object

DuctAPE.allocate_panel_containers! — Function

allocate_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

panels::NamedTuple : A named tuple of named tuples containing the dimensions needed to reshape the cache with regards to the panel objects

DuctAPE.allocate_panel_container! — Function

allocate_panel_container!(total_length, nn, np, tn, tp, nb)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

nn::Int : number of nodes in each body, rotor, or wake sheet

np::Int : number of panels in each body, rotor, or wake sheet

tn::Int : number of total nodes among the bodies, rotors, or wake sheets

tp::Int : number of total panels among the bodies, rotors, or wake sheets

nb::Int : number of bodies, rotors, or wake sheets

Returns

panel::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to an arbitrary panel set

DuctAPE.allocate_body_panel_container! — Function

allocate_body_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

body_vortex_panels::NamedTuple : A named tuple containing the dimensions needed to reshape the cache with regards to the body vortex panel object

DuctAPE.allocate_rotor_panel_container! — Function

allocate_rotor_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

rotor_source_panels::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to the rotor source panel object

DuctAPE.allocate_solve_parameter_extras! — Function

allocate_solve_parameter_extras!(
    solver_options::SolverOptionsType, input_length, total_length
)

Includes additional caching for various solvers. Currently only does anything for SIAMFANLEOptions types.

Arguments

input_length::Int : the number of state variables in the solver

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

Returns

solve_parameter_extras::NamedTuple : A named tuple containing dimensions related to extra caching parameters used in various solvers.

DuctAPE.allocate_grid_parameter_cache — Function

allocate_grid_parameter_cache(pg, x, n)

Allocate a cache used inside the elliptic grid solve.

Arguments

pg::AbstractArray{Float,3} : the proposed grid array

x::AbstractVector{Float} : the array of ξ values used in the solve

n::AbstractVector{Float} : the array of η values used in the solve

Returns

grid_parameter_cache::NamedTuple : A named tuple containing the PreallocationTools DiffCache and dimensions for accessing it.

DuctAPE.allocate_integration_containers — Function

allocate_integration_containers(
    integration_options::IntegrationMethod, dispatch_type; cache_size=20
)

Description

Arguments

integration_options::IntegrationMethod : options for integration used for dispatch and to size cache

dispatch_type:: : an object with eltype(dispatch_type) with which to define the type for cache initialization.

Keyword Arguments

cache_size::Int=20 : size needed for intermediate calculations for integration.

Returns

integration_containers::NamedTuple : A named tuple containing the cache(s) needed for integration.

Reshaping

The following are used internally to reshape the cache vectors into more usable formats.

DuctAPE.withdraw_prepost_container_cache — Function

withdraw_prepost_container_cache(vec, dims)

Reshape the prepost cache vector using the saved dimensions tuple.

Arguments

vec::Vector{Float} : vector cache of pre- and post-processing intermediate containers.

dims::NamedTuple : Named tuple containing the indices and shape of the various items stored in the cache vector.

Returns

prepost_container_caching::NamedTuple : Named tuple containing reshaped views of sections of the cache vector.

DuctAPE.withdraw_solve_parameter_cache — Function

withdraw_solve_parameter_cache(solver_options::SolverOptionsType, vec, dims)

Reshape the solve parameter cache vector using the saved dimensions tuple.

Arguments

solver_options::SolverOptionsType : Solver options type for dispatch.

vec::Vector{Float} : vector cache of pre- and post-processing intermediate containers.

dims::NamedTuple : Named tuple containing the indices and shape of the various items stored in the cache vector.

Returns

solve_parameter_caching::NamedTuple : Named tuple containing reshaped views of sections of the cache vector.

DuctAPE.withdraw_solve_container_cache — Function

withdraw_solve_container_cache(solver_options::SolverOptionsType, vec, dims)

Reshape the intermediate solve container cache vector using the saved dimensions tuple.

Arguments

solver_options::SolverOptionsType : Solver options type for dispatch.

vec::Vector{Float} : vector cache of pre- and post-processing intermediate containers.

dims::NamedTuple : Named tuple containing the indices and shape of the various items stored in the cache vector.

Returns

solve_container_caching::NamedTuple : Named tuple containing reshaped views of sections of the cache vector.

DuctAPE.withdraw_grid_parameter_cache — Function

withdraw_grid_parameter_cache(vec, dims)

Reshape the cache used inside the elliptic grid solve.

Arguments

vec::Vector{Float} : the cache vector

dims::NamedTuple : the named tuple of dimensions used to reshape the cache vector

Returns

proposed_grid::AbstractArray{Float,3} : the proposed grid array

xi::AbstractVector{Float} : the array of ξ values used in the solve

eta::AbstractVector{Float} : the array of η values used in the solve