Public API

Public API

Input Types

Preallocations

Options

General Options

Integration Options

Solver Options

Elliptic Grid Solve

Aerodynamics Solve

Preprocess

Analysis

Miscellaneous

Airfoil/Geometry Manipulation

NACA 6-Series Cascade Geometry Generation

Input Types

DuctAPE.Propulsor — Type

Propulsor(duct_coordinates, centerbody_coordinates, rotorstator_parameters, operating_point, paneling_constants, reference_parameters)

Arguments

duct_coordinates::AbstractMatrix : The [z, r] coordinates of the duct geometry beginning at the inner (casing) side trailing edge and proceeding clockwise. Note that the duct geometry absolute radial position does not need to be included here if the autoshiftduct option is selected.

centerbody_coordinates::AbstractMatrix : The [z, r] coordinates of the centerbody beginning at the leading edge and ending at the trailing edge. Note that the leading edge is assumed to be placed at a radial distance of 0.0 from the axis of rotation.

operating_point::OperatingPoint : The operating point values.

paneling_constants::PanelingConstants : Constants used in re-paneling the geometry.

rotorstator_parameters::RotorStatorParameters : Rotor (and possibly stator) geometric paramters.

reference_parameters::ReferenceParameters : Reference Parameters.

DuctAPE.RotorStatorParameters — Type

RotorStatorParameters(
    B, rotorzloc, r, Rhub, Rtip, chords, twists, tip_gap, airfoils, fliplift
)

Composite type containing the rotor(s) geometric properties.

Note that the actual struct requires the inputs to be arrays, but there is a constructor function that will take in scalars and automatically build the array-based struct.

Arguments

B::AbstractVector{Float} : The number of blades for each rotor. May not be an integer, but usually is.

rotorzloc::AbstractVector{Float} : Dimensional, axial position of each rotor.

r::AbstractArray{Float} : Non-dimensional radial locations of each blade element.

Rhub::AbstractVector{Float} : Dimensional hub radius of rotor. (may be changed if it does not match the radial position of the centerbody geometry at the selected rotorzloc.

Rtip::AbstractVector{Float} : Dimensional tip radius of rotor. Is used to determine the radial position of the duct if the autoshiftduct option is selected.

chords::AbstractArray{Float} : Dimensional chord lengths of the blade elements.

twists::AbstractArray{Float} : Blade element angles, in radians.

tip_gap::AbstractVector{Float} : Currently unused, do not set to anything other than zeros.

airfoils::AbstractArray{AFType} : Airfoil types describing the airfoil polars for each blade element. Currently only fully tested with C4Blade.DFDCairfoil types.

fliplift::AbstractVector{Bool} : flag to indicate if the airfoil lift values should be flipped or not.

DuctAPE.OperatingPoint — Type

OperatingPoint(Vinf, rhoinf, muinf, asound, Omega)

Propulsor operating point information.

Note that the actual struct requires the inputs to be arrays, but there is a constructor function that will take in scalars and automatically build the array-based struct.

Also note that even though each field is required to be a vector, only Omega should have more than one entry, and only then if there are more than one rotor. The purpose behind having vector rather than constant scalar inputs here is for ease of redefinition in an optimization setting when freestream design variables may be present.

Arguments

Vinf::AbstractVector{Float} : Freestream velocity magnitude (which is only in the axial direction).

rhoinf::AbstractVector{Float} : Freestream density

muinf::AbstractVector{Float} : Freestream viscosity

asound::AbstractVector{Float} : Freestream speed of sound

Omega::AbstractVector{Float} : Rotor rototation rate(s)

DuctAPE.PanelingConstants — Type

PanelingConstants(
    nduct_inlet,
    ncenterbody_inlet,
    npanels,
    dte_minus_cbte,
    nwake_sheets,
    wake_length=1.0,
)

Constants used in re-paneling geometry.

Note that unlike other input structures, this one, in general, does not define fields as vectors. This is because these values should not change throughout an optimization, even if the geometry may change. Otherwise, discontinuities could be experienced.

Arguments

nduct_inlet::Int : The number of panels to use for the duct inlet (this number is used for both the casing and nacelle re-paneling)

ncenterbody_inlet::Int : The number of panels to use for the centerbody inlet.

npanels::AbstractVector{Int} : A vector containing the number of panels between discrete locations inside the wake. Specifically, the number of panels between the rotors, between the last rotor and the first body trailing edge, between the body trailing edges (if different), and between the last body trailing edge and the end of the wake. The length of this vector should be N+1 (where N is the number of rotors) if the duct and centerbody trailing edges are aligned, and N+2 if not.

dte_minus_cbte::Float : An indicator concerning the hub and duct trailing edge relative locations. Should be set to -1 if the duct trailing edge axial position minus the centerbody trailing edge axial position is negative, +1 if positive (though any positive or negative number will suffice), and zero if the trailing edges are aligned.

nwake_sheets::Int : The number of wake sheets to use. Note this will also be setting the number of blade elements to use.

wake_length::Float=1.0 : Non-dimensional (based on the length from the foremost body leading edge and the aftmost body trailing edge) length of the wake extending behind the aftmost body trailing edge.

DuctAPE.ReferenceParameters — Type

ReferenceParameters(Vref, Rref)

Reference parameters for post-process non-dimensionalization.

Note that the actual struct requires the inputs to be arrays, but there is a constructor function that will take in scalars and automatically build the array-based struct.

Arguments

Vref::AbstractVector{Float} : Reference velocity.

Rref::AbstractVector{Float} : Reference rotor tip radius.

Preallocations

DuctAPE.allocate_prepost_container_cache — Function

allocate_prepost_container_cache(paneling_constants::PanelingConstants)
allocate_prepost_container_cache(problem_dimensions::ProblemDimensions)

Allocate the pre- and post-processing cache (used for intermediate calculations) based on paneling constants or problem dimensions.

Arguments

paneling_constants::PanelingConstants : a PanelingConstants object

OR

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

prepost_container_caching::NamedTuple : a Named Tuple containing:

prepost_container_cache::PreallocationTools.DiffCache : the cache

prepost_container_cache_dims::NamedTuple : a named tuple containing the dimensions used for reshaping the cache when needed.

DuctAPE.allocate_solve_parameter_cache — Function

allocate_solve_parameter_cache(
    solve_type::SolverOptionsType,
    paneling_constants::PanelingConstants;
    fd_chunk_size=12,
    levels=1,
)
allocate_solve_parameter_cache(
    solve_type::SolverOptionsType,
    problem_dimensions::ProblemDimensions;
    fd_chunk_size=12,
    levels=1
)

Allocate the solve parameter cache for parameters passed into the solver(s).

Arguments

solve_type::SolverOptionsType : Solver options type used for dispatch

paneling_constants::PanelingConstants : a PanlingConstants object used for sizing

OR

problem_dimensions::ProblemDimensions : a ProblemDimensions object used for sizing

Keyword Arguments

fd_chunk_size::Int=12 : chunk size to use for PreallocationTools caches. Note that the automated chuck size for DuctAPE will always be the ForwardDiff threshold of 12 due to the size of the system, so it will be best to leave this at the default unless further development allows for chunk size selection for individual solvers.

levels::Int=1 : levels for nested duals. Note that since ImplicitAD is being used for all solves, there should be no need for more than 1 level.

Returns

solve_parameter_caching::NamedTuple : a Named Tuple containing:

solve_parameter_cache::PreallocationTools.DiffCache : the cache

solve_parameter_cache_dims::NamedTuple : a named tuple containing the dimensions used for reshaping the cache when needed.

DuctAPE.allocate_solve_container_cache — Function

allocate_solve_container_cache(
    solve_type::SolverOptionsType,
    paneling_constants::PanelingConstants;
    fd_chunk_size=12,
    levels=1,
)
allocate_solve_container_cache(
    solve_type::SolverOptionsType,
    problem_dimensions::ProblemDimensions;
    fd_chunk_size=12,
    levels=1,
)

Allocate the solve cache (used for intermediate calculations) based on paneling constants or problem dimensions.

Arguments

paneling_constants::PanelingConstants : a PanelingConstants object

OR

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Keyword Arguments

fd_chunk_size::Int=12 : chunk size to use for PreallocationTools caches. Note that the automated chuck size for DuctAPE will always be the ForwardDiff threshold of 12 due to the size of the system, so it will be best to leave this at the default unless further development allows for chunk size selection for individual solvers.

levels::Int=1 : levels for nested duals. Note that since ImplicitAD is being used for all solves, there should be no need for more than 1 level.

Returns

solve_container_caching::NamedTuple : a Named Tuple containing:

solve_container_cache::PreallocationTools.DiffCache : the cache

solve_container_cache_dims::NamedTuple : a named tuple containing the dimensions used for reshaping the cache when needed.

Options

General Options

DuctAPE.Options — Type

struct Options{
    TB,
    TBwo,
    TF,
    TI,
    TSf,
    TSt,
    Tin,
    TIo<:IntegrationOptions,
    TSo<:SolverOptionsType,
    WS<:GridSolverOptionsType,
}

Type containing (nearly) all the available user options.

Fields

General Options

verbose::TB = false : flag to print verbose statements

silence_warnings::TB = true : flag to silence warnings

multipoint_index::TI = [1] : holds current index of multi-point solver (no need for user to change this usually)

Geometry Re-interpolation and generation options` :

finterp::Tin = FLOWMath.akima : interpolation method used for re-paneling bodies

autoshiftduct::TB = true : flag as to whether duct geometry should be shifted based on rotor tip location

lu_decomp_flag::TB = false : flag indicating if panel method LHS matrix factorization was successful

paneling options

itcpshift::TF = 0.05 : factor for internal trailing edge psuedo-panel placement (default is DFDC hard-coded value)

axistol::TF = 1e-15 : tolerance for how close the the axis of rotation should be considered on the axis

tegaptol::TF = 1e1 * eps() : tolerance for how large of a trailing edge gap should be considered a gap

Integration Options

integration_options::TIo = IntegrationOptions() : integration options

Post-processing Options

write_outputs::TBwo = [false] : Bool for whether to write the outputs of the analysis to an external file (slow)

outfile::TSf = ["outputs.jl"] : External output file name (including path information) for files to write

checkoutfileexists::TB = false : Flag for whether to check if file exists before overwriting

output_tuple_name::TSt = ["outs"] : variable name for named tuple written to out file

Solving Options

grid_solver_options::WS = GridSolverOptions() : elliptic grid solver options

solver_options::TSo = ChainSolverOptions() : solver options

DuctAPE.set_options — Function

set_options(; kwargs...)
set_options(multipoint; kwargs...)

Set the options for DuctAPE to use.

Note that the vast majority of the available options are defined through keyword arguments. See the documentation for the various option types for more information.

Arguments

multipoint::AbstractVector{OperatingPoint} : a vector of operating points to use if running a multi-point analysis.

Integration Options

DuctAPE.IntegrationOptions — Type

struct IntegrationOptions{TN<:IntegrationMethod,TS<:IntegrationMethod}

A struct used to hold the integration options for both the nominal and singular cases.

Fields

nominal::IntegrationMethod=GaussLegendre(8) : the integration options to use for the nominal case.

singular::IntegrationMethod=GaussLegendre(8) : the integration options to use for the self-induced case.

DuctAPE.GaussLegendre — Type

struct GaussLegendre{TN,TW} <: IntegrationMethod

Options for Gauss-Legendre integration method

Fields

sample_points::TN : Sample Points

weights::TW : Gauss weights

DuctAPE.GaussKronrod — Type

struct GaussKronrod{TF,TI} <: IntegrationMethod

Options for Gauss-Kronrod integration method

Fields

order::TI = 7 : order of Legendre polynomial to use on each interval

maxevales::TI = 10^7 : maximum number of evaluations in the adaptive method

atol::TF = 0.0 : absolute error tolerance. (note, if zero, QuadGK uses sqrt(eps()) relative tolerance).

DuctAPE.Romberg — Type

struct Romberg{TF,TI} <: IntegrationMethod

Options for Romberg integration method

Fields

max_subdivisions::TI = 10 : maximum number of subdivisions. Note, total number of internvals is 2^N, where N is number of subdivisions.

atol::TF = 1e-6 : absolute error tolerance.

Solver Options

Elliptic Grid Solve

DuctAPE.SLORGridSolverOptions — Type

struct SLORGridSolverOptions{TB,TF,TI} <: GridSolverOptionsType

Options for SLOR (successive line over relaxation) elliptic grid solver.

Fields

iteration_limit::TI = 100 : maximum number of iterations

atol::TF = 1e-9 : absolute convergence tolerance

`converged::AbstractVector{TB} = [false]

DuctAPE.GridSolverOptions — Type

struct GridSolverOptions{TB,TF,TI,TSym} <: GridSolverOptionsType

Options for Newton elliptic grid solver.

Fields

iteration_limit::TI = 10 : maximum number of iterations

atol::TF = 1e-14 : absolute convergence tolerance

algorithm::TSym = :newton : algorithm to use in NLsolve.jl

autodiff::TSym = :forward : differentiation method to use in NLsolve.jl

`converged::AbstractVector{TB} = [false]

Aerodynamics Solve

DuctAPE.ChainSolverOptions — Type

struct ChainSolverOptions{TB,TS<:Union{ExternalSolverOptions,PolyAlgorithmOptions}} <:PolyAlgorithmOptions

Options for Chain Solvers (try one solver, if it doesn't converge, try another)

Fields

`solvers::AbstractVector{TS} = [ NLsolveOptions(; algorithm=:anderson, atol=1e-12), MinpackOptions(; atol=1e-12), NonlinearSolveOptions(; algorithm=SimpleNonlinearSolve.SimpleNewtonRaphson, atol=1e-12, additional_kwargs=(; autodiff=SimpleNonlinearSolve.AutoForwardDiff()), ), ] : Vector of solver options to use.

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.CompositeSolverOptions — Type

struct CompositeSolverOptions{
    TB,TS<:Union{ExternalSolverOptions,PolyAlgorithmOptions}
} <: PolyAlgorithmOptions

Options for Composite Solvers (start with a partial solve of one solve, then finish with another starting where the first left off).

Fields

`solvers::AbstractVector{TS} = [ NLsolveOptions(; algorithm=:newton, iteration_limit=3), NLsolveOptions(; algorithm=:anderson, atol=1e-12), ]' : Vector of solver options to use.

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.NLsolveOptions — Type

struct NLsolveOptions{TB,TF,TK,Tls,Tlsk,TSym} <: ExternalSolverOptions

Options for the NLsolve pacakge solvers

Fields

algorithm::TSym = :anderson : algorithm to use

additional_kwargs::TK = (;) : any additional keyword arguments for the solver

atol::TF = 1e-12 : absolute convergence tolerance

iteration_limit::TF = 25 : maximum number of iterations

linesearch_method::Tls = LineSearches.MoreThuente : line search method to use

linesearch_kwargs::Tlsk = (;) : any additional lineseach keyword arguments

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.NonlinearSolveOptions — Type

struct NonlinearSolveOptions{TA,TB,TF,TI,TT} <: ExternalSolverOptions

Options for the SimpleNonlinearSolve pacakge solvers

Fields

algorithm::TA = SimpleNonlinearSolve.SimpleNewtonRaphson : algorithm to use

additional_kwargs::TK = (;) : any additional keyword arguments for the solver

atol::TF = 1e-12 : absolute convergence tolerance

iteration_limit::TF = 25 : maximum number of iterations

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.MinpackOptions — Type

struct MinpackOptions{TB,TF,TI,TSym} <: ExternalSolverOptions

Options for the MINPACK's HYBRJ solver

Fields

algorithm::TSym = :hybr : algorithm to use in MINPACK.jl (hybr is HYBRJ when the jacobian is provided)

atol::TF = 1e-12 : absolute convergence tolerance

iteration_limit::TF = 100 : maximum number of iterations

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.SIAMFANLEOptions — Type

struct SIAMFANLEOptions{TA,TB,TF,TI,TK} <: ExternalSolverOptions

Options for the SIAMFANLEquations pacakge solvers

Fields

algorithm::TA = SIAMFANLEquations.nsoli : algorithm to use

rtol::TF = 0.0 : relative convergence tolerance

atol::TF = 1e-10 : absolute convergence tolerance

iteration_limit::TF = 1000 : maximum number of iterations

linear_iteration_limit::TF = 5 : maximum number of linear solve iterations (GMRES)

additional_kwargs::TK = (;) : any additional keyword arguments for the solver

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.SpeedMappingOptions — Type

struct SpeedMappingOptions{TB,TF,TI,TL,TSm,TU} <: ExternalSolverOptions

Options for the SpeedMapping.jl package solver

Fields

`orders::AbstractVector{TI} = [3, 2]

sig_min::TSm = 0 : maybe set to 1?

stabilize::TB = false : stabilizes before extrapolation

check_obj::TB = false : checks for inf's and nan's and starts from previous finite point

atol::TF = 1e-10 : absolute convergence tolerance

iteration_limit::TF = 1000 : maximum number of iterations

time_limit::TF = Inf : time limit in seconds

lower::TL = nothing : box lower bounds

upper::TU = nothing : box upper bounds

buffer::TF = 0.01 : if using bounds, buffer brings x inside bounds by buffer amountd

Lp::TF = Inf : p value for p-norm for convergence criteria

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.FixedPointOptions — Type

struct FixedPointOptions{TB,TF,TI} <: ExternalSolverOptions

Options for the FixedPoint.jl package solver

Fields

iteration_limit::TF = 1000 : maximum number of iterations

vel::TF = 0.9 : vel keyword argument, default is package default

ep::TF = 0.01 : ep keyword argument, default is package default

atol::TF = 1e-12 : absolute convergence tolerance

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

DuctAPE.CSORSolverOptions — Type

struct CSORSolverOptions{TB,TC<:ConvergenceType,TF,TS} <: SolverOptionsType

Type containing all the options for the CSOR (controlled successive over relaxation) solver.

Note that the defaults match DFDC with the exception of the relaxation schedule, which is an experimental feature.

Fields

var::type :

verbose::TB = false : flag to print verbose statements

iteration_limit::TF = 1e2 : maximum number of iterations

nrf::TF = 0.4 : nominal relaxation factor

bt1::TF = 0.2 : backtracking factor 1

bt2::TF = 0.6 : backtracking factor 2

pf1::TF = 0.4 : press forward factor 1

pf2::TF = 0.5 : press forward factor 2

btw::TF = 0.6 : backtracking factor for wake

pfw::TF = 1.2 : press forward factor for wake

relaxation_schedule::TS = [[0.0;1e-14;1e-13;1e10]), [1.0;1.0;0.0;0.0])] : values used in spline definition for scaling the relaxation factors (second vector) after various convergence values (first vector).

f_circ::TF = 1e-3 : convergence tolerance for rotor circulation

f_dgamw::TF = 2e-4 : convergence tolerance for wake vortex strength

convergence_type::TC = Relative() : dispatch for relative or absolute convergence criteria.

Vconv::TF = 1.0 : velocity used in relative convergence criteria (should be set to Vref).

converged::AbstractVector{TB} = [false] : flag to track if convergence took place.

Preprocess

DuctAPE.setup_analysis — Function

setup_analysis(
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
)

Perform pre-processing and cache setup (as needed) for propuslor analysis.

Arguments

propulsor::Propulsor : Propulsor input object (see docstring for Propulsor type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

Returns

problem_dimensions::NamedTuple : Named Tuple contiaining bookkeeping information (problem dimensions)

prepost_containers::NamedTuple : Named Tuple containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : Vector containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : Named Tuple containing dimensions used for reshaping the solve parameter cache

A_bb_LU::LinearAlgebra.LU : The LU factorization of the AIC matrix used in the panel method

lu_decomp_flag::Bool : flag indicating if the LU decomposition was successful

airfoils::Matrix{AFType} : Matrix contiaining the blade element airfoil polar objects

idmaps::NamedTuple : Named Tuple containing bookkeeping information (index mappings)

Analysis

DuctAPE.analyze — Function

analyze(
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze propulsor, including preprocessing.

Arguments

propulsor::Propulsor : Propulsor input object (see docstring for Propulsor type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options=set_options();
    return_inputs=false,
    solve_container_caching=nothing,
)

Analyze propulsor, assuming setup_analysis has been called and the outputs thereof are being passed in here.

Arguments

propulsor::Propulsor : Propulsor input object

prepost_containers::NamedTuple : An output from setup_analysis containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : An output from setup_analysis containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : An output from setup_analysis containing dimensions used for reshaping the solve parameter cache

airfoils::Vector{AFType} : An output from setup_analysis contiaining the blade element airfoil polar objects

A_bb_LU::LinearAlgebra.LU : An output from setup_analysis that is the LU decomposition of the AIC matrix used in the panel method

idmaps::NamedTuple : An output from setup_analysis containing bookkeeping information (index mappings)

problem_dimensions::NamedTuple : An output from setup_analysis contiaining bookkeeping information (problem dimensions)

options::Options=set_options() : Options object

Keyword Arguments

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    multipoint::AbstractVector{OperatingPoint},
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze propulsor, including preprocessing, for a set of operating points.

Arguments

multipoint::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the propulsor (note that the operating point within the propulsor input will be overwritten with these)

propulsor::Propulsor : Propulsor input object

options::Options=set_options() : Options object

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Vector of named tuples of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    multipoint::Vector{OperatingPoint},
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options=set_options();
    return_inputs=false,
    solve_container_caching=nothing,
)

Analyze propulsor, assuming setup_analysis has been called and the inputs are being passed in here.

Arguments

multipoint::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the propulsor (note that the operating point within the propulsor input will be overwritten with these)

propulsor::Propulsor : Propulsor input object

prepost_containers::NamedTuple : An output from setup_analysis containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : An output from setup_analysis containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : An output from setup_analysis containing dimensions used for reshaping the solve parameter cache

airfoils::Vector{AFType} : An output from setup_analysis contiaining the blade element airfoil polar objects

A_bb_LU::LinearAlgebra.LU : An output from setup_analysis that is the LU decomposition of the AIC matrix used in the panel method

idmaps::NamedTuple : An output from setup_analysis containing bookkeeping information (index mappings)

problem_dimensions::NamedTuple : An output from setup_analysis contiaining bookkeeping information (problem dimensions)

options::Options=set_options() : Options object

Keyword Arguments

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true. Note that some inputs will be overwritten (e.g. the linear system RHS components related to the freestream) and only those associated with the final operating point will be returned.

convergence_flag : Flag for successful solve convergence

Miscellaneous

Airfoil/Geometry Manipulation

NACA 6-Series Cascade Geometry Generation