API Reference

API Reference

Public API

Private API

Bookkeeping

Caching

Analysis

Post-process

Index

Public API

DuctAPE.setup_analysis — Function

setup_analysis(
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
)

Perform pre-processing and cache setup (as needed) for propuslor analysis.

Arguments

propulsor::Propulsor : Propulsor input object (see docstring for Propulsor type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

Returns

problem_dimensions::NamedTuple : Named Tuple contiaining bookkeeping information (problem dimensions)

prepost_containers::NamedTuple : Named Tuple containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : Vector containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : Named Tuple containing dimensions used for reshaping the solve parameter cache

ivb::NamedTuple : NamedTuple containing the unit induced velocities on the body

A_bb_LU::LinearAlgebra.LU : The LU factorization of the AIC matrix used in the panel method

lu_decomp_flag::Bool : flag indicating if the LU decomposition was successful

airfoils::Matrix{AFType} : Matrix contiaining the blade element airfoil polar objects

idmaps::NamedTuple : Named Tuple containing bookkeeping information (index mappings)

The remainder of the public API elements are found throughout the remainder of this documentation in the context of their usage. The index at the bottom of this page may be helpful in locating them.

Private API

Bookkeeping

DuctAPE.get_problem_dimensions — Function

get_problem_dimensions(paneling_constants::PanelingConstants)
get_problem_dimensions(body_vortex_panels, rotor_source_panels, wake_vortex_panels)

Determine all relevant dimensions to the problem based either on the paneling_constants or the panels themselves.

Arguments

paneling_constants::PanelingConstants : Rotor (and possibly stator) geometric paramters.

Returns

problem_dimensions::ProblemDimensions : ProblemDimensions object.

Caching

DuctAPE.allocate_wake_panel_container! — Function

allocate_wake_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

wake_vortex_panels::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to the wake vortex panel object

DuctAPE.allocate_panel_containers! — Function

allocate_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

panels::NamedTuple : A named tuple of named tuples containing the dimensions needed to reshape the cache with regards to the panel objects

DuctAPE.allocate_panel_container! — Function

allocate_panel_container!(total_length, nn, np, tn, tp, nb)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

nn::Int : number of nodes in each body, rotor, or wake sheet

np::Int : number of panels in each body, rotor, or wake sheet

tn::Int : number of total nodes among the bodies, rotors, or wake sheets

tp::Int : number of total panels among the bodies, rotors, or wake sheets

nb::Int : number of bodies, rotors, or wake sheets

Returns

panel::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to an arbitrary panel set

DuctAPE.allocate_body_panel_container! — Function

allocate_body_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

body_vortex_panels::NamedTuple : A named tuple containing the dimensions needed to reshape the cache with regards to the body vortex panel object

DuctAPE.allocate_rotor_panel_container! — Function

allocate_rotor_panel_containers!(total_length, problem_dimensions::ProblemDimensions)

A helper function is assembling the prepostcontainercache.

Arguments

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

problem_dimensions::ProblemDimensions : a ProblemDimensions object

Returns

rotor_source_panels::NamedTuple : A named containing the dimensions needed to reshape the cache with regards to the rotor source panel object

DuctAPE.allocate_solve_parameter_extras! — Function

allocate_solve_parameter_extras!(
    solver_options::SolverOptionsType, input_length, total_length
)

Includes additional caching for various solvers. Currently only does anything for SIAMFANLEOptions types.

Arguments

input_length::Int : the number of state variables in the solver

total_length::Vector{Int} : a one-element vector used to store the total length in order to know how large of a cache to allocate. Is updated in place.

Returns

solve_parameter_extras::NamedTuple : A named tuple containing dimensions related to extra caching parameters used in various solvers.

Analysis

DuctAPE.analyze — Function

analyze(
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze propulsor, including preprocessing.

Arguments

propulsor::Propulsor : Propulsor input object (see docstring for Propulsor type)

options::Options=set_options() : Options object (see set_options and related functions)

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    ivb,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options=set_options();
    return_inputs=false,
    solve_container_caching=nothing,
)

Analyze propulsor, assuming setup_analysis has been called and the outputs thereof are being passed in here.

Arguments

propulsor::Propulsor : Propulsor input object

prepost_containers::NamedTuple : An output from setup_analysis containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : An output from setup_analysis containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : An output from setup_analysis containing dimensions used for reshaping the solve parameter cache

airfoils::Vector{AFType} : An output from setup_analysis contiaining the blade element airfoil polar objects

ivb::NamedTuple : An output from setup_analysis containing the unit induced velocities on the body

A_bb_LU::LinearAlgebra.LU : An output from setup_analysis that is the LU decomposition of the AIC matrix used in the panel method

idmaps::NamedTuple : An output from setup_analysis containing bookkeeping information (index mappings)

problem_dimensions::NamedTuple : An output from setup_analysis contiaining bookkeeping information (problem dimensions)

options::Options=set_options() : Options object

Keyword Arguments

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::NamedTuple : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    multipoint::AbstractVector{OperatingPoint},
    propulsor::Propulsor,
    options::Options=set_options();
    prepost_container_caching=nothing,
    solve_parameter_caching=nothing,
    solve_container_caching=nothing,
    return_inputs=false,
)

Analyze propulsor, including preprocessing, for a set of operating points.

Arguments

multipoint::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the propulsor (note that the operating point within the propulsor input will be overwritten with these)

propulsor::Propulsor : Propulsor input object

options::Options=set_options() : Options object

Keyword Arguments

prepost_container_caching=nothing : Output of allocate_prepost_container_cache

solve_parameter_caching=nothing : Output of allocate_solve_parameter_container_cache

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Vector of named tuples of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true

convergence_flag : Flag for successful solve convergence

analyze(
    multipoint::Vector{OperatingPoint},
    propulsor::Propulsor,
    prepost_containers,
    solve_parameter_cache_vector,
    solve_parameter_cache_dims,
    airfoils,
    ivb,
    A_bb_LU,
    idmaps,
    problem_dimensions,
    options::Options=set_options();
    return_inputs=false,
    solve_container_caching=nothing,
)

Analyze propulsor, assuming setup_analysis has been called and the inputs are being passed in here.

Arguments

multipoint::AbstractVector{OperatingPoint} : Vector of Operating Points at which to analyze the propulsor (note that the operating point within the propulsor input will be overwritten with these)

propulsor::Propulsor : Propulsor input object

prepost_containers::NamedTuple : An output from setup_analysis containing reshaped views into the prepost cache

solve_parameter_cache_vector::Vector : An output from setup_analysis containing the relevant typed cache vector of solve parameters

solve_parameter_cache_dims::NamedTuple : An output from setup_analysis containing dimensions used for reshaping the solve parameter cache

airfoils::Vector{AFType} : An output from setup_analysis contiaining the blade element airfoil polar objects

ivb::NamedTuple : An output from setup_analysis containing the unit induced velocities on the body

A_bb_LU::LinearAlgebra.LU : An output from setup_analysis that is the LU decomposition of the AIC matrix used in the panel method

idmaps::NamedTuple : An output from setup_analysis containing bookkeeping information (index mappings)

problem_dimensions::NamedTuple : An output from setup_analysis contiaining bookkeeping information (problem dimensions)

options::Options=set_options() : Options object

Keyword Arguments

solve_container_caching=nothing : Output of allocate_solve_container_cache

return_inputs=false : flag as to whether or not to return the pre-processed inputs

Returns

outs::Vector{NamedTuple} : Named Tuple of various analysis outputs (see docstring for postprocess for details), note, if linear system decomposition fails, no solve is performed and an empty vector is returned.

ins::NamedTuple : Named Tuple of various pre-processed inputs (e.g. panels and body linear system), will only be returned if return_inputs=true. Note that some inputs will be overwritten (e.g. the linear system RHS components related to the freestream) and only those associated with the final operating point will be returned.

convergence_flag : Flag for successful solve convergence

Post-process

DuctAPE.post_process — Function

Index

DuctAPE.OperatingPoint

DuctAPE.PanelingConstants

DuctAPE.Propulsor

DuctAPE.ReferenceParameters

DuctAPE.RotorStatorParameters

DuctAPE.allocate_body_panel_container!

DuctAPE.allocate_panel_container!

DuctAPE.allocate_panel_containers!

DuctAPE.allocate_prepost_container_cache

DuctAPE.allocate_rotor_panel_container!

DuctAPE.allocate_solve_container_cache

DuctAPE.allocate_solve_parameter_cache

DuctAPE.allocate_solve_parameter_extras!

DuctAPE.allocate_wake_panel_container!

DuctAPE.analyze

DuctAPE.analyze

DuctAPE.get_problem_dimensions

DuctAPE.post_process

DuctAPE.set_options

DuctAPE.setup_analysis