API Reference

Public

Composite Types

There are several native DuctTAPE types that are required for the user to use as inputs.

The first is the freestream object, which defines the various freestream values.

DuctTAPE.Freestream — Type

Freestream{TVI,TVR,TF}

Fields:

vinf::Float : Freestream velocity

rho::Float : Air density value

mu::Float : Air dynamic viscosity value

asound::Float : Speed of sound value

Second is the rotor geometry object that contains all the required information do define a rotor as well as allow DuctTAPE to create the CCBlade rotor, section, and operation point inputs.

DuctTAPE.RotorGeometry — Type

RotorGeometry{TF, TI, TA, TC, TT, TAF, TR, TM}

Fields:

xlocation::Float : x location of rotor plane, non-dimensional based on duct chord (max TE location - min LE location of hub/wall)

numblades::Int : number of rotor blades

radialstations::Array{Float} : array of radial stations defining rotor blade, non-dimensional with hub=0, tip=1

chords::Array{Float} : array of chord lengths at radial stations defining rotor blade, non-dimensional based on blade tip radius

twists::Array{Float} : array of twist values (in degrees) at radial stations defining rotor blade

airfoils::Array{Airfoil} : array of airfoil data objects at radial stations defining rotor blade

rpm::Float : RPM of rotor

Functions

The current main function the user will employ from DuctTAPE is ff2ccb which is the FLOWFoil to CCBlade coupling function. This function calls several private functions listed below.

DuctTAPE.ff2ccb — Function

ff2ccb(ff_solution, rotor, freestream)

Takes the FLOWFoil solution object, rotor geometry object, and freestream object and genereates the inputs to the CCBlade solve function.

Arguments:

ff_solution::FLOWFoil.InviscidSolution : Inviscid solution object from FLOWFoil.

rotor::RotorGeometry : rotor geometry object

freestream::Freestream : freestream object

returns:

ccbrotor::ccblade.rotor : ccblade rotor object

ccbsections::array{ccblade.section} : array of ccblade section objects

ccbop::array{ccblade.operatingpoint} : array of ccblade operation points

rdist::array{float} : array of refined, dimensional rotot radial station r locations.

Private

Auxiliary Coupling Functions

These functions are called by the ff2ccb function as part of converting the FLOWFoil outputs into CCBlade inputs.

DuctTAPE.extract_ff_geom — Function

extract_ff_geom(ff_solution)

extract x and r coordiantes of duct and hub geometries from the FLOWFoil solution object.

arguments:

ff_solution::FLOWFoil.inviscidsolution : inviscid solution object from FLOWFoil.

returns:

duct_x::array{float} : array of x-coordinates of duct wall control points.

duct_r::array{float} : array of r-coordinates of duct wall control points.

hub_x::Array{Float} : Array of x-coordinates of hub wall control points.

hub_r::Array{Float} : Array of r-coordinates of hub wall control points.

DuctTAPE.ff2ccb_velocity — Function

ff2ccb_velocity(rotor, dim_rad_stash, velocities, rdist)

Convert velocities from FLOWFoil to the CCBlade reference frame.

Arguments:

rotor::RotorGeometry : Rotor geometry object

dim_rad_stash::Array{Float} : Dimensional Radial Station locations

velocities::Array{Array{Float}} : Array of [x; r] velocities from FLOWFoil.

rdist::array{float} : array of refined, dimensional rotot radial station r locations.

Returns:

Vx::Array{Float} : x-velocities in the CCBlade reference frame

Vy::Array{Float} : y-velocities in the CCBlade reference frame

DuctTAPE.generate_ccb_sections — Function

generate_ccb_sections(rotor, dim_rad_stash)

Generates CCBlade section objects from rotor information and dimensional radial stations.

Arguments:

rotor::RotorGeometry : Rotor geometry object

dim_rad_stash::Array{Float} : Dimensional Radial Station locations

Returns:

ccbsections::Array{CCBlade.Section} : Array of CCBlade section objects

rdist::array{float} : array of refined, dimensional rotot radial station r locations.

Utility Functions

The following functions are used as various utility functions throughout the code.

DuctTAPE.get_omega — Method

get_omega(rpm)

Calculates rad/s from RPM

Arguments:

rpm::Float : RPM

Returns:

omega::Float : radians per second

DuctTAPE.ismonotonic — Method

ismonotonic(A)

Checks if array, A is monotonic

Arguments:

A::Array : Array in question

Returns:

true::Bool : true or false.

DuctTAPE.lintran — Method

lintran(rb1, rbend, ra1, raend, ra)

Linear transfrom of ra from range (ra1, raend) to (rb1, rbend)

Arguments:

rb1::Float : lower bound of range b

rbend::Float : upper bound of range b

ra1::Float : lower bound of range a

raend::Float : upper bound of range a

ra::Array{Float} : array of values on range a

Returns:

rb::Array{Float} : array of transformed values on range b

DuctTAPE.split_wall — Method

split_wall(x,r)

Splits full airfoil coordinates into upper and lower halves.

Only works based on geometry. Splits at lowest x-value, does not split based on stagnation point.

Arguments:

x::Array{Float} : Array of x-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

r::Array{Float} : Array of r-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

Returns:

xlower::Array{Float} : Array of lower x-coordinates

xupper::Array{Float} : Array of upper x-coordinates

rlower::Array{Float} : Array of lower r-coordinates

rupper::Array{Float} : Array of upper r-coordinates