Cascade Types

Warning

Cascade types are currently in development and not ready for general use.

Cascade types are defined similarly to CCBlade airfoil types. Instead of angle of attack, however, cascade types take in both inflow and stagger angles. In addition, cascade types are dependent on local solidity. Currently only one such method is implemented, taking in inflow, stagger, solidity, Reynolds number, and Mach number. Thus the lift and drag fields for this type are 5D arrays. Note that since the chord and twist are already fixed inputs to DuctAPE, the solidity and stagger angles are already known a priori. In many cases, the CCBlade-like airfoil types are sufficient not only for airfoils, but also higher solidity cases.

DuctAPE.C4Blade.InReStSoMaCAS — Type

InReStSoMaCAS(inflow, Re, stagger, solidity, Mach, cl, cd)
InReStSoMaCAS(filenames::Matrix{String}; radians=true)

Data is fit recursively with Akima splines.

Arguments

inflow::Vector{Float64}: inflow angles

Re::Vector{Float64}: Reynolds numbers

stagger::Vector{Float64}: stagger angles

solidity::Vector{Float64}: local solidity

Mach::Vector{Float64}: Mach numbers

cl::Array{Float64}: lift coefficients where cl[i, j, k, l, m] corresponds to inflow[i], Re[j], stagger[k], solidity[l], Mach[m]

cd::Array{Float64}: lift coefficients where cd[i, j, k, l, m] corresponds to inflow[i], Re[j], stagger[k], solidity[l], Mach[m]

or files with one per Re/Stagger/Solidty/Mach combination

Arguments

filenames::Matrix{String}: name/path of files to read in. filenames[i, j, k, l] corresponds to Re[i], Stagger[j], Solidity[k], and Mach[l] with each in ascending order.

radians::Bool: true if angle of attack in file is given in radians