Utility Functions

DuctAPE.promote_propulsor_type — Function

promote_propulsor_type(propulsor)

Convenience function for promoting types based on any potential elements of the propulsor object dependent on optimization design variables.

Arguments

propulsor::Propulsor : the propulsor input

Returns

TP::Type : the promoted type

DuctAPE.update_operating_point! — Function

update_operating_point!(op_old, op_new)

Overwrites all the values of an OperatingPoint object with another OperatingPoint object's values (or NamedTuple with the same field names).

Arguments

op_old::OperatingPoint : the OperatingPoint to be overwritten (can also be a NamedTuple with the same field names as an OperatingPoint).

op_new::OperatingPoint : the OperatingPoint values to be used (can also be a NamedTuple with the same field names as an OperatingPoint).

DuctAPE.isscalar — Function

isscalar(x::T) where {T} = isscalar(T)
isscalar(::Type{T}) where {T} = BroadcastStyle(T) isa Broadcast.DefaultArrayStyle{0}

Determines if the input is a scalar. Note that Base.BroadcastStyle is imported.

DuctAPE.dot — Function

dot(A, B) = sum(a * b for (a, b) in zip(A, B))

A faster dot product.

DuctAPE.norm — Function

norm(A) = sqrt(mapreduce(x -> x^2, +, A))

A faster 2-norm.

DuctAPE.cross2mag — Function

cross2mag(A, B) = A[1] * B[2] - A[2] * B[1]

2D "cross product" magnitude

DuctAPE.linear_transform — Function

linear_transform(range1, range2, values)

Linear transfrom of values from range (source_range[1], source_range[end]) to (target_range[1], target_range[end])

Arguments

source_range::Vector{Float} : range values come from (can also be a Tuple)

target_range::Vector{Float} : range onto which we are transforming (can also be a Tuple)

source_values::Array{Float} : array of source values to transform

Returns

target_values::Array{Float} : array of transformed sourcevalues onto target range

DuctAPE.extract_primals! — Function

extract_primals!(Avalue, A::AbstractMatrix{T}) where {T}

Extracts primals of A and places them in Avalue.

DuctAPE.lfs — Function

lfs(shape)

Determines length from shape (output of size function).

DuctAPE.reset_containers! — Function

reset_containers!(containers; exception_keys=[])

Resets all fields (not incluing any contained in exception keys) of containers–-which must be arrays, structs of arrays, or tuples of arrays–-to zeros.

DuctAPE.cache_dims! — Function

cache_dims!(total_length, l, s)

A function that returns a named tuple containing an index range and shape and increases total_length by l.

This function is used heavily in the cache allocation functions for setting up the dimension maps used to access the vectorized caches.

Arguments

total_length::Vector{Int} : single element vector containing the current total length of the eventual cache vector. Modified in place.

l::Int : total length of the object in question

s::Int : size of the object in question

Returns

dims::NamedTuple : A named tuple containing index and shape fields

DuctAPE.write_data — Function

write_data(outs, filename; checkoutfileexists=false)

Writes NamedTuples, specifically for writing out the output of the post_procces() function.

Arguments:

outs::NamedTuple : Named tuple to write to file.

filename::String : file name (including full desired path and file type) for file to write

Keyword Arguments:

output_tuple_name::String : desired variable name of written NamedTuple

checkoutfileexists::Bool=false : boolean for whether to check if the outfile already exists and whether or not to overwrite it.