API Reference

Public

Additional Airfoil Functions

These functions are nearly identical to CCBlade implementations, but include a solidity factor as one of the inputs.

DuctTAPE.AlphaReMachsolidityAF — Type

AlphaReMachsolidityAF(alpha, Re, Mach, solidity, cl, cd, info)
AlphaReMachsolidityAF(alpha, Re, Mach, solidity, cl, cd)
AlphaReMachsolidityAF(filenames::Matrix{String}; radians=true)

Airfoil data that varies with angle of attack, Reynolds number, and Mach number. Data is fit with a recursive Akima spline.

Arguments

alpha::Vector{Float64}: angles of attack

Re::Vector{Float64}: Reynolds numbers

Mach::Vector{Float64}: Mach numbers

solidity::Vector{Float64} : Solidity parameter

cl::Array{Float64}: lift coefficients where cl[i, j, k] corresponds to alpha[i], Re[j], Mach[k]

cd::Array{Float64}: drag coefficients where cd[i, j, k] corresponds to alpha[i], Re[j], Mach[k]

info::String: a description of this airfoil data (just informational)

or files with one per Re/Mach combination

Arguments

filenames::Matrix{String}: name/path of files to read in. filenames[i, j] corresponds to Re[i] Mach[j] with Reynolds number and Mach number in ascending order.

radians::Bool: true if angle of attack in file is given in radians

DuctTAPE.AlphaResolidityAF — Type

AlphaResolidityAF(alpha, Re, solidity, cl, cd, info)
AlphaResolidityAF(alpha, Re, solidity, cl, cd)
AlphaResolidityAF(filenames::Matrix{String}; radians=true)

Airfoil data that varies with angle of attack, Reynolds number, and solidity number. Data is fit with a recursive Akima spline.

Arguments

alpha::Vector{Float64}: angles of attack

Re::Vector{Float64}: Reynolds numbers

solidity::Vector{Float64}: solidity numbers

cl::Array{Float64}: lift coefficients where cl[i, j, k] corresponds to alpha[i], Re[j], solidity[k]

cd::Array{Float64}: drag coefficients where cd[i, j, k] corresponds to alpha[i], Re[j], solidity[k]

info::String: a description of this airfoil data (just informational)

or files with one per Re/solidity combination

Arguments

filenames::Matrix{String}: name/path of files to read in. filenames[i, j] corresponds to Re[i] solidity[j] with Reynolds number and solidity number in ascending order.

radians::Bool: true if angle of attack in file is given in radians

DuctTAPE.AlphaSolidityAF — Type

AlphaSolidityAF(alpha, solidity, cl, cd, info, Mach)
AlphaSolidityAF(alpha, solidity, cl, cd, info)
AlphaSolidityAF(alpha, solidity, cl, cd)
read_AlphaSolidityAF(filenames::Vector{String}; radians=true)

Airfoil data that varies with angle of attack and Reynolds number. Data is fit with a recursive Akima spline.

Arguments

alpha::Vector{Float64}: angles of attack

solidity::Vector{Float64}: solidity parameter

cl::Matrix{Float64}: lift coefficients where cl[i, j] corresponds to alpha[i], solidity[j]

cd::Matrix{Float64}: drag coefficients where cd[i, j] corresponds to alpha[i], solidity[j]

info::String: a description of this airfoil data (just informational)

Mach::Float64: Mach number data was taken at (just informational)

or

filenames with one file per Reynolds number.

Arguments

filenames::Vector{String}: name/path of files to read in, each at a different Reynolds number in ascending order

radians::Bool: true if angle of attack in file is given in radians

DuctTAPE.get_clcd — Method

get_clcd(af, alpha, reynolds, mach, solidity)

Return lift and drag coefficients based on airfoil object type and flow conditions.

Arguments:

af::AFType : Airfoil object either of a CCBlade airfoil type or custom type (custom if solidity information is included).

alpha::Float : angle of attack in degrees

reynolds::Float : Reynolds number

mach::Float : Mach number

solidity::Float : Solidity factor (local chord length over distance between radial blade stations on adjacent blades). Set to -1 if not being used.

Returns:

cl::Float : section lift coefficient

cd::Float : section drag coefficient

DuctTAPE.parsefile — Method

overload parsefile function from CCBlade assumes solidity parameter after mach number

DuctTAPE.writefile — Method

overload writefile function from CCBlade to include solidity parameter in file header

Private

Wake Grid Geometry Functions

The wake grid initalization function actually calls two functions. The first is a method for initializing the grid geometry based on conservation of mass:

DuctTAPE.generate_grid_points — Function

generate_grid_points(ductgeometry, ductsplines, rotors, grid_options, debug=false)

Get grid boundary and initial interior points.

Arguments:

ductgeometry::DuctTAPE.DuctGeometry : Duct Geometry Object.

rotors::Array{DuctTAPE.Rotor} : Array of Rotor objects

grid_options::DuctTAPE.GridOptions : GridOptions object

Returns:

x_grid_points::Matrix{Float64,2} : 2D Array of x grid points

r_grid_points::Matrix{Float64,2} : 2D Array of r grid points

The second is an elliptic grid solver using successive line over relaxation (SLOR) to relax the grid, meaning more accurately place the radial grid points along streamlines:

DuctTAPE.relax_grid — Function

relax_grid(xg, rg, nxi, neta; max_iterations, tol)

Relax grid using elliptic grid solver.

Arguments:

xg::Matrix{Float64} : Initial x grid points guess

rg::Matrix{Float64} : Initial r grid points guess

nxi::Int : number of xi (x) stations in the grid

neta::Int : number of eta (r) stations in the grid

Keyword Arguments:

max_iterations::Int : maximum number of iterations to run, default=100

tol::Float : convergence tolerance, default = 1e-9

Returns:

x_relax_points::Matrix{Float64} : Relaxed x grid points guess

r_relax_points::Matrix{Float64} : Relaxed r grid points guess

Rotor Geometric Functions

DuctTAPE.reinterpolate_rotor! — Function

reinterpolate_rotor!(wakegrid, rotor, rotoridx)

Since the wake grid relaxes and is not aligned with aft rotor radial stations, this function reinterpolates rotor data based on updated radial stations.

(The rotor inputs is the only one updated by this function.)

Arguments:

wakegrid::DuctTAPE.WakeGridGeometry : wake grid geometry object

rotor::DuctTAPE.RotorGeometry : the rotor geometry to update

rotoridx::Int : index in the x direction for where the rotor lies on the wake grid

Rotor & Wake Grid Aerodynamic Functions

DuctTAPE.set_grid_aero! — Function

set_grid_aero!(
    b_gamma_grid,
    delta_enthalpy_grid,
    delta_entropy_grid,
    b_circ_rotors,
    rotor_source_strengths,
    control_point_velocities,
    omegas,
    rotoridxs,
)

Sets grid aero data from rotor disk jump data

Aero data includes the three "tilde" values in the DFDC theory: circulation (Gamma), enthalpy (H), and entropy (S), which are all defined based on rotor values, then carried downstream at each wake station, summing up effects from each rotor.

Arguments:

b_gamma_grid::Matrix{Float} : B*Γ values on wake grid

delta_enthalpy_grid::Matrix{Float} : ΔH (enthalpy) values on wake grid

delta_entropy_grid::Matrix{Float} : ΔS (entropy) values on wake grid

b_circ_rotors::Array{Array{Float}} : B*γ values at rotor blades (one array per rotor in increasing order of x location)

rotor_source_strengths::Array{Array{Float}} : σ (source strength) values at rotor blades (one array per rotor in increasing order of x location)

control_point_velocities::Matrix{Float} : velocities at control points

omegas::Array{Float} : rotation rate (rad/s) of the rotors

rotoridxs::Array{Int} : x-indicies of rotor locations

Other Dispatches:

set_grid_aero!(grid_aerodynamics, omegas, rotoridxs)

Same, but inputs are mostly contained in grid_aerodynamics object.

Unique Arguments:

grid_aerodynamics::DuctTAPE.SystemAero : circulations, source strengths, velocities, etc. of system grid.

set_grid_aero!(b_gamma_grid, b_circ_rotor, delta_enthalpy_grid, omega, rotoridx)

Doesn't set entropy on grid (used in initializing rotor aerodynamics). Also only does one rotor at a time (so inputs are for only a single rotor).

TODOs:

probably actually want to have a initializegridaero function for initialization, then an update function for later use.

That way, things can be broken up a bit better so that not everything has to be done in the initializesystemaerodynamics function (or maybe that can run everything, but the current content can be broken out into smaller functions for testing).

DuctTAPE.set_rotor_velocities — Function

set_rotor_velocities(
    vax, vrad, vtan, vinf, omega, radialstations, wwa=0.0, wwt=0.0, vfac=1.0
)

Define RotorVelocities object.

Arguments:

vax::Array{Float} : axial velocities

vrad::Array{Float} : radial velocities

vtan::Array{Float} : tangential velocities

vinf::Array{Float} : freestream velocity

omega::Float : rotation rate in rad/s

`wwa::Array{Float} : user defined additional axial velocity (unused)

`wwt::Array{Float} : user defined additional tangential velocity (unused)

vfac::Float : user defined velocity factor (unused)

Returns:

rotorvelocities::DuctTAPE.RotorVelocities : induced, absolute, and relative velocities along rotor blades.

DuctTAPE.calc_gamma_i — Function

equation 45 (or 61) in dfdc theory

Utility Functions

The following functions are used as various utility functions throughout the code.

DuctTAPE.calc_normal — Method

calc_normal(xs, rs)

Calculate unit normal vector of panel with edges defined by xs and rs.

Returns (0.0, 0.0) if panel has zero length.

Arguments:

xs::Tuple{Float, Float} : x-coordinates of panel edges

rs::Tuple{Float, Float} : r-coordinates of panel edges

Keyword Arguments:

return_tangent::Bool : if true, return tangent vector (as tuple) as a second output.

Returns:

panel_normal::Tuple{Float} : unit normal vector (x, r) of panel.

DuctTAPE.cosinespace — Method

cosinespace(n)

Calculates cosine spaced points from 0 to pi/2 (viz. cosin(theta), theta ∈ [0, pi/2])

Arguments:

n::Int : Number of points to calculate

Returns:

pts::Array{Float} : Array of cosine spaced points

DuctTAPE.get_omega — Method

get_omega(rpm)

Calculates rad/s from RPM

Arguments:

rpm::Float : RPM

Returns:

omega::Float : radians per second

DuctTAPE.ismonotonic — Method

ismonotonic(A)

Checks if array, A is monotonic

Arguments:

A::Array : Array in question

Returns:

true::Bool : true or false.

DuctTAPE.lintran — Method

lintran(rb1, rbend, ra1, raend, ra)

Linear transfrom of ra from range (ra1, raend) to (rb1, rbend)

Arguments:

rb1::Float : lower bound of range b

rbend::Float : upper bound of range b

ra1::Float : lower bound of range a

raend::Float : upper bound of range a

ra::Array{Float} : array of values on range a

Returns:

rb::Array{Float} : array of transformed values on range b

DuctTAPE.sinespace — Method

sinespace(n)

Calculates sine spaced points from 0 to pi/2 (viz. sin(theta), theta ∈ [0, pi/2])

Arguments:

n::Int : Number of points to calculate

Returns:

pts::Array{Float} : Array of sine spaced points