API Reference

Public

Private

Grid Functions

The wake grid initalization function actually calls two functions. The first is a method for initializing the grid geometry based on conservation of mass:

DuctTAPE.generate_grid_points — Function

generate_grid_points(ductgeometry, ductsplines, rotors, grid_options, debug=false)

Get grid boundary and initial interior points.

Arguments:

duct::DuctTAPE.Duct : Duct Object.

rotors::Array{DuctTAPE.Rotor} : Array of Rotor objects

grid_options::DuctTAPE.GridOptions : GridOptions object

Returns:

x_grid_points::Matrix{Float64,2} : 2D Array of x grid points

r_grid_points::Matrix{Float64,2} : 2D Array of r grid points

The second is an elliptic grid solver using successive line over relaxation (SLOR) to relax the grid, meaning more accurately place the radial grid points along streamlines:

DuctTAPE.relax_grid — Function

relax_grid(xg, rg, nxi, neta; max_iterations, tol)

Relax grid using elliptic grid solver.

Arguments:

xg::Matrix{Float64} : Initial x grid points guess

rg::Matrix{Float64} : Initial r grid points guess

nxi::Int : number of xi (x) stations in the grid

neta::Int : number of eta (r) stations in the grid

Keyword Arguments:

max_iterations::Int : maximum number of iterations to run, default=100

tol::Float : convergence tolerance, default = 1e-9

Returns:

x_relax_points::Matrix{Float64} : Relaxed x grid points guess

r_relax_points::Matrix{Float64} : Relaxed r grid points guess