Quick Start Guide

Setting up Duct Geometry

For this example, we are going to set up an example available in the DFDC source files. We have taken one of the case files and transcribed it into a julia file located in the data directory.

We'll go ahead and load that file and grab pieces of it as we go.

# --- WALL GEOMETRY DEFINITION
include("../../data/dfdc/dstestr2_case.jl")

107-element Vector{Float64}:
 0.159526
 0.16136
 0.163786
 0.166523
 0.169314
 0.17207
 0.174771
 0.177409
 0.17999
 0.182504
 ⋮
 0.158336
 0.158373
 0.1584
 0.15842
 0.158434
 0.15844
 0.158442
 0.158441
 0.158439

The geometry is defined for a complete airfoil as the duct wall. We actually want to split the duct wall coordinates for easier navigation as we set up the flow field.

DuctTAPE.split_wall — Function

split_wall(x,r)

Splits full airfoil coordinates into upper and lower halves.

Only works based geometry. Splits at lowest x-value, does not split based on stagnation point.

Arguments:

x::Array{Float} : Array of x-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

r::Array{Float} : Array of r-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

Returns:

xlower::Array{Float} : Array of lower x-coordinates

xupper::Array{Float} : Array of upper x-coordinates

rlower::Array{Float} : Array of lower r-coordinates

rupper::Array{Float} : Array of upper r-coordinates

# - Split Wall Coordinates
innerwallx, innerwallr, outerwallx, outerwallr = DuctTAPE.split_wall(ductx, ductr)

# - Plot Geometry
plot!(innerwallx,innerwallr,aspectratio=:equal)
plot!(outerwallx,outerwallr,linestyle=:dash)
plot!(hubx,hubr,linestyle=:dash, color=2)

Now we want to put all the geometry together in a DuctGeometry object.

DuctTAPE.DuctGeometry — Type

DuctGeometry{TA,TF}

Fields:

wallinnerxcoordinates::Array{Float} : x coordinates of inner (lower) wall geometry

wallouterxcoordinates::Array{Float} : x coordinates of outer (upper) wall geometry

wallinnerrcoordinates::Array{Float} : r coordinates of inner (lower) wall geometry

wallouterrcoordinates::Array{Float} : r coordinates of outer (upper) wall geometry

hubxcoordinates::Array{Float} : x coordinates of hub geometry

hubrcoordinates::Array{Float} : r coordinates of hub geometry

LEx::Float : x-position of leading edge

TEx::Float : x-position of trailing edge

chord::Float : chord length

wallbluntTE::Bool : flag for blunt trailing edge on wall

hubbluntTE::Bool : flag for blunt trailing edge on hub

DuctTAPE.defineDuctGeometry — Function

defineDuctGeometry(
    wallinnerxcoordinates,
    wallinnerrcoordinates,
    wallouterxcoordinates,
    wallouterrcoordinates,
    hubxcoordinates=nothing,
    hubrcoordinates=nothing;
    LEx=nothing,
    TEx=nothing,
    chord=nothing,
)

Constructor function for the DuctGeometry Object.

Required Arguments:

wallinnerxcoordinates::Array{Float} : x coordinates of inner (lower) wall geometry

wallouterxcoordinates::Array{Float} : x coordinates of outer (upper) wall geometry

wallinnerrcoordinates::Array{Float} : r coordinates of inner (lower) wall geometry

wallouterrcoordinates::Array{Float} : r coordinates of outer (upper) wall geometry

Optional Arguments:

hubxcoordinates::Array{Float} : x coordinates of hub geometry

hubrcoordinates::Array{Float} : r coordinates of hub geometry

Note, if hub x and r coordinates are not set, the x coordinates for the inner wall will be used and the r coordinates will be set to zero. Also note that if one of these is unset, the other must also be unset.

Keyword Arguments:

LEx::Float : x-position of manually defined leading edge. Set to foremost x-coordinate of duct and hub geometry otherwise.

TEx::Float : x-position of mannually defined trailing edge. Set to the rear-most x-coordinate of duct and hub geometry otherwise.

chord::Float : manuall defined chord length. Set to difference between leading and trailing edges otherwise.

Using the defineDuctGeometry contructor function, we can input our wall and hub geometries and let the leading and trailing edges and chord length be calculated automatically. Note that this function also outputs a spline object for the inner duct wall and the hub wall. These splines are used throughout the initialization process to help with rotor placement.

# --- DEFINE DUCT OBJECT
duct, ductsplines = DuctTAPE.defineDuctGeometry(
    innerwallx,
    innerwallr,
    outerwallx,
    outerwallr,
    hubx,
    hubr
)

(DuctTAPE.DuctGeometry{Vector{Float64}, Float64, Bool}([0.304542, 0.30009, 0.29409, 0.287159, 0.279916, 0.272566, 0.26517, 0.257742, 0.250277, 0.242797  …  0.020704, 0.017416, 0.014505, 0.011977, 0.009841, 0.008099, 0.006756, 0.005823, 0.005308, 0.005218], [0.005218, 0.005585, 0.006426, 0.007724, 0.009454, 0.011592, 0.014134, 0.017087, 0.020467, 0.024287  …  0.255908, 0.261895, 0.267881, 0.273874, 0.279879, 0.28585, 0.291638, 0.296942, 0.301211, 0.304466], [0.159526, 0.16136, 0.163786, 0.166523, 0.169314, 0.17207, 0.174771, 0.177409, 0.17999, 0.182504  …  0.199618, 0.197448, 0.195197, 0.192881, 0.190513, 0.1881, 0.185649, 0.183168, 0.180677, 0.178173], [0.178173, 0.175718, 0.173376, 0.171183, 0.169158, 0.167296, 0.165587, 0.164018, 0.162581, 0.161275  …  0.15829, 0.158336, 0.158373, 0.1584, 0.15842, 0.158434, 0.15844, 0.158442, 0.158441, 0.158439], [0.306379, 0.304411, 0.299968, 0.295347, 0.290619, 0.285765, 0.280781, 0.275668, 0.270466, 0.265187  …  0.014561, 0.011615, 0.00898, 0.00667, 0.004695, 0.003057, 0.001763, 0.000817, 0.000219, 0.0], [0.035928, 0.03666, 0.038174, 0.039551, 0.040757, 0.041795, 0.042673, 0.043397, 0.043968, 0.044388  …  0.024326, 0.022009, 0.01959, 0.017084, 0.014491, 0.011818, 0.00907, 0.006233, 0.003268, 0.0], 0.0, 0.306379, 0.306379, true, false), DuctTAPE.DuctSplines{FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}, FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}}(FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}([0.304542, 0.30009, 0.29409, 0.287159, 0.279916, 0.272566, 0.26517, 0.257742, 0.250277, 0.242797  …  0.020704, 0.017416, 0.014505, 0.011977, 0.009841, 0.008099, 0.006756, 0.005823, 0.005308, 0.005218], [0.005218, 0.005585, 0.006426, 0.007724, 0.009454, 0.011592, 0.014134, 0.017087, 0.020467, 0.024287  …  0.255908, 0.261895, 0.267881, 0.273874, 0.279879, 0.28585, 0.291638, 0.296942, 0.301211, 0.304466], [0.005218, 0.005585, 0.006426, 0.007724, 0.009454, 0.011592, 0.014134, 0.017087, 0.020467, 0.024287  …  0.249928, 0.255908, 0.261895, 0.267881, 0.273874, 0.279879, 0.28585, 0.291638, 0.296942, 0.301211], [-0.05356895777178894, -0.11422588385426305, -0.16504691650535908, -0.21178181211393987, -0.26455924417326104, -0.3168385475132872, -0.37002275403497115, -0.4241603451877591, -0.4806379741757129, -0.5399783206095436  …  -1.5539417873469252, -1.7086835799900866, -1.9062788960797397, -2.165798244822906, -2.5194808851464114, -3.0166638450158656, -3.7006652802609468, -4.657661364863906, -5.807319547525475, -10.671224611029983], [5.826770622732226, 4.5002192976857796, 2.8780905411149704, 3.9253135058458204, 3.632112294513922, 3.7044404664978923, 3.829238195233005, 3.935733474231462, 4.121830552988461, 3.9998063772323196  …  30.010366476254546, 42.258267888068126, 65.49572226885873, 103.19178312478283, 177.1363055881247, 315.16364319558403, 648.0051518931392, 2070.7610955711234, 5013.781473723795, 411691.23302350176], [-147.58200240297504, 29.459249269754675, -47.45337851688161, 25.953600904685743, 6.865950232641603, 9.822381557360972, 16.611160285352266, 13.655280987972263, 13.835102989295327, -6.120832794637456  …  1604.6674525907347, 2475.733649848496, 4791.042474360526, 8765.487674928809, 18962.086951456193, 45479.207518041236, 144807.50844205538, 1.0394088922717187e6, 377402.6511213752, 1.426761594626947e9]), FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}([0.306379, 0.304411, 0.299968, 0.295347, 0.290619, 0.285765, 0.280781, 0.275668, 0.270466, 0.265187  …  0.014561, 0.011615, 0.00898, 0.00667, 0.004695, 0.003057, 0.001763, 0.000817, 0.000219, 0.0], [0.035928, 0.03666, 0.038174, 0.039551, 0.040757, 0.041795, 0.042673, 0.043397, 0.043968, 0.044388  …  0.024326, 0.022009, 0.01959, 0.017084, 0.014491, 0.011818, 0.00907, 0.006233, 0.003268, 0.0], [0.035928, 0.03666, 0.038174, 0.039551, 0.040757, 0.041795, 0.042673, 0.043397, 0.043968, 0.044388  …  0.026536, 0.024326, 0.022009, 0.01959, 0.017084, 0.014491, 0.011818, 0.00907, 0.006233, 0.003268], [-0.38754645564685475, -0.3587982184085237, -0.32275686922921326, -0.2761381062345554, -0.23312212814218525, -0.1933465464538242, -0.1574283833227317, -0.12461143611708475, -0.09394113158063368, -0.0641637451856979  …  0.630376481988316, 0.7232052108919896, 0.8373069904345483, 0.9790476943766118, 1.1631678825534388, 1.4139643498387235, 1.7632109055465794, 2.2992685365282504, 3.157153822751898, 6.59550209144081], [-9.165381689871582, -4.06731134766873, -5.992101034490946, -4.266056306125101, -3.720240265493685, -3.136096404650486, -2.86889735073257, -2.665816870947827, -2.5316233885371435, -1.9749619591538528  …  -18.63865414158088, -25.713868368379135, -38.109305938891595, -57.69791482198846, -100.47294292131004, -185.8759404031548, -421.3697568482151, -1311.9850495077249, -3285.572245576463, -53295.225803290174], [-630.5757545012166, -1.7017592850302152, -136.74849433729474, 39.905633251639806, 51.7725635663145, 62.500082789494975, 44.366671042074366, 36.15494709752692, 36.46385402635212, 99.20271843687578  …  -889.3997374681618, -1436.5977227237267, -2837.0744313531254, -5150.089534851246, -12482.756588486083, -32262.239587185122, -110374.9361459641, -605043.9100597023, -457858.54093777615, -6.973962413497666e7])))

The duct object now contains all the geometry information we'll need.

Wake Grid Initialization

For the sake of this example, we're going to define a rotor object that has a location only (we won't need the rest of the rotor defnition yet).

DuctTAPE.Rotor — Type

Rotor{TF, TI, TA, TC, TT, TAF, TR, TM}

Fields:

xlocation::Float : x location of rotor plane

numblades::Int : number of rotor blades

radialstations::Array{Float} : array of radial stations defining rotor blade, non-dimensional with hub=0, tip=1

chords::Array{Float} : array of chord lengths at radial stations defining rotor blade

twists::Array{Float} : array of twist values (in degrees) at radial stations defining rotor blade

skews::Array{Float} : array of skew values (similar to sweep) at radial stations defining rotor blade

rakes::Array{Float} : array of rake values (similar to dihedral) at radial stations defining rotor blade

airfoils::Array{Airfoil} : array of airfoil data objects at radial stations defining rotor blade

reynolds::Array{Float} : array of reynolds numbers at radial stations defining rotor blade

solidity:Array{Float} : array of rotor solidity at radial stations defining rotor blade

machs::Array{Float} : array of machs numbers at radial stations defining rotor blade (currently unsupported).

Note that we want to create an array, even if we only have one rotor. When we initialize the grid, it will expect an array. Also, our rotor object has more fields than are used in the original dfdc, for now, we'll set the section skew, rake, reynolds, airfoil, solidity, and mach to nothing.

# --- GENERATE ROTOR OBJECT ARRAY

#generate rotor object
rotor1 = DuctTAPE.Rotor(xdisk1, nblade1, 0.0, chord1, beta1, nothing, nothing, nothing, nothing, nothing, nothing, rpm)

#generate stator object (rpm is zero for stator)
rotor2 = DuctTAPE.Rotor(xdisk2, nblade2, 0.0, chord2, beta2, nothing, nothing, nothing, nothing, nothing, nothing, 0.0)

#assemble array
rotors = [rotor1; rotor2]

We can next define some grid options.

Note

The number of radial stations set for grid options must match the number of radial stations defined for the rotor objects. Furthermore, the rotors must have the same number of radial stations defined.

DuctTAPE.GridOptions — Type

GridOptions{TF,TI}

Fields:

num_radial_stations::Integer : Number of radial stations (equal to number of rotor blade elements used in analysis)

inlet_length::Float : inlet length (unused)

wake_length::Float : length of wake behind duct relative to chord length

wake_expansion_factor::Float : expansion factor to apply to wake grid generation

DuctTAPE.defineGridOptions — Function

defineGridOptions(
    num_radial_stations;
    inlet_length=0.5,
    wake_length=1.0,
    wake_expansion_factor=1.1
)

Constructor function for the GridOptions object.

Required Argument:

num_radial_stations::Integer : Number of radial stations (equal to number of rotor blade elements used in analysis)

Keyword Arguments:

inlet_length::Float : inlet length (unused)

wake_length::Float : length of wake behind duct in terms of chord length

wake_expansion_factor::Float : expansion factor to apply to wake grid generation

# --- SET GRID OPTIONS
num_radial_stations = 10
grid_options = DuctTAPE.defineGridOptions(num_radial_stations)

DuctTAPE.GridOptions{Float64, Int64}(10, 0.5, 1.0, 1.1)

With options set, rotor location chosen, and the wall and hub geometry available, we're finally ready to initialize the wake grid object.

DuctTAPE.initialize_grid — Function

initialize_grid(ductgeometry, ductsplines, rotors, grid_options; max_iterations=-1, tol=1e-9)

Initialize grid via zero-thrust, unit freestream solution.

Arguments:

ductgeometry::DuctTAPE.DuctGeometry : ductgeometry Geometry Object

ductsplines::DuctTAPE.DuctSplines : ductsplines object

rotors::Array{DuctTAPE.Rotor} : Array of rotor objects

grid_options::DuctTAPE.GridOptions : Grid options object

Keyword Arguments:

max_iterations::Int : maximum number of iterations to run, default=100

tol::Float : convergence tolerance, default = 1e-9

Returns:

grid::DuctTAPE.Grid : Grid Object

DuctTAPE.Grid — Type

Grid{TF,TI}

Grid Object

Fields:

x_grid_points::Matrix{Float} : 2D Array of x grid points

r_grid_points::Matrix{Float} : 2D Array of radial grid points

nx::Int : number of x stations

nr::Int : number of radial stations

wallTEidx::Int : index of duct wall trailing edge x location

hubTEidx::Int : index of hub wall trailing edge x location

rotoridxs::Array{Int} : array of indices of rotor x locations

# --- INITIALIZE GRID
grid = DuctTAPE.initialize_grid(duct, rotors, grid_options)

xg = grid.x_grid_points
rg = grid.r_grid_points
nx = grid.nx
nr = grid.nr

plot!(xg, rg, color=3, linewidth=0.5)
plot!(xg', rg', color=3, linewidth=0.5)

Note

This process should be relatively robust in that most combinations of wall and hub and rotor position should generate a grid. However, in the case that there is no overlap in the x-positions of the wall and hub, problems may occur later in the solver. In addition if the rotor is positioned somewhere not between the wall and hub, for example, if the rotor is out in front of the duct, then if a solution is found, it will likely be inaccurate. And finally, things will break if the hub and wall overlap in the radial direction, in other words, if the duct is blocked.

System Paneling