Quick Start Guide

Setting up Duct Geometry

For this example, we are going to set up an example available in the DFDC source files. We have taken one of the case files and transcribed it into a julia file located in the data directory.

We'll go ahead and load that file and grab pieces of it as we go.

# --- WALL GEOMETRY DEFINITION
include("../../data/dfdc/dstestr2_case.jl");

DFDC Example Case Information Loaded

The geometry is defined for a complete airfoil as the duct wall. We actually want to split the duct wall coordinates for easier navigation as we set up the flow field.

DuctTAPE.split_wall — Function

split_wall(x,r)

Splits full airfoil coordinates into upper and lower halves.

Only works based on geometry. Splits at lowest x-value, does not split based on stagnation point.

Arguments:

x::Array{Float} : Array of x-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

r::Array{Float} : Array of r-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

Returns:

xlower::Array{Float} : Array of lower x-coordinates

xupper::Array{Float} : Array of upper x-coordinates

rlower::Array{Float} : Array of lower r-coordinates

rupper::Array{Float} : Array of upper r-coordinates

# - Split Wall Coordinates
outerwallx, innerwallx, outerwallr, innerwallr = DuctTAPE.split_wall(ductx, ductr)

# - Plot Geometry
plot!(innerwallx, innerwallr, linewidth=2)
plot!(outerwallx, outerwallr, linestyle=:dash, color=1, linewidth=2)
plot!(hubx, hubr, color=2, linewidth=2)

Now we want to put all the geometry together in a DuctGeometry object.

DuctTAPE.DuctGeometry — Type

DuctGeometry{TA,TF}

Fields:

wallinnerxcoordinates::Array{Float} : x coordinates of inner (lower) wall geometry

wallinnerrcoordinates::Array{Float} : r coordinates of inner (lower) wall geometry

wallinnerspline::FLOWMath.Akima : Spline of inner coordinates of duct wall

wallouterxcoordinates::Array{Float} : x coordinates of outer (upper) wall geometry

wallouterrcoordinates::Array{Float} : r coordinates of outer (upper) wall geometry

wallouterspline::FLOWMath.Akima : Spline of outer coordinates of duct wall

hubxcoordinates::Array{Float} : x coordinates of hub geometry

hubrcoordinates::Array{Float} : r coordinates of hub geometry

hubspline::FLOWMath.Akima : Spline of hub coordinates

LEx::Float : x-position of leading edge

TEx::Float : x-position of trailing edge

chord::Float : chord length

wallbluntTE::Bool : flag for blunt trailing edge on wall

hubbluntTE::Bool : flag for blunt trailing edge on hub, unused

DuctTAPE.defineDuctGeometry — Function

defineDuctGeometry(
    wallinnerxcoordinates,
    wallinnerrcoordinates,
    wallouterxcoordinates,
    wallouterrcoordinates,
    hubxcoordinates=nothing,
    hubrcoordinates=nothing;
    LEx=nothing,
    TEx=nothing,
    chord=nothing,
    bluntTEtol=1e-9,
)

Constructor function for the DuctGeometry Object.

Required Arguments:

wallinnerxcoordinates::Array{Float} : x coordinates of inner (lower) wall geometry

wallouterxcoordinates::Array{Float} : x coordinates of outer (upper) wall geometry

wallinnerrcoordinates::Array{Float} : r coordinates of inner (lower) wall geometry

wallouterrcoordinates::Array{Float} : r coordinates of outer (upper) wall geometry

Optional Arguments:

hubxcoordinates::Array{Float} : x coordinates of hub geometry

hubrcoordinates::Array{Float} : r coordinates of hub geometry

Note, if hub x and r coordinates are not set, the x coordinates for the inner wall will be used and the r coordinates will be set to zero. Also note that if one of these is unset, the other must also be unset.

Keyword Arguments:

LEx::Float : x-position of manually defined leading edge. Set to foremost x-coordinate of duct and hub geometry otherwise.

TEx::Float : x-position of mannually defined trailing edge. Set to the rear-most x-coordinate of duct and hub geometry otherwise.

chord::Float : manually defined chord length. Set to difference between leading and trailing edges otherwise.

bluntTEtol::Float : tolerance for how close trailing edge points need to be before being considered a blunt trailing edge.

Notes

The wall inner and outer and hub coordinates are all saved from leading to trailing edge (this helps with splining things later). It is assumed that the wall is similar to an airfoil that may or may not have a blunt trailing edge.

Using the defineDuctGeometry contructor function, we can input our wall and hub geometries and let the leading and trailing edges and chord length be calculated automatically. Note that this function also outputs a spline object for the inner duct wall and the hub wall. These splines are used throughout the initialization process to help with rotor placement.

# --- DEFINE DUCT OBJECT
ductgeometry, ductsplines = DuctTAPE.defineDuctGeometry(
    innerwallx,
    innerwallr,
    outerwallx,
    outerwallr,
    hubx,
    hubr
)

The ductgeometry and ductsplines objects now contains all the geometry information we'll need.

Defining Rotors

Along with the duct geometry, we will need to define an array of Rotor objects.

DuctTAPE.RotorGeometry — Type

RotorGeometry{TF, TI, TA, TC, TT, TAF, TR, TM}

Fields:

xlocation::Float : x location of rotor plane, non-dimensional based on duct chord (max TE location - min LE location of hub/wall)

numblades::Int : number of rotor blades

radialstations::Array{Float} : array of radial stations defining rotor blade, non-dimensional with hub=0, tip=1

tipgap::Float : gap between blade tip and duct wall (not implemented yet)

chords::Array{Float} : array of chord lengths at radial stations defining rotor blade, non-dimensional based on blade tip radius

twists::Array{Float} : array of twist values (in degrees) at radial stations defining rotor blade

skews::Array{Float} : array of skew values (similar to sweep) at radial stations defining rotor blade, non-dimensional based on rotor tip radius. (note: this is for reference only, the solver can't use this information)

rakes::Array{Float} : array of rake values (similar to dihedral) at radial stations defining rotor blade, non-dimensional based on rotor tip radius. (note: this is for reference only right now. it may be implemented into the grid initialization functions later.)

airfoils::Array{Airfoil} : array of airfoil data objects at radial stations defining rotor blade

solidity:Array{Float} : array of rotor solidity at radial stations defining rotor blade, chord/distance between blade sections

rpm::Float : RPM of rotor

Note that we want to create an array, even if we only have one rotor. When we initialize the grid, it will expect an array. Also, our rotor object has more fields than are used in the original dfdc, for now, we'll set the section skew, rake, and solidity to nothing. For the airfoils, we will use the CCBlade.jl functionality and for our case here, define our airfoils as ccb.AlphaReAF objects (where CCBlade has been included in the DuctTAPE package and renamed 'ccb') using data files from digitized dfdc plots.

We can define rotor objects directly, or we can use the initialize_rotor_geometry function to help us.

DuctTAPE.initialize_rotor_geometry — Function

function initialize_rotor_geometry(
    xlocation,
    numblades,
    numstations,
    chords,
    twists,
    airfoils,
    rpm;
    radialstations=nothing,
    tipgap=0.0, #non-dimensional relative to blade length
    skews=nothing,
    rakes=nothing,
    solidities=nothing,
)

Initialize non-dimensional rotor geometry and other relative information.

Arguments:

xlocation::Float : x location of rotor relative to duct chord

numblades::Float : number of rotor blades

numstations::Float : number of radial stations (the length of the below sectional properties)

chords::Array{Float} : array of section chord lengths (relative to blade tip radius)

twists::Array{Float} : array of section twists (0 degrees is aligned with the axial direction) in degrees

airfoils::Array{AFType} : airfoil objects at each section

rpm::Float : RPM of rotor

Keyword Arguments:

radialstations::Array{Float} : non-dimensional radial stations (0 = hub radius, 1 = tip radius)

tipgap::Float : non-dimensional relative to blade length (Not implemented yet)

skews::Array{Float} : sectional skew distance relative to tip radius (not implemented yet)

rakes::Array{Float} : sectional rake distances relative to tip radius (not implemented yet)

solidities::Array{Float} or Nothing : sectional rotor solidities, if using.

Returns:

rotorgeometry::DuctTAPE.RotorGeometry : Rotor geometry object

TODOs: if rotor rake is present, need to redo parts of grid initialization (and check that the unused portions of the code work now) to account for different x-locations of start of wake. NOTE: not sure if rake will work with this solver. Need to think about that more before implementing.

# -- GENERATE ROTOR OBJECT ARRAY

# set datapath for airfoil data files. Remember to grab these files and change the data path for your case.
datapath = "../../data/dfdc/airfoils/"

# generate airfoils for rotor
af1 = DuctTAPE.ccb.AlphaReAF([
    datapath * "disk1_re5e5.dat",
    datapath * "disk1_re1e6.dat",
    datapath * "disk1_re1.5e6.dat",
    datapath * "disk1_re2e6.dat",
])

# generate airfoils for stator
af2 = DuctTAPE.ccb.AlphaReAF([
    datapath * "disk2_re5e5.dat",
    datapath * "disk2_re1e6.dat",
    datapath * "disk2_re1.5e6.dat",
    datapath * "disk2_re2e6.dat",
])

#generate rotor object using initialization function
rotor1 = DuctTAPE.initialize_rotor_geometry(
    xdisk1, #x position of rotor
    nblade1, # number of blades
    length(rnondim1), #number of radial stations
    chord1, #chords
    beta1, #twists
    fill(af1, length(rnondim1)), #airfoils
    rpm; #RPM
    radialstations=rnondim1 #radial station locations
)

#generate stator object (rpm is zero for stator) using direct definition
rotor2 = DuctTAPE.RotorGeometry(
    xdisk2, #x position of rotor
    nblade2, #number of blades
    rnondim2, #radial stations
    0.0, #tip gap
    chord2, #chords
    beta2, #twists
    zeros(length(rnondim2)), #skews
    zeros(length(rnondim2)), #rakes
    fill(af2, length(rnondim2)), #airfoils
    -1 .* ones(Int, length(rnondim2)), #solidities
    0.0, #RPM
)

#assemble array
rotors = [rotor1; rotor2]

2-element Vector{RotorGeometry{Float64, Int64, Vector{Float64}, Float64, Vector{Float64}, Vector{Float64}, Vector{Float64}, Vector{Float64}, Vector{CCBlade.AlphaReAF{Float64, String}}, TSo, Float64} where TSo}:
 RotorGeometry{Float64, Int64, Vector{Float64}, Float64, Vector{Float64}, Vector{Float64}, Vector{Float64}, Vector{Float64}, Vector{CCBlade.AlphaReAF{Float64, String}}, Vector{Float64}, Float64}(0.39167175295956963, 6, [0.0, 0.11110776761354521, 0.22221553522709042, 0.33331327234793784, 0.4444210399614829, 0.5555488685604237, 0.6665864227250843, 0.7777242818167229, 0.8888621409083616, 0.9999999999999999], 0.0, [0.3744190691790399, 0.3571542712564086, 0.3402423596777416, 0.3250882215859911, 0.3122378320793662, 0.3017244823223917, 0.2933950329582529, 0.2870430787668953, 0.2824688727611692, 0.27951927558425993], [19.171999999999997, 30.241999999999997, 38.293, 43.571, 47.791, 51.234, 54.103, 56.543, 58.663, 60.547], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], CCBlade.AlphaReAF{Float64, String}[CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15)], [-1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0, -1.0], 8000.0)
 RotorGeometry{Float64, Int64, Vector{Float64}, Float64, Vector{Float64}, Vector{Float64}, Vector{Float64}, Vector{Float64}, Vector{CCBlade.AlphaReAF{Float64, String}}, Vector{Int64}, Float64}(0.7180648804258777, 9, [0.0, 0.11064323803152241, 0.2212963145156531, 0.3320871293362981, 0.44301568249345763, 0.5541213277975642, 0.6652958422699281, 0.7766671257944551, 0.8882351783711461, 1.0], 0.0, [0.38141720896601594, 0.3786629856044173, 0.37502136330769736, 0.3708604483007954, 0.366384013672517, 0.3617432459081049, 0.3570433182146848, 0.3523828304739368, 0.3478472359166503, 0.3434956944718333], [-2.7109999999999985, -1.7390000000000043, -1.2279999999999944, -0.867999999999995, -0.6380000000000052, -0.4410000000000025, -0.12399999999999523, 0.4170000000000016, 1.0450000000000017, 1.6119999999999948], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0], CCBlade.AlphaReAF{Float64, String}[CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-1.5707963267948966, -1.5533430342749532, -1.53588974175501, -1.5184364492350666, -1.5009831567151235, -1.4835298641951802, -1.4660765716752369, -1.4486232791552935, -1.4311699866353502, -1.413716694115407  …  1.413716694115407, 1.4311699866353502, 1.4486232791552935, 1.4660765716752369, 1.4835298641951802, 1.5009831567151235, 1.5184364492350666, 1.53588974175501, 1.5533430342749532, 1.5707963267948966], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15)], [-1, -1, -1, -1, -1, -1, -1, -1, -1, -1], 0.0)

Take note that the rotor objects are defined using non-dimensional parameters. If we want to visualize things at this point, we'll have to dimensionalize things, which is done internally in the code using BladeDimensions objects.

DuctTAPE.BladeDimensions — Type

BladeDimensions{TF, TA}

Fields:

rhub::Float : hub radius (dimensional)

rtip::Float : tip radius (dimensional)

rdim::Array{Float} : array of dimensional radial stations

cdim::Array{Float} : array of dimensional chords

tdim::Array{Float} : array of twists (already dimensional in rotorgeometry)

sweptannulus::Float : area of blade swept annulus

sweptarea::Float : area of blade tip swept disk

DuctTAPE.initialize_blade_dimensions — Function

initialize_blade_dimensions(rotors, ductgeometry)

Initilialize needed blade information for various calculations during the solution process.

Arguments:

ductsplines::DuctTAPE.ductsplines : ductsplines object containing splines for duct wall and hub

Rotor::DuctTAPE.Rotor : Rotor object for which to define blade information

blade1 = DuctTAPE.initialize_blade_dimensions(ductgeometry, ductsplines, rotor1)
blade2 = DuctTAPE.initialize_blade_dimensions(ductgeometry, ductsplines, rotor2)

blades = [blade1; blade2]

nr = length(blade1.rdim)
plot!(rotor1.xlocation.*ductgeometry.chord*ones(nr), blade1.rdim, color=4, linewidth=2)
plot!(rotor2.xlocation.*ductgeometry.chord*ones(nr), blade2.rdim, color=4, linewidth=2)

Wake Grid Initialization

With the duct geometry and rotor data defined, we can now initialize the rotor wake grid. We'll begin by setting some grid options.

Note

The number of radial stations set for grid options must match the number of radial stations defined for the rotor objects. Furthermore, the rotors must have the same number of radial stations defined.

DuctTAPE.GridOptions — Type

GridOptions{TF,TI}

Fields:

num_radial_stations::Integer : Number of radial stations (equal to number of rotor blade elements used in analysis)

inlet_length::Float : inlet length (unused)

wake_length::Float : length of wake behind duct relative to chord length

wake_expansion_factor::Float : expansion factor to apply to wake grid generation

DuctTAPE.defineGridOptions — Function

defineGridOptions(
    num_radial_stations;
    inlet_length=0.5,
    wake_length=1.0,
    wake_expansion_factor=1.1
)

Constructor function for the GridOptions object.

Required Argument:

num_radial_stations::Integer : Number of radial stations (equal to number of rotor blade elements used in analysis)

Keyword Arguments:

inlet_length::Float : inlet length (unused)

wake_length::Float : length of wake behind duct in terms of chord length

wake_expansion_factor::Float : expansion factor to apply to wake grid generation

# --- SET GRID OPTIONS
num_radial_stations = length(rnondim1)
grid_options = DuctTAPE.defineGridOptions(num_radial_stations)

GridOptions{Float64, Int64}(10, 0.0, 1.0, 1.1)

With options set, rotor location chosen, and the wall and hub geometry available, we're finally ready to initialize the wake grid object.

DuctTAPE.generate_wake_grid — Function

generate_wake_grid(ductgeometry, rotors, grid_options; max_iterations=-1, tol=1e-9)

Initialize grid via zero-thrust, unit freestream solution.

Arguments:

ductgeometry::DuctTAPE.DuctGeometry : ductgeometry Geometry Object

rotors::Array{DuctTAPE.Rotor} : Array of rotor objects

grid_options::DuctTAPE.GridOptions : Grid options object

Keyword Arguments:

max_iterations::Int : maximum number of iterations to run, default=100

tol::Float : convergence tolerance, default = 1e-9

Returns:

wakegridgeometry::DuctTAPE.WakeGridGeometry : WakeGridGeometry Object

NOTES:

The grid is designed such that the grid spacing in the axial direction will line up with the rotor positions as well as the trailing edge positions of the duct wall and hub (as well as the leading edge positions if they happen to be behind a rotor, which is not advised).

In addition, the grid axial spacing is taken to be as close as possible to the radial spacing, which is defined directly from the rotor radial station positions.

The wake spacing is started at the average of the axial spacing inside the duct area and then expanded by an expansion factor that can be defined by the user and is set to 1.1 by default. This means that the end of the wake will actually not lie directly at the length input by the user (default 2x duct chord), but should be close enough.

If more than one rotor is being analyzed, the rotor radial stations have more than likely changed for aft rotors. Therefore, rotor information for all but the foremost rotor are reinterpolated so that rotor stations line up with wake grid stations.

DuctTAPE.WakeGridGeometry — Type

WakeGridGeometry{TF,TI,TA,TW,TH}

Wake grid geometry object

Fields:

x_grid_points::Matrix{Float} : 2D Array of x grid points

r_grid_points::Matrix{Float} : 2D Array of radial grid points

nx::Int : number of x stations

nr::Int : number of radial stations

wallTEidx::Int : index of duct wall trailing edge x location

hubTEidx::Int : index of hub wall trailing edge x location

rotoridxs::Array{Int} : array of indices of rotor x locations

# --- INITIALIZE GRID
wakegrid = DuctTAPE.generate_wake_grid(ductgeometry, ductsplines, rotors, grid_options)

xg = wakegrid.x_grid_points
rg = wakegrid.r_grid_points
nx = wakegrid.nx
nr = wakegrid.nr

plot(; xlabel="x", ylabel="r", aspectratio=:equal)
plot!(
    ductgeometry.wallinnerxcoordinates,
    ductgeometry.wallinnerrcoordinates;
    color=1,
    linewidth=2,
)
plot!(
    ductgeometry.wallouterxcoordinates,
    ductgeometry.wallouterrcoordinates;
    color=1,
    linestyle=:dash,
    linewidth=2,
)
plot!(ductgeometry.hubxcoordinates, ductgeometry.hubrcoordinates; color=2, linewidth=2)

plot!(rotor1.xlocation .* ductgeometry.chord * ones(nr), blade1.rdim; color=4, linewidth=2)
plot!(rotor2.xlocation .* ductgeometry.chord * ones(nr), blade2.rdim; color=4, linewidth=2)

plot!(xg, rg; color=3, linewidth=0.5)
plot!(xg', rg'; color=3, linewidth=0.5)

Note

This process should be relatively robust in that most combinations of wall and hub and rotor position should generate a grid. However, in the case that there is no overlap in the x-positions of the wall and hub, problems may occur later in the solver. In addition if the rotor is positioned somewhere not between the wall and hub, for example, if the rotor is out in front of the duct, then if a solution is found, it will likely be inaccurate. And finally, things will break if the hub and wall overlap in the radial direction, in other words, if the duct is blocked.

System Paneling

With the geometry defined and the grid initialized, we can define the system panels and their control points (centers). To do so, we use the generate_paneling function:

Missing docstring.

Missing docstring for DuctTAPE.generate_paneling. Check Documenter's build log for details.

For our case here, we are going to use a rough sampling of the geometry data in order to more clearly see the paneling. The generated panels will be contained in Panels objects.

DuctTAPE.Panels — Type

Panels{TPEx,TPEr,TPC,TPN,TPT}

Fields:

panel_edges_x::Array{Tuple{Float, Float}} : Array of sets of x locations for panel edges

panel_edges_r::Array{Tuple{Float, Float}} : Array of sets of r locations for panel edges

panel_centers::Array{Tuple{Float, Float}} : Array of sets of x,r locations for panel centers

panel_normals::Array{Tuple{Float, Float}} : Array of panel unit normal vectors

panel_tyes::Array{String} : Array of panel types (for use in assembling linear system)

# Get paneling of various objects

wall_panels, hub_panels, wake_panels, rotor_source_panels = DuctTAPE.generate_paneling(
    ductgeometry, ductsplines, rotors, wakegrid
)

# PLOT PANELS

plot(; xlabel="x", ylabel="r", aspectratio=:equal, legend=true, label="")

# wall panels:
for i in 1:length(wall_panels.panel_edges_x)
    plot!(
        [wall_panels.panel_edges_x[i][1]; wall_panels.panel_edges_x[i][2]],
        [wall_panels.panel_edges_r[i][1]; wall_panels.panel_edges_r[i][2]];
        color=1,
        linewidth=0.5,
        markershape=:diamond,
        markersize=2,
        label="",
    )
end

scatter!(
    getindex.(wall_panels.panel_centers, 1),
    getindex.(wall_panels.panel_centers, 2);
    color=1,
    markersize=3,
    markershape=:circle,
    label="wall panel centers",
)

#hub panels:
for i in 1:length(hub_panels.panel_edges_x)
    plot!(
        [hub_panels.panel_edges_x[i][1]; hub_panels.panel_edges_x[i][2]],
        [hub_panels.panel_edges_r[i][1]; hub_panels.panel_edges_r[i][2]];
        markersize=2,
        markershape=:diamond,
        color=2,
        linewidth=0.5,
        label="",
    )
end

scatter!(
    getindex.(hub_panels.panel_centers, 1),
    getindex.(hub_panels.panel_centers, 2);
    markersize=3,
    markershape=:circle,
    color=2,
    label="hub panel centers",
)

#vortex sheet panels
for i in 1:length(wake_panels.panel_centers)
    plot!(
        [wake_panels.panel_edges_x[i][1]; wake_panels.panel_edges_x[i][2]],
        [wake_panels.panel_edges_r[i][1]; wake_panels.panel_edges_r[i][2]];
        markersize=2,
        markershape=:diamond,
        color=3,
        linewidth=0.5,
        label="",
    )
end

scatter!(
    getindex.(wake_panels.panel_centers, 1),
    getindex.(wake_panels.panel_centers, 2);
    markersize=3,
    markershape=:circle,
    color=3,
    label="vortex sheet panel centers",
)

#rotor source panels:
for i in 1:length(rotor_source_panels[1].panel_centers)
    plot!(
        [
            rotor_source_panels[1].panel_edges_x[i][1]
            rotor_source_panels[1].panel_edges_x[i][2]
        ],
        [
            rotor_source_panels[1].panel_edges_r[i][1]
            rotor_source_panels[1].panel_edges_r[i][2]
        ];
        markersize=2,
        markershape=:diamond,
        color=4,
        linewidth=0.5,
        label="",
    )
end

scatter!(
    getindex.(rotor_source_panels[1].panel_centers, 1),
    getindex.(rotor_source_panels[1].panel_centers, 2);
    markersize=3,
    markershape=:circle,
    color=4,
    label="rotor source panel centers",
)

# stator source panels
for i in 1:length(rotor_source_panels[2].panel_centers)
    plot!(
        [
            rotor_source_panels[2].panel_edges_x[i][1]
            rotor_source_panels[2].panel_edges_x[i][2]
        ],
        [
            rotor_source_panels[2].panel_edges_r[i][1]
            rotor_source_panels[2].panel_edges_r[i][2]
        ];
        markersize=2,
        markershape=:diamond,
        color=4,
        linewidth=0.5,
        label="",
    )
end

scatter!(
    getindex.(rotor_source_panels[2].panel_centers, 1),
    getindex.(rotor_source_panels[2].panel_centers, 2);
    markersize=3,
    markershape=:circle,
    color=4,
    label="",
)

We could also get all the various panels into a single object using the PanelSystem object and associated generate_panel_system function.

Missing docstring.

Missing docstring for DuctTAPE.PanelSystem. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DuctTAPE.generate_panel_system. Check Documenter's build log for details.

Initializing Rotor Aerodynamics

Now that we have all the geometry set up, we're ready to start initializing the aerodynamics of our system. We first will define the freestream conditions by defining a Freestream object.

DuctTAPE.Freestream — Type

Freestream{TVI,TVR,TF}

Fields:

vinf::Float : Freestream velocities

vref::Float : Reference velocities

rho::Float : Air density value

vso::Float : Speed of sound value

mu::Float : Air dynamic viscosity value

#define freestream object from example case values
freestream = DuctTAPE.Freestream(vinf, vref, rho, vso, rmu)

Freestream{Float64, Float64, Float64}(0.0, 50.0, 1.226, 340.0, 1.78e-5)

After we have created a freestream object, we can initialize the rotor aerodynamics (circulations and velocities) as well as initialize the rotor influence on the rotor wake grid aerodynamics. We'll do so using the initialize_system_aerodynamics function.

DuctTAPE.initialize_system_aerodynamics — Function

initialize_system_aerodynamics(
    rotors, blades, wakegrid, freestream; niter=10, rlx=0.5
)

Initialize system aerodynamics for rotors and wakes.

Arguments:

rotors::Array{DuctTAPE.RotorGeometry} : array of rotor geometries

blades::Array{DuctTAPE.BladeDimensions} : array of dimensional blade geometries

wakegrid::DuctTAPE.WakeGridGeometry : wake grid geometry

freestream::DuctTAPE.Freestream : freestream information

Returns:

systemaero::DuctTAPE.SystemAero : aerodynamic values for rotor sections and wake grid

rotorvelocities::DuctTAPE.RotorVelocities : velocities along rotor blades

Vm_avg::Float : initial guess for average axial velocity in the wakes

Notes:

This function starts by initializing the rotor operating point from a constant freestream. An iteration using momentum theory is performed to find the self-induced tangential velocities on the rotor. set_grid_flow_data is called at the end to initialize the circulation and enthalpy values on the wake grid (the entropy values are set to zero for now and are initialized later). set_rotor_velocities is also called at the end of this function to save the various rotor velocities.

The returns of this function include the SystemAero object and the RotorVelocities object (see below), as well as an average axial velocity that is used later.

DuctTAPE.SystemAero — Type

SystemAero{TG,TH,TS,TC,TR}

Fields:

b_gamma_grid::Matrix{Float} : B*Γ values on wake grid

delta_enthalpy_grid::Matrix{Float} : ΔH (enthalpy) values on wake grid

delta_entropy_grid::Matrix{Float} : ΔS (entropy) values on wake grid

b_circ_rotors::Array{Array{Float}} : B*γ values at rotor blades (one array per rotor in increasing order of x location)

rotor_source_strengths::Array{Array{Float}} : σ (source strength) values at rotor blades (one array per rotor in increasing order of x location)

control_point_velocities::Matrix{Float} : velocities at control points

DuctTAPE.RotorVelocities — Type

RotorVelocities{TA}

Fields:

induced_axial_velocities::Array{Float} : local section induced axial velocities

induced_radial_velocities::Array{Float} : local section induced radial velocities

induced_tangential_velocities::Array{Float} : local section induced tangential (circumferential) velocities

absolute_axial_velocities::Array{Float} : local section absolute axial velocities

absolute_radial_velocities::Array{Float} : local section absolute radial velocities

absolute_tangential_velocities::Array{Float} : local section absolute tangential velocities

relative_axial_velocities::Array{Float} : local section relative axial velocities

relative_radial_velocities::Array{Float} : local section relative radial velocities

relative_tangential_velocities::Array{Float} : local section relateive tangential velocities

# initialize rotor and associated grid aerodynamics
system_aero, rotor_velocities, average_axial_velocity = initialize_system_aerodynamics(
    rotors, blades, wakegrid, rotor_source_panels, freestream)

Please note that the initialization that takes place here is soley based on the rotor aerodynamics. At this point, we have not taken into account the duct influence and the wake has only been initialized with some of the rotor influence. To be specific, the entropy and control point velocities are set to zeros in this function as they will be properly initialized later. The circulations, velocities, and enthalpies initialized thus far are soley based on the rotors and simply provide a first guesss so we have somewhere to start in our solution process.