Quick Start Guide

Setting up Duct Geometry

For this example, we are going to set up an example available in the DFDC source files. We have taken one of the case files and transcribed it into a julia file located in the data directory.

We'll go ahead and load that file and grab pieces of it as we go.

# --- WALL GEOMETRY DEFINITION
include("../../data/dfdc/dstestr2_case.jl");

DFDC Example Case Information Loaded

The geometry is defined for a complete airfoil as the duct wall. We actually want to split the duct wall coordinates for easier navigation as we set up the flow field.

DuctTAPE.split_wall — Function

split_wall(x,r)

Splits full airfoil coordinates into upper and lower halves.

Only works based on geometry. Splits at lowest x-value, does not split based on stagnation point.

Arguments:

x::Array{Float} : Array of x-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

r::Array{Float} : Array of r-coordinates, assumed to start at the bottom trailing edge and proceed clockwise

Returns:

xlower::Array{Float} : Array of lower x-coordinates

xupper::Array{Float} : Array of upper x-coordinates

rlower::Array{Float} : Array of lower r-coordinates

rupper::Array{Float} : Array of upper r-coordinates

# - Split Wall Coordinates
outerwallx, innerwallx, outerwallr, innerwallr = DuctTAPE.split_wall(ductx, ductr)

# - Plot Geometry
plot!(innerwallx, innerwallr, linewidth=2)
plot!(outerwallx, outerwallr, linestyle=:dash, color=1, linewidth=2)
plot!(hubx, hubr, color=2, linewidth=2)

Now we want to put all the geometry together in a DuctGeometry object.

DuctTAPE.DuctGeometry — Type

DuctGeometry{TA,TF}

Fields:

wallinnerxcoordinates::Array{Float} : x coordinates of inner (lower) wall geometry

wallouterxcoordinates::Array{Float} : x coordinates of outer (upper) wall geometry

wallinnerrcoordinates::Array{Float} : r coordinates of inner (lower) wall geometry

wallouterrcoordinates::Array{Float} : r coordinates of outer (upper) wall geometry

hubxcoordinates::Array{Float} : x coordinates of hub geometry

hubrcoordinates::Array{Float} : r coordinates of hub geometry

LEx::Float : x-position of leading edge

TEx::Float : x-position of trailing edge

chord::Float : chord length

wallbluntTE::Bool : flag for blunt trailing edge on wall

hubbluntTE::Bool : flag for blunt trailing edge on hub

DuctTAPE.defineDuctGeometry — Function

defineDuctGeometry(
    wallinnerxcoordinates,
    wallinnerrcoordinates,
    wallouterxcoordinates,
    wallouterrcoordinates,
    hubxcoordinates=nothing,
    hubrcoordinates=nothing;
    LEx=nothing,
    TEx=nothing,
    chord=nothing,
)

Constructor function for the DuctGeometry Object.

Required Arguments:

wallinnerxcoordinates::Array{Float} : x coordinates of inner (lower) wall geometry

wallouterxcoordinates::Array{Float} : x coordinates of outer (upper) wall geometry

wallinnerrcoordinates::Array{Float} : r coordinates of inner (lower) wall geometry

wallouterrcoordinates::Array{Float} : r coordinates of outer (upper) wall geometry

Optional Arguments:

hubxcoordinates::Array{Float} : x coordinates of hub geometry

hubrcoordinates::Array{Float} : r coordinates of hub geometry

Note, if hub x and r coordinates are not set, the x coordinates for the inner wall will be used and the r coordinates will be set to zero. Also note that if one of these is unset, the other must also be unset.

Keyword Arguments:

LEx::Float : x-position of manually defined leading edge. Set to foremost x-coordinate of duct and hub geometry otherwise.

TEx::Float : x-position of mannually defined trailing edge. Set to the rear-most x-coordinate of duct and hub geometry otherwise.

chord::Float : manuall defined chord length. Set to difference between leading and trailing edges otherwise.

Using the defineDuctGeometry contructor function, we can input our wall and hub geometries and let the leading and trailing edges and chord length be calculated automatically. Note that this function also outputs a spline object for the inner duct wall and the hub wall. These splines are used throughout the initialization process to help with rotor placement.

# --- DEFINE DUCT OBJECT
ductgeometry, ductsplines = DuctTAPE.defineDuctGeometry(
    innerwallx,
    innerwallr,
    outerwallx,
    outerwallr,
    hubx,
    hubr
)

(DuctTAPE.DuctGeometry{Vector{Float64}, Float64, Bool}([0.005218, 0.005585, 0.006426, 0.007724, 0.009454, 0.011592, 0.014134, 0.017087, 0.020467, 0.024287  …  0.255908, 0.261895, 0.267881, 0.273874, 0.279879, 0.28585, 0.291638, 0.296942, 0.301211, 0.304466], [0.178173, 0.175718, 0.173376, 0.171183, 0.169158, 0.167296, 0.165587, 0.164018, 0.162581, 0.161275  …  0.15829, 0.158336, 0.158373, 0.1584, 0.15842, 0.158434, 0.15844, 0.158442, 0.158441, 0.158439], [0.005218, 0.005308, 0.005823, 0.006756, 0.008099, 0.009841, 0.011977, 0.014505, 0.017416, 0.020704  …  0.242797, 0.250277, 0.257742, 0.26517, 0.272566, 0.279916, 0.287159, 0.29409, 0.30009, 0.304542], [0.178173, 0.180677, 0.183168, 0.185649, 0.1881, 0.190513, 0.192881, 0.195197, 0.197448, 0.199618  …  0.182504, 0.17999, 0.177409, 0.174771, 0.17207, 0.169314, 0.166523, 0.163786, 0.16136, 0.159526], [0.0, 0.000219, 0.000817, 0.001763, 0.003057, 0.004695, 0.00667, 0.00898, 0.011615, 0.014561  …  0.265187, 0.270466, 0.275668, 0.280781, 0.285765, 0.290619, 0.295347, 0.299968, 0.304411, 0.306379], [0.0, 0.003268, 0.006233, 0.00907, 0.011818, 0.014491, 0.017084, 0.01959, 0.022009, 0.024326  …  0.044388, 0.043968, 0.043397, 0.042673, 0.041795, 0.040757, 0.039551, 0.038174, 0.03666, 0.035928], 0.0, 0.306379, 0.306379, true, true), DuctTAPE.DuctSplines{FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}, FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}, FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}}(FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}([0.005218, 0.005585, 0.006426, 0.007724, 0.009454, 0.011592, 0.014134, 0.017087, 0.020467, 0.024287  …  0.255908, 0.261895, 0.267881, 0.273874, 0.279879, 0.28585, 0.291638, 0.296942, 0.301211, 0.304466], [0.178173, 0.175718, 0.173376, 0.171183, 0.169158, 0.167296, 0.165587, 0.164018, 0.162581, 0.161275  …  0.15829, 0.158336, 0.158373, 0.1584, 0.15842, 0.158434, 0.15844, 0.158442, 0.158441, 0.158439], [0.178173, 0.175718, 0.173376, 0.171183, 0.169158, 0.167296, 0.165587, 0.164018, 0.162581, 0.161275  …  0.158235, 0.15829, 0.158336, 0.158373, 0.1584, 0.15842, 0.158434, 0.15844, 0.158442, 0.158441], [-8.641669933613432, -3.640112675697738, -1.8180243860769707, -1.281999895051796, -0.9538273933055692, -0.7358539389569719, -0.5804379471991367, -0.4645717015983914, -0.3742571854264559, -0.3013064067502596  …  0.009920816625699514, 0.008435134507539459, 0.006970305865060831, 0.005240652061192317, 0.0037656612601192534, 0.0028640864833284875, 0.0015609361970887626, 0.0005871434125151924, -1.0711980720754257e-5, -0.0004686558108345489], [2330.6066809714234, 884.5537052660821, -115.96175576919096, 3.6222484205465384, 14.399698463766912, 13.859170257144568, 10.658775607110739, 8.271243089051719, 6.326263701934602, 4.947051657459788  …  -0.11451413290830172, -0.13205834977904302, -0.10658138437224386, -0.12200850759916244, -0.06723320536918355, -0.04272508933487499, -0.1035130081444716, -0.006099759587383213, -0.049815159618298285, -0.031173828762324322], [8.144421479810734e6, 157537.08326356643, 165610.0074552678, 35154.25643034464, 11405.151187746897, 4382.496284838639, 2022.715925927524, 1003.7263193966782, 562.3503557426164, 319.34454195688454  …  -1.0821382364922425, 1.0828171036282959, -4.2202406476753875, -0.11690993618198246, -0.8698827423337829, -7.413408928642808, 2.233507188203756, -6.317138960893495, -0.5966738201458531, -4.182406427681041]), FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}([0.005218, 0.005308, 0.005823, 0.006756, 0.008099, 0.009841, 0.011977, 0.014505, 0.017416, 0.020704  …  0.242797, 0.250277, 0.257742, 0.26517, 0.272566, 0.279916, 0.287159, 0.29409, 0.30009, 0.304542], [0.178173, 0.180677, 0.183168, 0.185649, 0.1881, 0.190513, 0.192881, 0.195197, 0.197448, 0.199618  …  0.182504, 0.17999, 0.177409, 0.174771, 0.17207, 0.169314, 0.166523, 0.163786, 0.16136, 0.159526], [0.178173, 0.180677, 0.183168, 0.185649, 0.1881, 0.190513, 0.192881, 0.195197, 0.197448, 0.199618  …  0.18495, 0.182504, 0.17999, 0.177409, 0.174771, 0.17207, 0.169314, 0.166523, 0.163786, 0.16136], [39.31488673139179, 6.826151508342943, 2.7354269104107782, 1.965180329793822, 1.494709192030506, 1.1928044256531551, 0.9816980415974484, 0.8262097289553377, 0.7040784087498327, 0.6060034042497369  …  -0.32180980843287643, -0.33130101385541294, -0.34096611134602334, -0.3503484099059978, -0.3600722851345816, -0.37000592344644057, -0.3802091101127905, -0.39033947869909613, -0.400145844063976, -0.40854883952653115], [-22102.87004954967, -3644.7578940703224, 580.3406332255618, 37.21965264680915, -15.301089224707505, -19.41207690755193, -16.292757369949875, -12.599278403555049, -10.411561572430738, -8.009661830277677  …  -0.6213936547241775, -0.6311004006176549, -0.6644102961504678, -0.6272219854643111, -0.7357650474440298, -0.6363268269390138, -0.7255285095143441, -0.5558698956815055, -0.6932453909179268, -0.6723980043793025], [-1.1732600252728407e9, -423067.16566403786, -709624.9740915189, -105423.89802979854, -27307.13775449223, -9364.610329276264, -3813.4148267069204, -1918.7623352482462, -912.9198338212042, -596.6481595025995  …  -1.0780565237508377, -1.3335352822628004, 3.2142087592947663, -2.4519804200721294, 5.787837302298633, -5.239682912035092, 2.4121858645965277, -14.577789734855536, -0.7782478846270652, -20.551152624506233]), FLOWMath.Akima{Vector{Float64}, Vector{Float64}, Float64}([0.0, 0.000219, 0.000817, 0.001763, 0.003057, 0.004695, 0.00667, 0.00898, 0.011615, 0.014561  …  0.265187, 0.270466, 0.275668, 0.280781, 0.285765, 0.290619, 0.295347, 0.299968, 0.304411, 0.306379], [0.0, 0.003268, 0.006233, 0.00907, 0.011818, 0.014491, 0.017084, 0.01959, 0.022009, 0.024326  …  0.044388, 0.043968, 0.043397, 0.042673, 0.041795, 0.040757, 0.039551, 0.038174, 0.03666, 0.035928], [0.0, 0.003268, 0.006233, 0.00907, 0.011818, 0.014491, 0.017084, 0.01959, 0.022009, 0.024326  …  0.044661, 0.044388, 0.043968, 0.043397, 0.042673, 0.041795, 0.040757, 0.039551, 0.038174, 0.03666], [19.904464653869063, 6.59550209144081, 3.157153822751898, 2.2992685365282504, 1.7632109055465794, 1.4139643498387235, 1.1631678825534388, 0.9790476943766118, 0.8373069904345483, 0.7232052108919896  …  -0.03421319590702168, -0.0641637451856979, -0.09394113158063368, -0.12461143611708475, -0.1574283833227317, -0.1933465464538242, -0.23312212814218525, -0.2761381062345554, -0.32275686922921326, -0.3587982184085237], [-7476.2927466105175, -2464.1740231340927, 405.12956724171266, 7.105745270416773, -27.339295071726628, -26.51261013453036, -22.00779434546913, -15.682232559045218, -13.017217694946956, -9.985684095752958  …  -3.5796651325887603, -3.109101444752477, -3.230050975351834, -3.5494377178469514, -4.070597642518973, -4.474152336146372, -4.832077808166373, -4.096356657493034, -4.044628598158562, -5.442462435296436], [-6.973962413497666e7, -457858.54093777615, -605043.9100597023, -110374.9361459641, -32262.239587185122, -12482.756588486083, -5150.089534851246, -2837.0744313531254, -1436.5977227237267, -889.3997374681618  …  99.20271843687578, 36.46385402635212, 36.15494709752692, 44.366671042074366, 62.500082789494975, 51.7725635663145, 39.905633251639806, -136.74849433729474, -1.7017592850302152, -630.5757545012166])))

The ductgeometry and ductsplines objects now contains all the geometry information we'll need.

Defining Rotors

Along with the duct geometry, we will need to define an array of Rotor objects.

DuctTAPE.RotorGeometry — Type

RotorGeometry{TF, TI, TA, TC, TT, TAF, TR, TM}

Fields:

xlocation::Float : x location of rotor plane, non-dimensional based on duct chord (max TE location - min LE location of hub/wall)

numblades::Int : number of rotor blades

radialstations::Array{Float} : array of radial stations defining rotor blade, non-dimensional with hub=0, tip=1

tipgap::Float : gap between blade tip and duct wall (not implemented yet)

chords::Array{Float} : array of chord lengths at radial stations defining rotor blade, non-dimensional based on blade tip radius

twists::Array{Float} : array of twist values (in degrees) at radial stations defining rotor blade

skews::Array{Float} : array of skew values (similar to sweep) at radial stations defining rotor blade, non-dimensional based on rotor tip radius. (note: this is for reference only, the solver can't use this information)

rakes::Array{Float} : array of rake values (similar to dihedral) at radial stations defining rotor blade, non-dimensional based on rotor tip radius. (note: this is for reference only right now. it may be implemented into the grid initialization functions later.)

airfoils::Array{Airfoil} : array of airfoil data objects at radial stations defining rotor blade

solidity:Array{Float} : array of rotor solidity at radial stations defining rotor blade, chord/distance between blade sections

rpm::Float : RPM of rotor

Note that we want to create an array, even if we only have one rotor. When we initialize the grid, it will expect an array. Also, our rotor object has more fields than are used in the original dfdc, for now, we'll set the section skew, rake, and solidity to nothing. For the airfoils, we will use the CCBlade.jl functionality and for our case here, define our airfoils as ccb.AlphaReAF objects (where CCBlade has been included in the DuctTAPE package and renamed 'ccb') using data files from digitized dfdc plots.

# -- GENERATE ROTOR OBJECT ARRAY

# set datapath for airfoil data files. Remember to grab these files and change the data path for your case.
datapath = "../../data/dfdc/airfoils/"

# generate airfoils for rotor
af1 = DuctTAPE.ccb.AlphaReAF([
    datapath * "disk1_re5e5.dat",
    datapath * "disk1_re1e6.dat",
    datapath * "disk1_re1.5e6.dat",
    datapath * "disk1_re2e6.dat",
])

# generate airfoils for stator
af2 = DuctTAPE.ccb.AlphaReAF([
    datapath * "disk2_re5e5.dat",
    datapath * "disk2_re1e6.dat",
    datapath * "disk2_re1.5e6.dat",
    datapath * "disk2_re2e6.dat",
])

#generate rotor object
rotor1 = DuctTAPE.RotorGeometry(
    xdisk1, #x position of rotor
    nblade1, #number of blades
    rnondim1, #radial stations
    0.0, #tip gap
    chord1, #chords
    beta1, #twists
    nothing, #skews
    nothing, #rakes
    fill(af1,length(rnondim1)), #airfoils
    nothing, #solidities
    rpm, #RPM
)

#generate stator object (rpm is zero for stator)
rotor2 = DuctTAPE.RotorGeometry(
    xdisk2, #x position of rotor
    nblade2, #number of blades
    rnondim2, #radial stations
    0.0, #tip gap
    chord2, #chords
    beta2, #twists
    nothing, #skews
    nothing, #rakes
    fill(af2,length(rnondim2)), #airfoils
    nothing, #solidities
    0.0, #RPM
)

#assemble array
rotors = [rotor1; rotor2]

2-element Vector{DuctTAPE.RotorGeometry{Float64, Int64, Vector{Float64}, Float64, Vector{Float64}, Vector{Float64}, Nothing, Nothing, Vector{CCBlade.AlphaReAF{Float64, String}}, Nothing, Float64}}:
 DuctTAPE.RotorGeometry{Float64, Int64, Vector{Float64}, Float64, Vector{Float64}, Vector{Float64}, Nothing, Nothing, Vector{CCBlade.AlphaReAF{Float64, String}}, Nothing, Float64}(0.39167175295956963, 6, [0.0, 0.11110776761354521, 0.22221553522709042, 0.33331327234793784, 0.4444210399614829, 0.5555488685604237, 0.6665864227250843, 0.7777242818167229, 0.8888621409083616, 0.9999999999999999], 0.0, [0.3744190691790399, 0.3571542712564086, 0.3402423596777416, 0.3250882215859911, 0.3122378320793662, 0.3017244823223917, 0.2933950329582529, 0.2870430787668953, 0.2824688727611692, 0.27951927558425993], [70.828, 59.758, 51.707, 46.429, 42.209, 38.766, 35.897, 33.457, 31.337, 29.453], nothing, nothing, CCBlade.AlphaReAF{Float64, String}[CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-1.83247422837987 -1.83247422837987 -1.83247422837987 -1.83247422837987; -1.8220217655802948 -1.8220217655802948 -1.8220217655802948 -1.8220217655802948; … ; 2.201288675740593 2.201288675740593 2.201288675740593 2.201288675740593; 2.2106958925035443 2.2106958925035443 2.2106958925035443 2.2106958925035443], [0.7587569421558654 0.5719166039184465 0.6427366145549502 0.5536068691841667; 0.7491961764011059 0.5649976280688906 0.6349370410615811 0.5470652926058168; … ; 0.4865868352978613 0.42183733286048736 0.45944804534869116 0.3622685436104774; 0.4928516534112002 0.427045434339654 0.4651090252425691 0.3664576687162524], "DFDC Disk 1 Airfoil", 0.15)], nothing, 8000.0)
 DuctTAPE.RotorGeometry{Float64, Int64, Vector{Float64}, Float64, Vector{Float64}, Vector{Float64}, Nothing, Nothing, Vector{CCBlade.AlphaReAF{Float64, String}}, Nothing, Float64}(0.7180648804258777, 9, [0.0, 0.11064323803152241, 0.2212963145156531, 0.3320871293362981, 0.44301568249345763, 0.5541213277975642, 0.6652958422699281, 0.7766671257944551, 0.8882351783711461, 1.0], 0.0, [0.38141720896601594, 0.3786629856044173, 0.37502136330769736, 0.3708604483007954, 0.366384013672517, 0.3617432459081049, 0.3570433182146848, 0.3523828304739368, 0.3478472359166503, 0.3434956944718333], [92.711, 91.739, 91.228, 90.868, 90.638, 90.441, 90.124, 89.583, 88.955, 88.388], nothing, nothing, CCBlade.AlphaReAF{Float64, String}[CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15), CCBlade.AlphaReAF{Float64, String}([-90.0, -89.0, -88.0, -87.0, -86.0, -85.0, -84.0, -83.0, -82.0, -81.0  …  81.0, 82.0, 83.0, 84.0, 85.0, 86.0, 87.0, 88.0, 89.0, 90.0], [100000.0, 1.0e6, 1.5e6, 2.0e6], [-2.1875623555013735 -2.1875623555013735 -2.1875623555013735 -2.1875623555013735; -2.178599934469161 -2.178599934469161 -2.178599934469161 -2.178599934469161; … ; 1.8240979418446208 1.8240979418446208 1.8240979418446208 1.8240979418446208; 1.8346219968858473 1.8346219968858473 1.8346219968858473 1.8346219968858473], [0.5569924531654876 0.5449484868984782 0.43162527216141156 0.31260669435394906; 0.5498304299041235 0.5380729449908267 0.4264686157347317 0.309168923392014; … ; 0.6031592027561349 0.6292733053856933 0.7439713642510304 0.3790396444056832; 0.6107281591876736 0.6370945598941206 0.7533064099512512 0.38332871946022884], "DFDC Disk 2 Airfoil", 0.15)], nothing, 0.0)

Take note that the rotor objects are defined using non-dimensional parameters. If we want to visualize things at this point, we'll have to dimensionalize things, which is done internally in the code using BladeDimensions objects.

DuctTAPE.BladeDimensions — Type

BladeDimensions{TF, TA}

Fields:

rhub::Float : hub radius (dimensional)

rtip::Float : tip radius (dimensional)

rdim::Array{Float} : array of dimensional radial stations

cdim::Array{Float} : array of dimensional chords

tdim::Array{Float} : array of twists (already dimensional in rotorgeometry)

sweptannulus::Float : area of blade swept annulus

sweptarea::Float : area of blade tip swept disk

DuctTAPE.initialize_blade_dimensions — Function

initialize_blade_dimensions(ductsplines, Rotor)

Initilialize needed blade information for various calculations during the solution process.

Arguments:

ductsplines::DuctTAPE.ductsplines : ductsplines object containing splines for duct wall and hub

Rotor::DuctTAPE.Rotor : Rotor object for which to define blade information

blade1 = DuctTAPE.initialize_blade_dimensions(ductgeometry, ductsplines, rotor1)
blade2 = DuctTAPE.initialize_blade_dimensions(ductgeometry, ductsplines, rotor2)

nr = length(blade1.rdim)

plot!(rotor1.xlocation.*ductgeometry.chord*ones(nr), blade1.rdim, color=4, linewidth=2)
plot!(rotor2.xlocation.*ductgeometry.chord*ones(nr), blade2.rdim, color=4, linewidth=2)

Wake Grid Initialization

With the duct geometry and rotor data defined, we can now initialize the rotor wake grid. We'll begin by setting some grid options.

Note

The number of radial stations set for grid options must match the number of radial stations defined for the rotor objects. Furthermore, the rotors must have the same number of radial stations defined.

DuctTAPE.GridOptions — Type

GridOptions{TF,TI}

Fields:

num_radial_stations::Integer : Number of radial stations (equal to number of rotor blade elements used in analysis)

inlet_length::Float : inlet length (unused)

wake_length::Float : length of wake behind duct relative to chord length

wake_expansion_factor::Float : expansion factor to apply to wake grid generation

DuctTAPE.defineGridOptions — Function

defineGridOptions(
    num_radial_stations;
    inlet_length=0.5,
    wake_length=1.0,
    wake_expansion_factor=1.1
)

Constructor function for the GridOptions object.

Required Argument:

num_radial_stations::Integer : Number of radial stations (equal to number of rotor blade elements used in analysis)

Keyword Arguments:

inlet_length::Float : inlet length (unused)

wake_length::Float : length of wake behind duct in terms of chord length

wake_expansion_factor::Float : expansion factor to apply to wake grid generation

# --- SET GRID OPTIONS
num_radial_stations = length(rnondim1)
grid_options = DuctTAPE.defineGridOptions(num_radial_stations)

DuctTAPE.GridOptions{Float64, Int64}(10, 0.5, 1.0, 1.1)

With options set, rotor location chosen, and the wall and hub geometry available, we're finally ready to initialize the wake grid object.

DuctTAPE.initialize_grid — Function

initialize_grid(ductgeometry, ductsplines, rotors, grid_options; max_iterations=-1, tol=1e-9)

Initialize grid via zero-thrust, unit freestream solution.

Arguments:

ductgeometry::DuctTAPE.DuctGeometry : ductgeometry Geometry Object

ductsplines::DuctTAPE.DuctSplines : ductsplines object

rotors::Array{DuctTAPE.Rotor} : Array of rotor objects

grid_options::DuctTAPE.GridOptions : Grid options object

Keyword Arguments:

max_iterations::Int : maximum number of iterations to run, default=100

tol::Float : convergence tolerance, default = 1e-9

Returns:

WakeGrid::DuctTAPE.WakeGrid : WakeGrid Object

DuctTAPE.WakeGrid — Type

WakeGrid{TF,TI,TA,TW,TH}

Grid Object

Fields:

x_grid_points::Matrix{Float} : 2D Array of x grid points

r_grid_points::Matrix{Float} : 2D Array of radial grid points

nx::Int : number of x stations

nr::Int : number of radial stations

wallTEidx::Int : index of duct wall trailing edge x location

hubTEidx::Int : index of hub wall trailing edge x location

rotoridxs::Array{Int} : array of indices of rotor x locations

# --- INITIALIZE GRID
wakegrid = DuctTAPE.initialize_grid(ductgeometry, ductsplines, rotors, grid_options)

xg = wakegrid.x_grid_points
rg = wakegrid.r_grid_points
nx = wakegrid.nx
nr = wakegrid.nr

plot(xlabel="x", ylabel="r",aspectratio=:equal)
plot!(ductgeometry.wallinnerxcoordinates, ductgeometry.wallinnerrcoordinates, color=1, linewidth=2)
plot!(ductgeometry.wallouterxcoordinates, ductgeometry.wallouterrcoordinates, color=1, linestyle=:dash, linewidth=2)
plot!(ductgeometry.hubxcoordinates, ductgeometry.hubrcoordinates, color=2, linewidth=2)

plot!(rotor1.xlocation.*ductgeometry.chord*ones(nr), blade1.rdim, color=4, linewidth=2)
plot!(rotor2.xlocation.*ductgeometry.chord*ones(nr), blade2.rdim, color=4, linewidth=2)

plot!(xg, rg, color=3, linewidth=0.5)
plot!(xg', rg', color=3, linewidth=0.5)

Note

This process should be relatively robust in that most combinations of wall and hub and rotor position should generate a grid. However, in the case that there is no overlap in the x-positions of the wall and hub, problems may occur later in the solver. In addition if the rotor is positioned somewhere not between the wall and hub, for example, if the rotor is out in front of the duct, then if a solution is found, it will likely be inaccurate. And finally, things will break if the hub and wall overlap in the radial direction, in other words, if the duct is blocked.

As can be seen in the plot above, there are a few features of the wake grid that deserve mention for the user's information. First, the grid is designed such that the grid spacing in the axial direction will line up with the rotor positions as well as the trailing edge positions of the duct wall and hub (as well as the leading edge positions if they happen to be behind a rotor, which is not advised). In addition, the grid axial spacing is taken to be as close as possible to the radial spacing, which is defined directly from the rotor radial station positions. The wake spacing is started at the average of the axial spacing inside the duct area and then expanded by an expansion factor that can be defined by the user and is set to 1.1 by default. This means that the end of the wake will actually not lie directly at the length input by the user (default 2x duct chord), but should be close enough.

System Paneling

With the geometry defined and the grid initialized, we can define the system panels and their control points (centers). To do so, we use the generate_paneling function:

DuctTAPE.generate_paneling — Function

generate_paneling(ductgeometry, ductsplines, rotors, wakegrid)

Generate panel edges and centers.

Arguments:

ductgeometry::DuctTAPE.DuctGeometry : Duct Geometry object

ductsplines::DuctTAPE.DuctSplines : Duct Splines object

rotors::Array{DuctTAPE.Rotor} : Array of rotor objects

wakegrid::DuctTAPE.WakeGrid : Wake Grid object

Returns:

wall_panels::DuctTAPE.Panels : Panels object for duct wall

hub_panels::DuctTAPE.Panels : Panels object for hub

wake_panels::DuctTAPE.Panels : Panels object for vortex wake sheets

rotor_source_panels::Array{DuctTAPE.Panels} : Array of Panels objects for each rotor

For our case here, we are going to use a rough sampling of the geometry data in order to more clearly see the paneling.

# Get paneling of various objects

wall_panels, hub_panels, wake_panels, rotor_source_panels = DuctTAPE.generate_paneling(
    ductgeometry, ductsplines, rotors, wakegrid
)

(DuctTAPE.Panels{Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, String}([(0.304411, 0.301211), (0.301211, 0.2598989241576609), (0.2598989241576609, 0.21858684831532185), (0.21858684831532185, 0.1854676288736064), (0.1854676288736064, 0.15234840943189099), (0.15234840943189099, 0.11922918999017555), (0.11922918999017555, 0.08399779473430041), (0.08399779473430041, 0.05221509391763513), (0.05221509391763513, 0.026992199739212637), (0.026992199739212637, 0.010798104819454313)  …  (0.026992199739212637, 0.05221509391763513), (0.05221509391763513, 0.08399779473430041), (0.08399779473430041, 0.11922918999017555), (0.11922918999017555, 0.15234840943189099), (0.15234840943189099, 0.1854676288736064), (0.1854676288736064, 0.21858684831532185), (0.21858684831532185, 0.2598989241576609), (0.2598989241576609, 0.301211), (0.301211, 0.304411), (0.304411, 0.304411)], [(0.1584357794458491, 0.158441), (0.158441, 0.15832013170161133), (0.15832013170161133, 0.15779422955018368), (0.15779422955018368, 0.15706356222783383), (0.15706356222783383, 0.15625871421756632), (0.15625871421756632, 0.15569243840130506), (0.15569243840130506, 0.15553332936144448), (0.15553332936144448, 0.15666087745744797), (0.15666087745744797, 0.16070206709552126), (0.16070206709552126, 0.16801256543962217)  …  (0.20283950416761726, 0.2111466860423776), (0.2111466860423776, 0.2143735120654717), (0.2143735120654717, 0.212389453342464), (0.212389453342464, 0.20692135663993397), (0.20692135663993397, 0.19934880741319233), (0.19934880741319233, 0.19014786781386786), (0.19014786781386786, 0.17664654297261145), (0.17664654297261145, 0.16090350259495168), (0.16090350259495168, 0.15958049570629068), (0.15958049570629068, 0.1584357794458491)], [(0.302811, 0.15843838972292457), (0.28055496207883046, 0.15838056585080568), (0.23924288623649137, 0.15805718062589752), (0.20202723859446414, 0.15742889588900877), (0.1689080191527487, 0.15666113822270006), (0.13578879971103328, 0.1559755763094357), (0.10161349236223799, 0.15561288388137479), (0.06810644432596777, 0.15609710340944621), (0.039603646828423884, 0.1586814722764846), (0.018895152279333475, 0.16435731626757172)  …  (0.039603646828423884, 0.20699309510499742), (0.06810644432596777, 0.21276009905392465), (0.10161349236223799, 0.21338148270396784), (0.13578879971103328, 0.209655404991199), (0.1689080191527487, 0.20313508202656316), (0.20202723859446414, 0.1947483376135301), (0.23924288623649137, 0.18339720539323967), (0.28055496207883046, 0.16877502278378156), (0.302811, 0.16024199915062118), (0.304411, 0.1590081375760699)], "w"), DuctTAPE.Panels{Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, String}([(0.0, 0.005835491917426202), (0.005835491917426202, 0.02277074906256416), (0.02277074906256416, 0.049148030471173006), (0.049148030471173006, 0.08238534405765194), (0.08238534405765194, 0.11922918999017555), (0.11922918999017555, 0.15234840943189099), (0.15234840943189099, 0.1854676288736064), (0.1854676288736064, 0.21858684831532185), (0.21858684831532185, 0.2598989241576609), (0.2598989241576609, 0.301211), (0.301211, 0.304411), (0.304411, 0.304411)], [(0.0, 0.016016085927272015), (0.016016085927272015, 0.029450308667391695), (0.029450308667391695, 0.038937936828243126), (0.038937936828243126, 0.044084613831847655), (0.044084613831847655, 0.044975455240187236), (0.044975455240187236, 0.04492940207091704), (0.04492940207091704, 0.044897075915259585), (0.044897075915259585, 0.044863552514875366), (0.044863552514875366, 0.044539087992498276), (0.044539087992498276, 0.037744824126883444), (0.037744824126883444, 0.03665999999999999), (0.03665999999999999, 0.0)], [(0.002917745958713101, 0.008008042963636008), (0.01430312048999518, 0.022733197297331857), (0.035959389766868584, 0.03419412274781741), (0.06576668726441247, 0.041511275330045394), (0.10080726702391374, 0.04453003453601745), (0.13578879971103328, 0.04495242865555214), (0.1689080191527487, 0.044913238993088314), (0.20202723859446414, 0.044880314215067475), (0.23924288623649137, 0.044701320253686824), (0.28055496207883046, 0.04114195605969086), (0.302811, 0.037202412063441714), (0.304411, 0.018329999999999996)], "w"), DuctTAPE.Panels{Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, String}([(0.304411, 0.33209024579027946), (0.33209024579027946, 0.36253741615958684), (0.36253741615958684, 0.396029303565825), (0.396029303565825, 0.43287037971268694), (0.43287037971268694, 0.4733955634742351), (0.4733955634742351, 0.5179732656119381), (0.5179732656119381, 0.5670087379634114), (0.5670087379634114, 0.6209477575500321), (0.11922918999017555, 0.15234840943189099), (0.15234840943189099, 0.1854676288736064)  …  (0.5670087379634114, 0.6209477575500321), (0.301211, 0.304411), (0.304411, 0.33209024579027946), (0.33209024579027946, 0.36253741615958684), (0.36253741615958684, 0.396029303565825), (0.396029303565825, 0.43287037971268694), (0.43287037971268694, 0.4733955634742351), (0.4733955634742351, 0.5179732656119381), (0.5179732656119381, 0.5670087379634114), (0.5670087379634114, 0.6209477575500321)], [(0.03666, 0.03666), (0.03666, 0.03666), (0.03666, 0.03666), (0.03666, 0.03666), (0.03666, 0.03666), (0.03666, 0.03666), (0.03666, 0.03666), (0.03666, 0.03666), (0.0726547010304667, 0.07280453968516014), (0.07280453968516014, 0.07293542514113499)  …  (0.1292429505396089, 0.1292425196161467), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441), (0.158441, 0.158441)], [(0.3182506228951397, 0.03666), (0.3473138309749332, 0.03666), (0.3792833598627059, 0.03666), (0.41444984163925597, 0.03666), (0.453132971593461, 0.03666), (0.49568441454308665, 0.03666), (0.5424910017876747, 0.03666), (0.5939782477567217, 0.03666), (0.13578879971103328, 0.07272962035781341), (0.1689080191527487, 0.07286998241314757)  …  (0.5939782477567217, 0.1292427350778778), (0.302811, 0.158441), (0.3182506228951397, 0.158441), (0.3473138309749332, 0.158441), (0.3792833598627059, 0.158441), (0.41444984163925597, 0.158441), (0.453132971593461, 0.158441), (0.49568441454308665, 0.158441), (0.5424910017876747, 0.158441), (0.5939782477567217, 0.158441)], "v"), DuctTAPE.Panels[DuctTAPE.Panels{Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, String}([(0.11922918999017555, 0.11922918999017555), (0.11922918999017555, 0.11922918999017555), (0.11922918999017555, 0.11922918999017555), (0.11922918999017555, 0.11922918999017555)], [(0.044975455240187236, 0.0726547010304667), (0.0726547010304667, 0.10033269741902229), (0.10033269741902229, 0.12800569620068247), (0.12800569620068247, 0.15569243840130506)], [(0.11922918999017555, 0.058815078135326965), (0.11922918999017555, 0.0864936992247445), (0.11922918999017555, 0.11416919680985238), (0.11922918999017555, 0.14184906730099378)], "s"), DuctTAPE.Panels{Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, Vector{Tuple{Float64, Float64}}, String}([(0.21858684831532185, 0.21858684831532185), (0.21858684831532185, 0.21858684831532185), (0.21858684831532185, 0.21858684831532185), (0.21858684831532185, 0.21858684831532185)], [(0.044863552514875366, 0.07296445190393547), (0.07296445190393547, 0.10115280764303497), (0.10115280764303497, 0.12942095989706012), (0.12942095989706012, 0.15779422955018368)], [(0.21858684831532185, 0.05891400220940542), (0.21858684831532185, 0.08705862977348522), (0.21858684831532185, 0.11528688377004755), (0.21858684831532185, 0.1436075947236219)], "s")])

# PLOT PANELS

plot(; xlabel="x", ylabel="r", aspectratio=:equal, legend=true, label="")

# wall panels:
for i in 1:length(wall_panels.panel_edges_x)
    plot!(
        [wall_panels.panel_edges_x[i][1]; wall_panels.panel_edges_x[i][2]],
        [wall_panels.panel_edges_r[i][1]; wall_panels.panel_edges_r[i][2]];
        color=1,
        linewidth=0.5,
        markershape=:diamond,
        markersize=2,
        label="",
    )
end

scatter!(
    getindex.(wall_panels.panel_centers, 1),
    getindex.(wall_panels.panel_centers, 2);
    color=1,
    markersize=3,
    markershape=:circle,
    label="wall panel centers",
)

#hub panels:
for i in 1:length(hub_panels.panel_edges_x)
    plot!(
        [hub_panels.panel_edges_x[i][1]; hub_panels.panel_edges_x[i][2]],
        [hub_panels.panel_edges_r[i][1]; hub_panels.panel_edges_r[i][2]];
        markersize=2,
        markershape=:diamond,
        color=2,
        linewidth=0.5,
        label="",
    )
end

scatter!(
    getindex.(hub_panels.panel_centers, 1),
    getindex.(hub_panels.panel_centers, 2);
    markersize=3,
    markershape=:circle,
    color=2,
    label="hub panel centers",
)

#vortex sheet panels
for i in 1:length(wake_panels.panel_centers)
    plot!(
        [wake_panels.panel_edges_x[i][1]; wake_panels.panel_edges_x[i][2]],
        [wake_panels.panel_edges_r[i][1]; wake_panels.panel_edges_r[i][2]];
        markersize=2,
        markershape=:diamond,
        color=3,
        linewidth=0.5,
        label="",
    )
end

scatter!(
    getindex.(wake_panels.panel_centers, 1),
    getindex.(wake_panels.panel_centers, 2);
    markersize=3,
    markershape=:circle,
    color=3,
    label="vortex sheet panel centers",
)

#rotor source panels:
    for i in 1:length(rotor_source_panels[1].panel_centers)
        plot!(
            [
                rotor_source_panels[1].panel_edges_x[i][1]
                rotor_source_panels[1].panel_edges_x[i][2]
            ],
            [
                rotor_source_panels[1].panel_edges_r[i][1]
                rotor_source_panels[1].panel_edges_r[i][2]
            ];
            markersize=2,
            markershape=:diamond,
            color=4,
            linewidth=0.5,
            label="",
        )
    end

    scatter!(
        getindex.(rotor_source_panels[1].panel_centers, 1),
        getindex.(rotor_source_panels[1].panel_centers, 2);
        markersize=3,
        markershape=:circle,
        color=4,
        label="rotor source panel centers",
    )

    for i in 1:length(rotor_source_panels[2].panel_centers)
        plot!(
            [
                rotor_source_panels[2].panel_edges_x[i][1]
                rotor_source_panels[2].panel_edges_x[i][2]
            ],
            [
                rotor_source_panels[2].panel_edges_r[i][1]
                rotor_source_panels[2].panel_edges_r[i][2]
            ];
            markersize=2,
            markershape=:diamond,
            color=4,
            linewidth=0.5,
            label="",
        )
    end

    scatter!(
        getindex.(rotor_source_panels[2].panel_centers, 1),
        getindex.(rotor_source_panels[2].panel_centers, 2);
        markersize=3,
        markershape=:circle,
        color=4,
        label="",
    )

The paneling for the rotor sources and the vortex wake sheets is based directly on the wake grid initialized in the previous section (remember that we used a rougher dataset here for visual clarity). The paneling of the duct wall and hub are set such that the panels aft of the foremost rotor also align perfectly with the wake grid. The points in front of the foremost rotor are set using cosine spacing such that the last panel before the foremost rotor is roughly similar in length to the average of the panel lenghts in the remainder of the duct. (Note that the estimation process for this is not particularly robust at this point.)